@@ -1,3 +1,3 @@
 <?php
-return ['domain'=>'flexible-shipping-usps','plural-forms'=>'nplurals=2; plural=(n != 1);','language'=>'de_DE','project-id-version'=>'Shipping Live Rates for USPS for WooCommerce 3.0.7','pot-creation-date'=>'2025-06-23T23:09:39+00:00','po-revision-date'=>'YEAR-MO-DA HO:MI+ZONE','x-generator'=>'WP-CLI 2.11.0','messages'=>['https://wordpress.org/plugins/flexible-shipping-usps/'=>'https://de.wordpress.org/plugins/flexible-shipping-usps/','Octolize'=>'Octolize','https://octol.io/usps-author'=>'https://octol.io/usps-author','Get USPS WooCommerce Live Rates PRO!'=>'Hole dir die USPS Live-Versandtarife PRO!','Handling Fees'=>'Bearbeitungsgebühren','Automatic Box Packing'=>'Automatische Verpackungsanpassung','Premium Support'=>'Premium-Support','Multicurrency Support'=>'Unterstützung mehrerer Währungen','Service Code'=>'USPS Servicecode','Live rates for USPS and WooCommerce'=>'USPS WooCommerce Live-Versandtarife','Settings'=>'Einstellungen','Docs'=>'Docs','Support'=>'Support','Help us improve Octolize plugins\' experience'=>'Verbessere mit uns die Erfahrung mit den Octolize-Plugins!','Hi %1$s, with your helping hand we can build effective solutions, launch the new features and shape better plugins experience. By agreeing to anonymously share non-sensitive %2$susage data%3$s of our plugins, you will help us develop them in the right direction. No personal data is tracked or stored and you can opt-out any time. Will you give the thumbs up to our efforts?'=>'Hallo %1$s, mit deiner hilfreichen Hand können wir effektive Lösungen entwickeln, neue Funktionen einführen und die Erfahrung mit den Plugins verbessern. Indem du zustimmst, nicht sensible %2$sNutzungsdaten%3$s unserer Plugins anonym zu teilen, hilfst du uns, sie in die erwünschte Richtung zu entwickeln. Es werden keine persönlichen Daten nachverfolgt oder gespeichert und du kannst sich jederzeit abmelden. Können wir uns auf dich verlassen?','Allow'=>'Erlauben','Page titleShipping Extensions'=>'Versanderweiterungen','Menu TitleShipping
-Extensions'=>'Versanderweiterungen','Buy bundle →'=>'Paket kaufen →','Buy plugin →'=>'Plugin kaufen →','Flexible Shipping PRO'=>'Flexible Shipping PRO','The best and the most powerful Table Rate shipping plugin for WooCommerce. Define the shipping rules based on numerous conditions and configure even the most complex shipping scenarios with ease.'=>'Das beste und leistungsstärkste Table Rate Versand-Plugin für WooCommerce. Definiere die Versandregeln basierend auf zahlreichen Bedingungen und konfiguriere leicht selbst die komplexesten Versandszenarien.','All Plugins Bundle'=>'Alle Plugins-Paket','Grab a pack of all Octolize plugins as a cut-price tailor-made limited offer for developers, agencies and freelancers. Move the WooCommerce shipping to a whole new level. No strings attached, each plugin\'s 25‑sites subscription included.'=>'Hol dir ein Paket mit allen Octolize-Plugins als maßgeschneidertes, preisgünstiges Sonderangebot für Entwickler, Agenturen und Freelancers. Bringe den Versand mit WooCommerce auf ein ganz neues Level. Keine Verpflichtungen, inklusive einem Abonnement für 25 Websites pro Plugin.','Complete UPS Integration Bundle'=>'Vollständiges UPS-Integrationspaket','Connect your WooCommerce store with your UPS account, offer your customers real-time shipping rates and generate printable shipping labels for each placed order.'=>'Verbinde deinen WooCommerce-Shop mit deinem UPS-Konto, biete deinen Kunden Echtzeit-Versandtarife an und generiere druckbare Versandetiketten für jede Bestellung.','Advanced DPD UK Integration Bundle'=>'Erweitertes DPD UK-Integrationspaket','Integrate your WooCommerce store with your DPD UK account, customize the UK locations and automate the whole order processing including generating and printing the shipping labels.'=>'Integriere deinen WooCommerce-Shop mit deinem DPD UK-Konto, passe die Standorte in Großbritannien an und automatisiere den gesamten Bestellvorgang, inklusive der Erstellung und des Ausdrucks der Versandetiketten.','Flexible Shipping Bundle'=>'Flexible Shipping-Paket','Fully customize the shipping cost in your shop, define its calculation rules based on numerous conditions, hide and display the shipping methods and divide orders into separate packages.'=>'Passe die Versandkosten in deinem Shop vollständig an, definiere die Berechnungsregeln basierend auf zahlreichen Bedingungen, verberge und zeige die Versandmethoden und teile Bestellungen in unterschiedliche Pakete auf.','Live Rates Bundle'=>'Live-Versandtarife-Paket','Serve your customers the UPS, FedEx, DHL Express or USPS shipping methods with automatically calculated rates and use multiple conditions to display them according to your terms and needs.'=>'Biete deinen Kunden die Versandmethoden von UPS, FedEx, DHL Express oder USPS mit automatisch berechneten Tarifen an und verwende verschiedene Bedingungen, um sie gemäß deinen Bedingungen und Anforderungen anzuzeigen.','UPS Live Rates and Access Points PRO'=>'UPS WooCommerce Live-Versandtarife und Zugangspunkte PRO','WooCommerce UPS integration packed with many advanced features. Display the dynamically calculated live rates for UPS shipping methods and adjust them to your needs.'=>'WooCommerce UPS-Integration mit vielen erweiterten Funktionen. Zeige die dynamisch berechneten Live-Tarife für UPS-Versandmethoden an und pass an deine Bedürfnisse an.','FedEx WooCommerce Live Rates PRO'=>'FedEx WooCommerce Live-Versandtarife PRO','Enable the FedEx live rates for international delivery and integrate it with your shop in less than 5 minutes. Save your time and money – let the shipping cost be calculated automatically.'=>'Aktiviere die FedEx Live-Tarife für den internationalen Versand und integriere sie in weniger als 5 Minuten in deinen Shop. Spare Zeit und Geld - lass die Versandkosten automatisch berechnen.','USPS Live Rates PRO'=>'USPS WooCommerce Live-Versandtarife PRO','Serve your customers the automatically and real-time calculated USPS shipping rates. Add the handling fees, insurance and adjust them to your needs with just a few clicks.'=>'Biete deinen Kunden die automatisch und in Echtzeit berechneten USPS-Versandtarife an. Füge die Bearbeitungsgebühren und die Versicherung hinzu und passe sie mit nur wenigen Klicks an deine Bedürfnisse an.','Flexible Shipping Import / Export'=>'Flexible Shipping Import / Export','Use the CSV files to import or export your shipping methods. Edit, update, move or backup the ready configurations and shipping scenarios.'=>'Verwende die CSV-Dateien, um deine Versandmethoden zu importieren oder zu exportieren. Bearbeite, aktualisiere, verschiebe oder sichere die fertigen Konfigurationen und Versandszenarien.','Flexible Shipping Locations'=>'Flexible Shipping Standorte','Calculate the shipping cost based on location. Define your own custom locations, use the WooCommerce defaults or the ones created by 3rd party plugins.'=>'Berechne die Versandkosten basierend auf dem Standort. Bestätige deine eigenen benutzerdefinierten Standorte, verwende die WooCommerce-Standardwerte oder die von Drittanbieter-Plugins erstellten Standorten.','Distance Based Shipping Rates'=>'Entfernungsabhängige Versandtarife','Offer shipping rates based on Distance or Total Travel Time calculated by Google Distance Matrix API and don\'t overpay for shipping.'=>'Biete die Versandtarife an, die auf der von Google Distance Matrix API berechneten Entfernung oder Gesamtreisezeit basieren, und überzahle für den Versand nicht mehr.','Pickup Points PRO WooCommerce Plugin'=>'Packstationen PRO WooCommerce-Plugin','Provide your customers with multiple carriers\' pickup points at the checkout and let them choose the preferred one to collect their order from.'=>'Biete deinen Kunden an der Kasse mehrere Packstationen verschiedener Lieferdienste an und lass sie wählen, von welcher sie ihre Bestellung abholen möchten.','Shipping Cost on Product Page PRO'=>'Rechner der Versandkosten auf der Produktseite PRO','Let your customers calculate and see the shipping cost on product pages based on the entered shipping destination and cart contents. Decide how and when exactly you want the shipping cost calculator to display.'=>'Lass deine Kunden die Versandkosten auf den Produktseiten berechnen und anzeigen, basierend auf dem eingegebenen Versandziel und dem Inhalt des Warenkorbs. Entscheide wie und wann genau du den Versandkostenrechner anzeigen lassen möchtest.','Flexible Shipping Box Packing'=>'Flexible Shipping Verpackungsanpassung','Use the advanced box packing WooCommerce algorithm to fit the ordered products into your shipping boxes the most optimal way. Configure the shipping cost based on the type and number of the used shipping boxes.'=>'Verwende den fortgeschrittenen WooCommerce-Algorithmus für das optimale Verpacken der bestellten Produkte in deine Versandkartons. Konfiguriere die Versandkosten basierend auf der Art und Anzahl der verwendeten Versandkartons.','Shipping Packages'=>'Versandpakete','Split the WooCommerce cart content into multiple packages based on various conditions like shipping class.'=>'Teile den Inhalt des WooCommerce-Warenkorbs in mehrere Pakete auf, die auf verschiedenen Bedingungen wie der Versandklasse basieren.','Multi Vendor Shipping'=>'Multi Vendor Versand','Define precisely the shipping cost calculation rules for each Vendor / Product Author in your marketplace or multivendor store.'=>'Definiere genau die Regeln zur Berechnung der Versandkosten für jeden Verkäufer / Produktautor auf deinem Marktplatz oder Multivendor-Shop.','Delivery Date Picker'=>'Lieferterminwähler','Let your customers choose a convenient delivery date for the ordered products and make the shipping cost dependent on the date they choose.'=>'Lass deine Kunden ein günstiges Lieferdatum für die bestellten Produkte wählen und mache die Versandkosten von dem gewählten Datum abhängig.','DPD UK & DPD Local'=>'DPD UK und DPD Local','Ship your DPD orders faster with advanced DPD UK & DPD Local WooCommerce integration. Gather shipping details, download printable shipping labels and track parcels - everything is automated.'=>'Versende deine DPD Bestellungen schneller mit der erweiterten DPD UK & DPD Local WooCommerce Integration. Sammele die Versanddetails, lade die druckbare Versandetiketten herunter und verfolge Pakete - alles ist automatisiert.','Flexible Printing'=>'Flexibler Druck','Automate your shipping process. Print the shipping labels on thermal printers via PrintNode service. Let the labels be printed automatically the same time the order is placed.'=>'Automatisiere deinen Versandprozess. Drucke die Versandetiketten auf Thermodruckern über den PrintNode-Service. Lass die Versandetiketten automatisch drucken, wenn die Bestellung aufgegeben wird.','UPS Labels and Tracking'=>'UPS Versandaufkleber und Sendungsverfolgung','Create the shipments, generate the printable UPS shipping labels for the placed orders and track the parcels directly from your WooCommerce store.'=>'Erstelle die Sendungen, generiere die druckbaren UPS Versandaufkleber für die aufgegebenen Bestellungen und verfolge die Pakete direkt in deinem WooCommerce-Shop.','DHL Express Live Rates PRO'=>'DHL Express WooCommerce Live-Versandtarife PRO','WooCommerce DHL Express integration packed with many advanced features. Display the dynamically calculated live rates for DHL Express shipping methods and adjust them to your needs.'=>'WooCommerce DHL Express Integration verpackt mit vielen erweiterten Funktionen. Zeige die dynamisch berechneten Live-Tarife für DHL Express-Versandmethoden an und passe sie an deine Bedürfnisse an.','Canada Post WooCommerce Plugin - Live Rates PRO'=>'Canada Post WooCommerce Live-Versandtarife PRO','Offer your customers the Canada Post services with real-time calculated shipping rates. Add the handling fees, insurance and adjust them to your needs with just a few clicks.'=>'Biete deinen Kunden die automatisch und in Echtzeit berechneten Canada Post-Versandtarife an. Füge die Bearbeitungsgebühren und die Versicherung hinzu und passe sie mit nur wenigen Klicks an deine Bedürfnisse an.','Royal Mail WooCommerce - Live Rates PRO'=>'Royal Mail WooCommerce Live-Versandtarife PRO','Let your customers choose the Royal Mail shipping methods with the real-time calculated shipping rates. Add the handling fees, insurance and adjust them to your needs in no time.'=>'Lass deine Kunden die Versandmethoden der Royal Mail mit in Echtzeit berechneten Versandkosten wählen. Füge schnell Bearbeitungsgebühren, Versicherungskosten hinzu und passe sie an deine Bedürfnisse an.','Conditional Shipping Methods'=>'Bedingte Versandmethoden','Conditionally display and hide the shipping methods in your WooCommerce store. Define the rules when the specific shipping methods, e.g., live rates should be available to pick and when not to.'=>'Bedinge das Ein-und-Ausblenden der Versandmethoden in deinem WooCommerce-Shop. Definiere die Regeln, wann die spezifischen Versandmethoden, z. B. Live-Tarife, zur Auswahl verfügbar sein sollen und wann nicht.','All'=>'Alle','Bundles'=>'Pakete','Live Rates'=>'Live-Versandtarife','Customizable Rates'=>'Benutzerdefinierter Versandtarif','Shipping Labels'=>'Versandetiketten','Link on plugin list pageExtensions'=>'Erweiterungen','Shipping Extensions by'=>'Versanderweiterungen von','Dive into a system of Octolize ecommerce shipping plugins for WooCommerce. Don’t lose your customers, time and money. Let our plugins secure your sales!'=>'Tauche ein in ein System von Octolize E-Commerce Versand Plugins für WooCommerce. Verliere keine Kunden, Zeit und kein Geld. Lass unsere Plugins deinen Umsatz sichern!','Filter plugins:'=>'Filter Plugins nach:','Blackout Lead Days'=>'Ausschluss von Durchführungstagen','Blackout Lead Days are used to define days of the week when shop is not processing orders.'=>'Diese Option wird verwendet, um Tage in der Woche zu definieren, an denen der Shop keine Bestellungen bearbeitet.','Monday'=>'Montag','Tuesday'=>'Dienstag','Wednesday'=>'Mittwoch','Thursday'=>'Donnerstag','Friday'=>'Freitag','Saturday'=>'Samstag','Sunday'=>'Sonntag','Can\'t convert weight to target unit.'=>'Das Gewicht kann nicht in die Zieleinheit umgerechnet werden.','Please enter USPS User ID.'=>'Bitte die USPS Benutzer-ID eingeben.','Multi-currency is supported only in the USPS Live Rates PRO version. %1$sLearn more →%2$s'=>'Mehrere Währungen werden nur von der USPS Live-Versandtarife PRO Version unterstützt! %1$sMehr erfahren →%2$s','Priority Mail'=>'Priority Mail','Priority Mail Express; Hold For Pickup'=>'Priority Mail Express; Hold For Pickup','Priority Mail Express'=>'Priority Mail Express','Standard Post'=>'Standard Post','Media Mail'=>'Media Mail','Library Mail'=>'Library Mail','Priority Mail Express; Flat Rate Envelope'=>'Priority Mail Express; Flat Rate Envelope','First-Class Mail; Large Postcards'=>'First-Class Mail; Large Postcards','Priority Mail; Flat Rate Envelope'=>'Priority Mail; Flat Rate Envelope','Priority Mail; Medium Flat Rate Box'=>'Priority Mail; Medium Flat Rate Box','Priority Mail; Large Flat Rate Box'=>'Priority Mail; Large Flat Rate Box','Priority Mail Express; Sunday/Holiday Delivery'=>'Priority Mail Express; Sunday/Holiday Delivery','Priority Mail Express; Sunday/Holiday Delivery Flat Rate Envelope'=>'Priority Mail Express; Sunday/Holiday Delivery Flat Rate Envelope','Priority Mail Express; Flat Rate Envelope Hold For Pickup'=>'Priority Mail Express; Flat Rate Envelope Hold For Pickup','Priority Mail; Small Flat Rate Box'=>'Priority Mail; Small Flat Rate Box','Priority Mail; Padded Flat Rate Envelope'=>'Priority Mail; Padded Flat Rate Envelope','Priority Mail Express; Legal Flat Rate Envelope'=>'Priority Mail Express; Legal Flat Rate Envelope','Priority Mail Express; Legal Flat Rate Envelope Hold For Pickup'=>'Priority Mail Express; Legal Flat Rate Envelope Hold For Pickup','Priority Mail Express; Sunday/Holiday Delivery Legal Flat Rate Envelope'=>'Priority Mail Express; Sunday/Holiday Delivery Legal Flat Rate Envelope','Priority Mail; Hold For Pickup'=>'Priority Mail; Hold For Pickup','Priority Mail; Large Flat Rate Box Hold For Pickup'=>'Priority Mail; Large Flat Rate Box Hold For Pickup','Priority Mail; Medium Flat Rate Box Hold For Pickup'=>'Priority Mail; Medium Flat Rate Box Hold For Pickup','Priority Mail; Small Flat Rate Box Hold For Pickup'=>'Priority Mail; Small Flat Rate Box Hold For Pickup','Priority Mail; Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Flat Rate Envelope Hold For Pickup','Priority Mail; Gift Card Flat Rate Envelope'=>'Priority Mail; Gift Card Flat Rate Envelope','Priority Mail; Gift Card Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Gift Card Flat Rate Envelope Hold For Pickup','Priority Mail; Window Flat Rate Envelope'=>'Priority Mail; Window Flat Rate Envelope','Priority Mail; Window Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Window Flat Rate Envelope Hold For Pickup','Priority Mail; Small Flat Rate Envelope'=>'Priority Mail; Small Flat Rate Envelope','Priority Mail; Small Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Small Flat Rate Envelope Hold For Pickup','Priority Mail; Legal Flat Rate Envelope'=>'Priority Mail; Legal Flat Rate Envelope','Priority Mail; Legal Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Legal Flat Rate Envelope Hold For Pickup','Priority Mail; Padded Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Padded Flat Rate Envelope Hold For Pickup','Priority Mail; Regional Rate Box A'=>'Priority Mail; Regional Rate Box A','Priority Mail; Regional Rate Box A Hold For Pickup'=>'Priority Mail; Regional Rate Box A Hold For Pickup','Priority Mail; Regional Rate Box B'=>'Priority Mail; Regional Rate Box B','Priority Mail; Regional Rate Box B Hold For Pickup'=>'Priority Mail; Regional Rate Box B Hold For Pickup','First-Class; Package Service Hold For Pickup'=>'First-Class; Package Service Hold For Pickup','Priority Mail Express; Flat Rate Boxes'=>'Priority Mail Express; Flat Rate Boxes','Priority Mail Express; Flat Rate Boxes Hold For Pickup'=>'Priority Mail Express; Flat Rate Boxes Hold For Pickup','Priority Mail Express; Sunday/Holiday Delivery Flat Rate Boxes'=>'Priority Mail Express; Sunday/Holiday Delivery Flat Rate Boxes','Priority Mail; Regional Rate Box C'=>'Priority Mail; Regional Rate Box C','Priority Mail; Regional Rate Box C Hold For Pickup'=>'Priority Mail; Regional Rate Box C Hold For Pickup','First-Class; Package Service'=>'First-Class; Package Service','Priority Mail Express; Padded Flat Rate Envelope'=>'Priority Mail Express; Padded Flat Rate Envelope','Priority Mail Express; Padded Flat Rate Envelope Hold For Pickup'=>'Priority Mail Express; Padded Flat Rate Envelope Hold For Pickup','Priority Mail Express; Sunday/Holiday Delivery Padded Flat Rate Envelope'=>'Priority Mail Express; Sunday/Holiday Delivery Padded Flat Rate Envelope','Priority Mail 1-Day Return Service Padded Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Padded Flat Rate Envelope','Priority Mail 1-Day Return Service Gift Card Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Gift Card Flat Rate Envelope','Priority Mail 1-Day Return Service Window Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Window Flat Rate Envelope','Priority Mail 1-Day Return Service Small Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Small Flat Rate Envelope','Priority Mail 1-Day Return Service Legal Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Legal Flat Rate Envelope','Priority Mail 1-Day Return Service Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Flat Rate Envelope','Priority Mail 1-Day Return Service Regional Rate Box A'=>'Priority Mail 1-Day Return Service Regional Rate Box A','Priority Mail 1-Day Return Service Regional Rate Box B'=>'Priority Mail 1-Day Return Service Regional Rate Box B','Priority Mail 1-Day Return Service'=>'Priority Mail 1-Day Return Service','Priority Mail 1-Day Return Service Large Flat Rate Box'=>'Priority Mail 1-Day Return Service Large Flat Rate Box','Priority Mail 1-Day Return Service Medium Flat Rate Box'=>'Priority Mail 1-Day Return Service Medium Flat Rate Box','Priority Mail 1-Day Return Service Small Flat Rate Box'=>'Priority Mail 1-Day Return Service Small Flat Rate Box','Priority Mail Military Return Service Large Flat Rate Box APO/FPO/DPO'=>'Priority Mail Military Return Service Large Flat Rate Box APO/FPO/DPO','Priority Mail 1-Day Return Service Cubic'=>'Priority Mail 1-Day Return Service Cubic','First-Class Package Return Service'=>'First-Class Package Return Service','Ground Return Service'=>'Ground Return Service','Priority Mail Express International'=>'Priority Mail Express International','Priority Mail International'=>'Priority Mail International','Global Express Guaranteed; (GXG)**'=>'Global Express Guaranteed; (GXG)**','Global Express Guaranteed; Document'=>'Global Express Guaranteed; Document','Global Express Guarantee; Non-Document Rectangular'=>'Global Express Guarantee; Non-Document Rectangular','Global Express Guaranteed; Non-Document Non-Rectangular'=>'Global Express Guaranteed; Non-Document Non-Rectangular','Priority Mail International; Flat Rate Envelope**'=>'Priority Mail International; Flat Rate Envelope**','Priority Mail International; Medium Flat Rate Box'=>'Priority Mail International; Medium Flat Rate Box','Priority Mail Express International; Flat Rate Envelope'=>'Priority Mail Express International; Flat Rate Envelope','Priority Mail International; Large Flat Rate Box'=>'Priority Mail International; Large Flat Rate Box','USPS GXG; Envelopes**'=>'USPS GXG; Envelopes**','First-Class Mail; International Letter**'=>'First-Class Mail; International Letter**','First-Class Mail; International Large Envelope**'=>'First-Class Mail; International Large Envelope**','First-Class Package International Service**'=>'First-Class Package International Service**','Priority Mail International; Small Flat Rate Box**'=>'Priority Mail International; Small Flat Rate Box**','Priority Mail Express International; Legal Flat Rate Envelope'=>'Priority Mail Express International; Legal Flat Rate Envelope','Priority Mail International; Gift Card Flat Rate Envelope**'=>'Priority Mail International; Gift Card Flat Rate Envelope**','Priority Mail International; Window Flat Rate Envelope**'=>'Priority Mail International; Window Flat Rate Envelope**','Priority Mail International; Small Flat Rate Envelope**'=>'Priority Mail International; Small Flat Rate Envelope**','USPS'=>'USPS','These are the USPS Live Rates plugin general settings. In order to learn more about its configuration please refer to its %1$sdedicated documentation →%2$s'=>'Dies sind die allgemeinen Einstellungen des USPS WooCommerce Live-Versandtarife Plugins. Um mehr über die Konfiguration des Plugins zu erfahren, lies bitte die %1$sDokumentation →%2$s','API Settings'=>'API Einstellungen','Advanced Options'=>'Erweiterte Einstellungen','Debug Mode'=>'Debug-Modus','Enable debug mode'=>'Debug-Modus aktivieren','Enable debug mode to display the additional tech information, incl. the data sent to USPS API, visible only for Admins and Shop Managers in the cart and checkout.'=>'Aktiviere den Debug-Modus, um die technischen Nachrichten im Warenkorb/an der Kasse anzuzeigen. Administratoren und Shop-Manager sehen alle Nachrichten und Daten, die an UPS gesendet werden. Der Kunde sieht nur Nachrichten von der USPS API.','Method Settings'=>'Einstellungen der Methode','Manage the way how the USPS services are displayed in the cart and checkout.'=>'Verwalte wie die USPS-Services im Warenkorb und an der Kasse angezeigt werden.','Method Title'=>'Name der Methode','Define the USPS shipping method title which should be used in the cart/checkout when the Fallback option was triggered.'=>'Definiere den USPS Versandmethodetitel, den der Benutzer im Warenkorb und an der Kasse sieht, wenn Fallback verwendet wird.','USPS Live rates'=>'USPS Live-Tarife','Enable to offer flat rate cost for shipping so that the user can still checkout, if API for some reason returns no matching rates.'=>'Aktiviere die Möglichkeit die Pauschalpreisen für den Versand anzubieten, damit der Benutzer immer noch zur Kasse gehen kann, wenn die API aus irgendeinem Grund keine passenden Preise liefert.','Free Shipping'=>'Kostenloser Versand','Services'=>'USPS Dienstleistungen','Enable the services\' custom settings'=>'Benutzerdefinierte Einstellungen für Dienste aktivieren','Decide which services should be displayed and which not, change their names and order. Please mind that enabling a service does not guarantee it will be visible in the cart/checkout. It has to be available for the provided package weight, origin and destination in order to be displayed.'=>'Entscheide welche Dienstleistungen angezeigt werden sollen, ändere ihre Namen und Reihenfolge. Aktiviere diese Option, wenn du die verfügbaren Dienste auswählen möchtest. Die Aktivierung einer Dienstleistung garantiert nicht, dass sie angeboten wird, weil das Plugin nur die verfügbaren Tarife auf der Grundlage des Paketgewichts, der Absenderadresse und des Ziels anbietet.','Domestic Services Table'=>'Tabelle der USPS-Inlandsdienstleistungen','International Services Table'=>'Tabelle der internationalen USPS Dienstleistungen','Type'=>'Typ','Rates Adjustments'=>'Tarifanpassungen','Adjust these settings to get more accurate rates. Read %swhat affects the USPS rates in USPS WooCommerce plugin →%s'=>'Passe diese Einstellungen an, um genauere Tarife zu erhalten. Lese %swas die USPS Tarife im USPS WooCommerce Plugin beeinflusst →%s','https://octol.io/usps-free-rates'=>'https://octol.io/usps-free-rates','Commercial Rates'=>'Kommerzielle Tarife','Use the USPS Commercial Pricing if available'=>'Verwende die USPS Commercial Pricing, falls verfügbar','Tick this checkbox if you want to use and display the Commercial Pricing shipping rates to your customers instead of the standard ones.'=>'Aktiviere dieses Kontrollkästchen, wenn du die Versandtarife für gewerbliche Kunden anstelle der Standardtarife verwendest und deinen Kunden anzeigen möchten.','Insurance'=>'Versicherung','Request insurance to be included in the USPS rates'=>'Beantrage die Aufnahme der Versicherung in die USPS Tarife','Enable if you want to include insurance in the USPS rates if possible.'=>'Aktiviere diese Option, wenn du eine Versicherung in die USPS Tarife aufnehmen möchtest, sofern diese verfügbar ist.','Empty response!'=>'Leere Antwort!','USPS Live Rates'=>'USPS Live-Tarife','USPS integration'=>'USPS-Integration','None'=>'Keine','Set image'=>'Bild einstellen','Remove image'=>'Bild entfernen','Select or Upload Media'=>'Medien auswählen oder hochladen','Use this media'=>'Dieses Medium verwenden','Search for a product&hellip;'=>'Suche nach einem Produkt&hellip;','Disabled'=>'Deaktiviert','Emergency'=>'Notfall','Alert'=>'Alarm','Critical'=>'Kritisch','Error'=>'Fehler','Warning'=>'Warnung','Notice'=>'Hinweis','Info'=>'Info','Debug'=>'Debug','Can not enable WP Desk Debug log! Cannot create directory %s or this directory is not writeable!'=>'WP Desk Debug-Log kann nicht aktiviert werden! Verzeichnis %s kann nicht erstellt werden, oder dieses Verzeichnis kann nicht geschrieben werden!','Can not enable WP Desk Debug log! Cannot create file %s!'=>'WP Desk Debug-Log kann nicht aktiviert werden! %s Datei kann nicht erstellt werden!','Length %1$s'=>'Länge %1$s','Width %1$s'=>'Breite %1$s','Height %1$s'=>'Höhe %1$s','Max Weight %1$s'=>'Maximales Gewicht %1$s','Padding %1$s'=>'Polsterung %1$s','Box Weight %1$s'=>'Verpackungsgewicht %1$s','Delete'=>'Löschen','Add'=>'Hinzufügen','Select the package type the ordered products will be matched to. You can choose one or as many different packagings as you need. If selected, filling in the products\' weight and dimensions fields is required. %1$sLearn more about the sizes and package types →%2$s'=>'Wähle die Verpackungsart, der die bestellten Produkte zugeordnet werden sollen. Du kannst eine oder so viele verschiedene Verpackungen wählen, wie du benötigst. Falls ausgewählt, müsst du die Felder für Gewicht und Abmessungen der Produkte ausfüllen. %1$sErfahre mehr über die Größen und Verpackungsarten →%2$s','Shipping boxes'=>'Versandverpackungen','Pack into one box by weight'=>'Alle Produkte zusammen in einer Versandverpackung','Pack into custom boxes'=>'In eigenen Versandverpackungen','Pack items separately'=>'Alle Produkte separat verpacken','Pack into custom boxes (3D bin packing)'=>'In eigenen Versandverpackungen (3D-Bin-Verpacken)','Pack into custom boxes (volume packing)'=>'In eigenen Versandverpackungen (nach Volumen)','Parcel Packing Method'=>'Verpackungsmethode','Define the way how the ordered products should be packed. Changing your choice here may affect the rates.'=>'Definiere die Art und Weise, wie die bestellten Produkte verpackt werden sollen. Sollst du unterschiedliche Optionen hier markieren, kann es sich auf die Tarife auswirken.','Cannot pack item: %1$s. %2$s'=>'Dieses Produkt kann nicht gepackt werden: %1$s. %2$s','Cannot pack items: %1$s.'=>'Diese Produkte können nicht gepackt werden: %1$s.','weight'=>'Gewicht','width'=>'Breite','length'=>'Länge','height'=>'Höhe','Item %1$s not set!'=>'Produktdaten %1$s nicht gesetzt!','Custom (%1$sx%2$sx%3$s/%4$s)'=>'Benutzerdefinierter (%1$sx%2$sx%3$s/%4$s)','Dynamically calculated USPS live rates based on the established USPS API connection. %1$sLearn more →%2$s'=>'Dynamisch berechnete USPS Live-Tarife auf der Grundlage der hergestellten USPS API-Verbindung. %1$sMehr erfahren →%2$s','%1$sPlease add now the %2$s shipping method within the shipping zone you want it to be available in →%3$s'=>'%1$sFüge jetzt die %2$s Versandmethode innerhalb der Versandzone hinzu, in der sie verfügbar sein soll →%3$s','API Connection Status'=>'Status der API-Verbindung','Checking...'=>'Wird geprüft…','If you encounter any problems with establishing the API connection, the detailed information on its cause will be displayed here.'=>'Sollten Probleme beim Aufbau der API-Verbindung auftreten, werden hier detaillierte Informationen über die Ursache angezeigt.','OK'=>'OK','Not found HTML view for custom field %1$s.'=>'HTML-Ansicht für benutzerdefiniertes Feld %1$s nicht gefunden.','Fixed value'=>'Festwert','Percentage'=>'Prozentsatz','Use this option to apply the handling fees to the rates. You can use either fixed or percentage values, including the negative ones for discounts.'=>'Verwende diese Option, um die Bearbeitungsgebühren auf die Tarife anzuwenden. Du kannst entweder feste oder prozentuale Werte verwenden, einschließlich der negativen Werte für Rabatte.','Fees Amount'=>'Gebühren betragen','Enter a positive value to charge your customers additionally or a negative one to apply the discount. If you use the currency switcher, the final price incl. Handling Fee will be automatically converted to the currently active currency in your shop. The rates will also include the taxes based on your current Tax settings.'=>'Gib einen positiven Wert ein, um deinen Kunden einen Aufschlag zu berechnen, oder einen negativen Wert, um den Rabatt anzuwenden. Wenn du den Währungsumrechner verwenden, wird der Endpreis inkl. Bearbeitungsgebühr automatisch in die aktuell in deinem Shop aktive Währung umgerechnet. Die Preise enthalten auch die Steuern, die du in deinen aktuellen Steuereinstellungen festgelegt hast.','Code'=>'Code','Name'=>'Name','Enabled'=>'Aktiviert','Drag and drop the services to control their display order. Confirm by clicking Save changes button below.'=>'Ziehe die Dienste per Drag & Drop, um die Reihenfolge der Anzeige zu bestimmen. Bestätige, indem du unten auf die Schaltfläche Änderungen speichern klickst.','Custom Origin'=>'Benutzerdefinierter Ursprung','Origin Address'=>'Absenderadresse','Origin City'=>'Sendeort','Origin Postcode'=>'Versand-Postleitzahl','Origin Country/State'=>'Versandsland','Origin Settings for the whole store'=>'Adres nadania','Origin Settings'=>'Adres nadania','Enable custom origin for the whole store'=>'Benutzerdefinierten Ursprung aktivieren','Enable custom origin'=>'Benutzerdefinierten Ursprung aktivieren','Use this option to use a different shipper\'s address than the one defined in the WooCommerce settings. If not enabled, the one you\'ve entered in WooCommerce → Settings → General → Store Address will be used by default. Each shipping method can have their own custom origin.'=>'Verwende diese Option, um eine andere Versandadresse zu verwenden als die, die in den WooCommerce-Einstellungen definiert ist. Wenn diese Option nicht aktiviert ist, wird standardmäßig die Adresse verwendet, die du unter WooCommerce → Einstellungen → Allgemein → Ladenadresse eingegeben hast.','Use this option to use a different shipper\'s address than the one defined in the WooCommerce settings. If not enabled, the one you\'ve entered in WooCommerce → Settings → General → Store Address will be used by default.'=>'Verwende diese Option, um eine andere Versandadresse zu verwenden als die, die in den WooCommerce-Einstellungen definiert ist. Wenn diese Option nicht aktiviert ist, wird standardmäßig die Adresse verwendet, die du unter WooCommerce → Einstellungen → Allgemein → Ladenadresse eingegeben hast.','Enable custom origin for this shipping method'=>'Benutzerdefinierten Ursprung aktivieren','Cart value exceeds %s. Free shipping has been applied.'=>'Warenkorb-Wert überschreitet %s. Der kostenlose Versand wurde angewendet.','Cart value doesn\'t exceed %s. Free shipping hasn\'t been applied.'=>'Der Warenkorb-Wert übersteigt nicht %s. Der kostenlose Versand wurde nicht angewendet.','Subtotal'=>'Zwischensumme','Free Shipping Amount'=>'Betrag für kostenlosen Versand','Enable the free shipping over amount'=>'Aktiviere den kostenlosen Versand über den Betrag','Tick this checkbox to enter the order amount above which the shipping becomes free.'=>'Aktiviere dieses Kontrollkästchen, um den Bestellwert einzugeben, ab dem der Versand kostenlos wird.','Free Shipping Threshold'=>'Schwellenwert für kostenlosen Versand','Enter only a numeric value without the currency symbol.'=>'Gib nur einen numerischen Wert ohne das Währungssymbol ein.','Fallback reason'=>'Ursache des Fallbacks','Packages'=>'Pakete','Custom'=>'Benutzerdefiniert','No rates added from collection point rates!'=>'Keine Zuschläge von Sammelstellentarifen!','Collection point not found.'=>'Sammelstelle nicht gefunden.','There was no valid services available.'=>'Es waren keine gültigen Dienstleistungen verfügbar.','Fallback rate added with reason: %1$s'=>'Ursache für das Fallback: %1$s','No rates added from standard service rates!'=>'Keine Aufschläge auf die Standardtarife!','Fallback'=>'Fallback','Enable fallback'=>'Fallback einschalten','Fallback Cost'=>'Fallback-Kosten','We noticed that you are using the Germanized plugin along with %1$s. In order to display the accurate rates, please select the WooCommerce default option in the Germanized Taxes > Additional Costs > Tax calculation mode settings. %2$sNeed assistance? Get in touch →%3$s'=>'Wir haben festgestellt, dass du das Germanized Plugin zusammen mit %1$s verwendest. Um die korrekten Tarife anzuzeigen, wähle die WooCommerce Standard in den Germanized Einstellungen Steuern > Nebenkosten > Steuerberechnung. %2$sBrauchst du Hilfe? Nimm Kontakt auf →%3$s','Show %1$s'=>'Anzeigen %1$s','Hide %1$s'=>'Ausblenden %1$s','Copy %1$s to clipboard'=>'Kopieren %1$s in die Zwischenablage','(Delivery Days: 0 days)'=>'(Liefertage: 0 Tage)','(Delivery Days: %1$d day)'=>'(Liefertage: %1$d Tag)' . "\0" . '','(Delivery Date: %1$s)'=>'(Lieferdatum: %1$s)','Awesome, you\'ve been using %s for more than 2 weeks. Could you please do me a BIG favor and give it a 5-star rating on WordPress? ~ Peter'=>'Toll, du hast %s schon seit mehr als 2 Wochen benutzt. Könntest du mir bitte einen GROSSEN Gefallen tun und ihm eine 5-Sterne-Bewertung auf WordPress geben? ~ Peter','%1$sOk, you deserved it%2$s'=>'%1$s Ok, du hast es verdient %2$s','%1$sNope, maybe later%2$s'=>'%1$s Nein, vielleicht später%2$s','%1$sI already did%2$s'=>'%1$sSchon getan%2$s','Created with %1$s by %2$s - If you like %3$s you can %4$srate us %5$s in plugins repository &rarr;%6$s'=>'Erstellt mit %1$s von %2$s - Hat es dir %3$s gefallen, könntest du uns %4$s%5$s in plugins Repository bewerten &rarr;%6$s','The plugin suddenly stopped working'=>'Das PlugIn funktioniert plötzlich nicht mehr','The plugin broke my site'=>'Das Plugin hat meine Seite gestört','I have found a better plugin'=>'Ich habe ein besseres Plugin gefunden','What\'s the plugin\'s name?'=>'Wie heißt das PlugIn?','I only needed the plugin for a short period'=>'Ich brauchte das Plugin nur für einen kurzen Zeitraum','I no longer need the plugin'=>'Ich brauche das Plugin nicht mehr','It\'s a temporary deactivation. I\'m just debugging an issue.'=>'Es ist eine vorübergehende Deaktivierung. Ich debugge gerade ein Problem.','Other'=>'Andere','Please let us know how we can improve our plugin'=>'Bitte teile uns mit, wie wir unser Plugin verbessern können','Plugin deactivation'=>'Plugin-Deaktivierung','You are deactivating %s plugin.'=>'Das %s Plugin wird gerade deaktiviert.','Skip &amp; Deactivate'=>'Überspringen &amp; deaktivieren','Submit &amp; Deactivate'=>'Absenden &amp; Deaktivieren','Cancel'=>'Abbrechen','You successfully opted out of collecting usage data by %1$s. If you change your mind, you can always opt in later in the plugin\'s quick links.'=>'Du hast dich bis %1$s erfolgreich gegen die Erfassung von Nutzungsdaten entschieden. Wenn du deine Meinung änderst, kannst du dich später über die Quicklinks des Plugins wieder anmelden.','Enable usage tracking'=>'Nutzung verfolgen','Disable usage tracking'=>'Nutzung nicht verfolgen','Hey %s,'=>'Hallo %s,','Please help us improve our plugins! If you opt-in, we will collect some non-sensitive data and usage information anonymously. If you skip this, that\'s okay! All plugins will work just fine.'=>'Bitte hilf uns, unsere Plugins zu verbessern! Wenn du zustimmst, sammele wir einige nicht-sensible Daten und Nutzungsinformationen. Wenn du stattdessen überspringst, es ist okay! Alle Plugins werden dennoch ganz normal funktionieren.','Allow & Continue &rarr;'=>'Zulassen & Weiter &rarr;','Skip'=>'Überspringen','What permissions are being granted?'=>'Welche Berechtigungen werden erteilt?','Your Site Overview'=>'Deine Website-Übersicht','WP version, PHP info'=>'WP Version, PHP Info','Plugin Usage'=>'Plugin Nutzung','Current settings and usage information of %1$s plugins'=>'Aktuelle Einstellungen und Nutzungsinformationen von %1$s Plugins','Your Store Overview'=>'Überblick über deinen Shop','Anonymized and non-sensitive store usage information'=>'Anonymisierte und nicht sensible Speichernutzungsinformationen','Find out more &raquo;'=>'Erfahre mehr &raquo;','Current settings and usage information of WP Desk plugins'=>'Aktuelle Einstellungen und Nutzungsinformationen von WP Desk Plugins',' If you have a moment, please let us know why you are deactivating plugin (anonymous feedback):'=>' Wenn du einen Moment Zeit hast, lass uns bitte wissen, warum du das Plugin deaktivierst (anonymes Feedback):','I found a better plugin'=>'Ich habe ein besseres Plugin gefunden','Kindly tell us the reason so we can improve'=>'Nenne bitte den Grund, sodass wir uns stetig verbessern können','We need your help to improve <strong>WP Desk plugins</strong>, so they are more useful for you and the rest of <strong>30,000+ users</strong>. By collecting data on how you use our plugins, you will help us a lot. We will not collect any sensitive data, so you can feel safe.'=>'Wir benötigen Ihre Hilfe, um die <strong>WP-Desk-Plugins</strong> zu verbessern, so dass sie noch nützlicher für dich und den Rest der <strong>30,000+ Benutzer</strong> sind. Indem wir Daten darüber sammeln, wie du unsere Plugins verwendest, hilfst du uns sehr. Wir sorgen dafür, dass keine sensible Daten gesammelt werden.','You successfully opted out of collecting usage data by WP Desk. If you change your mind, you can always opt in later in the plugin\'s quick links.'=>'Du hast erfolgreich angegeben, dass du keine Nutzungsdaten mit WP Desk teilen möchtest. Wenn du deine Meinung ändern solltest, könntest du jederzeit über die Quicklinks des Plugins zustimmen.','Upgrade'=>'Upgrade','Upgrade Now &rarr;'=>'Upgrade jetzt &rarr;','USPS WooCommerce shipping methods with real-time calculated shipping rates based on the established USPS API connection.'=>'USPS WooCommerce-Versandmethoden mit in Echtzeit berechneten Versandtarifen basierend auf der bestehenden USPS API-Verbindung.','You need to provide the USPS account credentials to get live rates. If you do not have the USPS account yet, you can register a new account directly at %1$susps.com →%2$s'=>'Du musst die Anmeldedaten für dein USPS-Konto angeben, um die aktuellen Tarife zu erhalten. Wenn du noch kein USPS-Konto hast, kannst du ein neues Konto direkt auf %1$susps.com%2$s registrieren.','USPS User ID *'=>'USPS Benutzer-ID *','Enter the credentials you acquired during the USPS account registration process.'=>'Gib bitte die Anmeldedaten ein, die du bei der Registrierung des USPS-Kontos erhalten hast.','Test Credentials'=>'Testanmeldeinformationen','Enable to use test credentials'=>'Aktiviere die Verwendung von Test-Anmeldeinformationen','The &#8220;%s&#8221; plugin cannot run on PHP versions older than %s. Please contact your host and ask them to upgrade.'=>'Das &#8220;%s&#8221; Plugin kann nicht mit PHP-Versionen älter als %s ausgeführt werden. Bitte kontaktiere deinen Hoster und bitte ihn darum, ein Upgrade durchzuführen.','The &#8220;%s&#8221; plugin cannot run on WordPress versions older than %s. Please update WordPress.'=>'Das &#8220;%s&#8221; Plugin kann nicht auf WordPress-Versionen ausgeführt werden, die älter als %s sind. Bitte aktualisiere WordPress.','The &#8220;%s&#8221; plugin cannot run on WooCommerce versions older than %s. Please update WooCommerce.'=>'Das &#8220;%s&#8221; Plugin kann nicht auf WooCommerce-Versionen ausgeführt werden, die älter als %s sind. Bitte aktualisiere WooCommerce.','The &#8220;%s&#8221; plugin cannot run without OpenSSL module version at least %s. Please update OpenSSL module.'=>'Das &#8220;%s&#8221; Plugin kann nicht ohne mindestens %s OpenSSL-Modulversion ausgeführt werden. Bitte aktualisiere das OpenSSL-Modul.','The &#8220;%1$s&#8221; plugin requires at least %2$s version of %3$s to work correctly. Please update it to its latest release.'=>'Das Plugin &#8220;%1$s&#8221; benötigt mindestens %2$s Version von %3$s, um korrekt zu funktionieren. Bitte aktualisiere es auf die neueste Version.','The &#8220;%s&#8221; plugin cannot run without %s active. Please install and activate %s plugin.'=>'Das &#8220;%s&#8221; Plugin kann nicht ohne aktives %s ausgeführt werden. Bitte installiere und aktiviere das %s Plugin.','The &#8220;%s&#8221; plugin requires free %s plugin. <a href="%s">Install %s</a>'=>'Das &#8220;%s&#8221; Plugin erfordert ein kostenloses %s Plugin. <a href="%s">Installiere %s</a>','The &#8220;%s&#8221; plugin requires activating %s plugin. <a href="%s">Activate %s</a>'=>'Das &#8220;%s&#8221; Plugin erfordert die Aktivierung des %s Plugins. <a href="%s">Aktiviere %s</a>','The &#8220;%s&#8221; plugin cannot run without %s PHP module installed. Please contact your host and ask them to install %s.'=>'Das &#8220;%s&#8221; Plugin kann nicht laufen, wenn kein %s PHP-Modul installiert ist. Bitte kontaktiere deinen Hoster und bitte ihn, %s zu installieren.','The &#8220;%s&#8221; plugin cannot run without %s PHP setting set to %s. Please contact your host and ask them to set %s.'=>'Das &#8220;%s&#8221; Plugin kann nicht laufen ohne die %s PHP-Einstellung auf %s zu setzen. Bitte kontaktiere deinen Hoster und bitte ihn, %s zu setzen.','The &#8220;%s&#8221; plugin is temporarily disabled since the required %s plugin is being upgraded.'=>'Das Plugin &#8220;%s&#8221; ist vorübergehend deaktiviert, da das erforderliche %s Plugin gerade aktualisiert wird.','Shipping service should implements CanTestSettings interface! Or overwrite this method'=>'Der Versanddienst sollte die CanTestSettings-Schnittstelle implementieren oder überschreiben Sie diese Methode','If you have a moment, please let us know why you are deactivating the plugin (anonymous feedback):'=>'Wenn du einen Moment Zeit hast, lass uns bitte wissen, warum du das Plugin deaktiviert hast (anonymes Feedback):']];
+return ['domain'=>'flexible-shipping-usps','plural-forms'=>'nplurals=2; plural=(n != 1);','language'=>'de_DE','project-id-version'=>'Shipping Live Rates for USPS for WooCommerce 3.1.0','pot-creation-date'=>'2025-07-31T07:25:50+00:00','po-revision-date'=>'YEAR-MO-DA HO:MI+ZONE','x-generator'=>'WP-CLI 2.12.0','messages'=>['https://wordpress.org/plugins/flexible-shipping-usps/'=>'https://de.wordpress.org/plugins/flexible-shipping-usps/','Octolize'=>'Octolize','https://octol.io/usps-author'=>'https://octol.io/usps-author','Get USPS WooCommerce Live Rates PRO!'=>'Hole dir die USPS Live-Versandtarife PRO!','Handling Fees'=>'Bearbeitungsgebühren','Automatic Box Packing'=>'Automatische Verpackungsanpassung','Premium Support'=>'Premium-Support','Multicurrency Support'=>'Unterstützung mehrerer Währungen','Service Code'=>'USPS Servicecode','Live rates for USPS and WooCommerce'=>'USPS WooCommerce Live-Versandtarife','Settings'=>'Einstellungen','Docs'=>'Docs','Support'=>'Support','Help us improve Octolize plugins\' experience'=>'Verbessere mit uns die Erfahrung mit den Octolize-Plugins!','Hi %1$s, with your helping hand we can build effective solutions, launch the new features and shape better plugins experience. By agreeing to anonymously share non-sensitive %2$susage data%3$s of our plugins, you will help us develop them in the right direction. No personal data is tracked or stored and you can opt-out any time. Will you give the thumbs up to our efforts?'=>'Hallo %1$s, mit deiner hilfreichen Hand können wir effektive Lösungen entwickeln, neue Funktionen einführen und die Erfahrung mit den Plugins verbessern. Indem du zustimmst, nicht sensible %2$sNutzungsdaten%3$s unserer Plugins anonym zu teilen, hilfst du uns, sie in die erwünschte Richtung zu entwickeln. Es werden keine persönlichen Daten nachverfolgt oder gespeichert und du kannst sich jederzeit abmelden. Können wir uns auf dich verlassen?','Allow'=>'Erlauben','Page titleShipping Extensions'=>'Versanderweiterungen','Menu TitleShipping
+Extensions'=>'Versanderweiterungen','Buy bundle →'=>'Paket kaufen →','Buy plugin →'=>'Plugin kaufen →','Flexible Shipping PRO'=>'Flexible Shipping PRO','The best and the most powerful Table Rate shipping plugin for WooCommerce. Define the shipping rules based on numerous conditions and configure even the most complex shipping scenarios with ease.'=>'Das beste und leistungsstärkste Table Rate Versand-Plugin für WooCommerce. Definiere die Versandregeln basierend auf zahlreichen Bedingungen und konfiguriere leicht selbst die komplexesten Versandszenarien.','All Plugins Bundle'=>'Alle Plugins-Paket','Grab a pack of all Octolize plugins as a cut-price tailor-made limited offer for developers, agencies and freelancers. Move the WooCommerce shipping to a whole new level. No strings attached, each plugin\'s 25‑sites subscription included.'=>'Hol dir ein Paket mit allen Octolize-Plugins als maßgeschneidertes, preisgünstiges Sonderangebot für Entwickler, Agenturen und Freelancers. Bringe den Versand mit WooCommerce auf ein ganz neues Level. Keine Verpflichtungen, inklusive einem Abonnement für 25 Websites pro Plugin.','Complete UPS Integration Bundle'=>'Vollständiges UPS-Integrationspaket','Connect your WooCommerce store with your UPS account, offer your customers real-time shipping rates and generate printable shipping labels for each placed order.'=>'Verbinde deinen WooCommerce-Shop mit deinem UPS-Konto, biete deinen Kunden Echtzeit-Versandtarife an und generiere druckbare Versandetiketten für jede Bestellung.','Advanced DPD UK Integration Bundle'=>'Erweitertes DPD UK-Integrationspaket','Integrate your WooCommerce store with your DPD UK account, customize the UK locations and automate the whole order processing including generating and printing the shipping labels.'=>'Integriere deinen WooCommerce-Shop mit deinem DPD UK-Konto, passe die Standorte in Großbritannien an und automatisiere den gesamten Bestellvorgang, inklusive der Erstellung und des Ausdrucks der Versandetiketten.','Flexible Shipping Bundle'=>'Flexible Shipping-Paket','Fully customize the shipping cost in your shop, define its calculation rules based on numerous conditions, hide and display the shipping methods and divide orders into separate packages.'=>'Passe die Versandkosten in deinem Shop vollständig an, definiere die Berechnungsregeln basierend auf zahlreichen Bedingungen, verberge und zeige die Versandmethoden und teile Bestellungen in unterschiedliche Pakete auf.','Live Rates Bundle'=>'Live-Versandtarife-Paket','Serve your customers the UPS, FedEx, DHL Express or USPS shipping methods with automatically calculated rates and use multiple conditions to display them according to your terms and needs.'=>'Biete deinen Kunden die Versandmethoden von UPS, FedEx, DHL Express oder USPS mit automatisch berechneten Tarifen an und verwende verschiedene Bedingungen, um sie gemäß deinen Bedingungen und Anforderungen anzuzeigen.','UPS Live Rates and Access Points PRO'=>'UPS WooCommerce Live-Versandtarife und Zugangspunkte PRO','WooCommerce UPS integration packed with many advanced features. Display the dynamically calculated live rates for UPS shipping methods and adjust them to your needs.'=>'WooCommerce UPS-Integration mit vielen erweiterten Funktionen. Zeige die dynamisch berechneten Live-Tarife für UPS-Versandmethoden an und pass an deine Bedürfnisse an.','FedEx WooCommerce Live Rates PRO'=>'FedEx WooCommerce Live-Versandtarife PRO','Enable the FedEx live rates for international delivery and integrate it with your shop in less than 5 minutes. Save your time and money – let the shipping cost be calculated automatically.'=>'Aktiviere die FedEx Live-Tarife für den internationalen Versand und integriere sie in weniger als 5 Minuten in deinen Shop. Spare Zeit und Geld - lass die Versandkosten automatisch berechnen.','USPS Live Rates PRO'=>'USPS WooCommerce Live-Versandtarife PRO','Serve your customers the automatically and real-time calculated USPS shipping rates. Add the handling fees, insurance and adjust them to your needs with just a few clicks.'=>'Biete deinen Kunden die automatisch und in Echtzeit berechneten USPS-Versandtarife an. Füge die Bearbeitungsgebühren und die Versicherung hinzu und passe sie mit nur wenigen Klicks an deine Bedürfnisse an.','Flexible Shipping Import / Export'=>'Flexible Shipping Import / Export','Use the CSV files to import or export your shipping methods. Edit, update, move or backup the ready configurations and shipping scenarios.'=>'Verwende die CSV-Dateien, um deine Versandmethoden zu importieren oder zu exportieren. Bearbeite, aktualisiere, verschiebe oder sichere die fertigen Konfigurationen und Versandszenarien.','Flexible Shipping Locations'=>'Flexible Shipping Standorte','Calculate the shipping cost based on location. Define your own custom locations, use the WooCommerce defaults or the ones created by 3rd party plugins.'=>'Berechne die Versandkosten basierend auf dem Standort. Bestätige deine eigenen benutzerdefinierten Standorte, verwende die WooCommerce-Standardwerte oder die von Drittanbieter-Plugins erstellten Standorten.','Distance Based Shipping Rates'=>'Entfernungsabhängige Versandtarife','Offer shipping rates based on Distance or Total Travel Time calculated by Google Distance Matrix API and don\'t overpay for shipping.'=>'Biete die Versandtarife an, die auf der von Google Distance Matrix API berechneten Entfernung oder Gesamtreisezeit basieren, und überzahle für den Versand nicht mehr.','Pickup Points PRO WooCommerce Plugin'=>'Packstationen PRO WooCommerce-Plugin','Provide your customers with multiple carriers\' pickup points at the checkout and let them choose the preferred one to collect their order from.'=>'Biete deinen Kunden an der Kasse mehrere Packstationen verschiedener Lieferdienste an und lass sie wählen, von welcher sie ihre Bestellung abholen möchten.','Shipping Cost on Product Page PRO'=>'Rechner der Versandkosten auf der Produktseite PRO','Let your customers calculate and see the shipping cost on product pages based on the entered shipping destination and cart contents. Decide how and when exactly you want the shipping cost calculator to display.'=>'Lass deine Kunden die Versandkosten auf den Produktseiten berechnen und anzeigen, basierend auf dem eingegebenen Versandziel und dem Inhalt des Warenkorbs. Entscheide wie und wann genau du den Versandkostenrechner anzeigen lassen möchtest.','Flexible Shipping Box Packing'=>'Flexible Shipping Verpackungsanpassung','Use the advanced box packing WooCommerce algorithm to fit the ordered products into your shipping boxes the most optimal way. Configure the shipping cost based on the type and number of the used shipping boxes.'=>'Verwende den fortgeschrittenen WooCommerce-Algorithmus für das optimale Verpacken der bestellten Produkte in deine Versandkartons. Konfiguriere die Versandkosten basierend auf der Art und Anzahl der verwendeten Versandkartons.','Shipping Packages'=>'Versandpakete','Split the WooCommerce cart content into multiple packages based on various conditions like shipping class.'=>'Teile den Inhalt des WooCommerce-Warenkorbs in mehrere Pakete auf, die auf verschiedenen Bedingungen wie der Versandklasse basieren.','Multi Vendor Shipping'=>'Multi Vendor Versand','Define precisely the shipping cost calculation rules for each Vendor / Product Author in your marketplace or multivendor store.'=>'Definiere genau die Regeln zur Berechnung der Versandkosten für jeden Verkäufer / Produktautor auf deinem Marktplatz oder Multivendor-Shop.','Delivery Date Picker'=>'Lieferterminwähler','Let your customers choose a convenient delivery date for the ordered products and make the shipping cost dependent on the date they choose.'=>'Lass deine Kunden ein günstiges Lieferdatum für die bestellten Produkte wählen und mache die Versandkosten von dem gewählten Datum abhängig.','DPD UK & DPD Local'=>'DPD UK und DPD Local','Ship your DPD orders faster with advanced DPD UK & DPD Local WooCommerce integration. Gather shipping details, download printable shipping labels and track parcels - everything is automated.'=>'Versende deine DPD Bestellungen schneller mit der erweiterten DPD UK & DPD Local WooCommerce Integration. Sammele die Versanddetails, lade die druckbare Versandetiketten herunter und verfolge Pakete - alles ist automatisiert.','Flexible Printing'=>'Flexibler Druck','Automate your shipping process. Print the shipping labels on thermal printers via PrintNode service. Let the labels be printed automatically the same time the order is placed.'=>'Automatisiere deinen Versandprozess. Drucke die Versandetiketten auf Thermodruckern über den PrintNode-Service. Lass die Versandetiketten automatisch drucken, wenn die Bestellung aufgegeben wird.','UPS Labels and Tracking'=>'UPS Versandaufkleber und Sendungsverfolgung','Create the shipments, generate the printable UPS shipping labels for the placed orders and track the parcels directly from your WooCommerce store.'=>'Erstelle die Sendungen, generiere die druckbaren UPS Versandaufkleber für die aufgegebenen Bestellungen und verfolge die Pakete direkt in deinem WooCommerce-Shop.','DHL Express Live Rates PRO'=>'DHL Express WooCommerce Live-Versandtarife PRO','WooCommerce DHL Express integration packed with many advanced features. Display the dynamically calculated live rates for DHL Express shipping methods and adjust them to your needs.'=>'WooCommerce DHL Express Integration verpackt mit vielen erweiterten Funktionen. Zeige die dynamisch berechneten Live-Tarife für DHL Express-Versandmethoden an und passe sie an deine Bedürfnisse an.','Canada Post WooCommerce Plugin - Live Rates PRO'=>'Canada Post WooCommerce Live-Versandtarife PRO','Offer your customers the Canada Post services with real-time calculated shipping rates. Add the handling fees, insurance and adjust them to your needs with just a few clicks.'=>'Biete deinen Kunden die automatisch und in Echtzeit berechneten Canada Post-Versandtarife an. Füge die Bearbeitungsgebühren und die Versicherung hinzu und passe sie mit nur wenigen Klicks an deine Bedürfnisse an.','Royal Mail WooCommerce - Live Rates PRO'=>'Royal Mail WooCommerce Live-Versandtarife PRO','Let your customers choose the Royal Mail shipping methods with the real-time calculated shipping rates. Add the handling fees, insurance and adjust them to your needs in no time.'=>'Lass deine Kunden die Versandmethoden der Royal Mail mit in Echtzeit berechneten Versandkosten wählen. Füge schnell Bearbeitungsgebühren, Versicherungskosten hinzu und passe sie an deine Bedürfnisse an.','Conditional Shipping Methods'=>'Bedingte Versandmethoden','Conditionally display and hide the shipping methods in your WooCommerce store. Define the rules when the specific shipping methods, e.g., live rates should be available to pick and when not to.'=>'Bedinge das Ein-und-Ausblenden der Versandmethoden in deinem WooCommerce-Shop. Definiere die Regeln, wann die spezifischen Versandmethoden, z. B. Live-Tarife, zur Auswahl verfügbar sein sollen und wann nicht.','All'=>'Alle','Bundles'=>'Pakete','Live Rates'=>'Live-Versandtarife','Customizable Rates'=>'Benutzerdefinierter Versandtarif','Shipping Labels'=>'Versandetiketten','Link on plugin list pageExtensions'=>'Erweiterungen','Shipping Extensions by'=>'Versanderweiterungen von','Dive into a system of Octolize ecommerce shipping plugins for WooCommerce. Don’t lose your customers, time and money. Let our plugins secure your sales!'=>'Tauche ein in ein System von Octolize E-Commerce Versand Plugins für WooCommerce. Verliere keine Kunden, Zeit und kein Geld. Lass unsere Plugins deinen Umsatz sichern!','Filter plugins:'=>'Filter Plugins nach:','Blackout Lead Days'=>'Ausschluss von Durchführungstagen','Blackout Lead Days are used to define days of the week when shop is not processing orders.'=>'Diese Option wird verwendet, um Tage in der Woche zu definieren, an denen der Shop keine Bestellungen bearbeitet.','Monday'=>'Montag','Tuesday'=>'Dienstag','Wednesday'=>'Mittwoch','Thursday'=>'Donnerstag','Friday'=>'Freitag','Saturday'=>'Samstag','Sunday'=>'Sonntag','Can\'t convert weight to target unit.'=>'Das Gewicht kann nicht in die Zieleinheit umgerechnet werden.','Please enter USPS User ID.'=>'Bitte die USPS Benutzer-ID eingeben.','Multi-currency is supported only in the USPS Live Rates PRO version. %1$sLearn more →%2$s'=>'Mehrere Währungen werden nur von der USPS Live-Versandtarife PRO Version unterstützt! %1$sMehr erfahren →%2$s','Priority Mail'=>'Priority Mail','Priority Mail Express; Hold For Pickup'=>'Priority Mail Express; Hold For Pickup','Priority Mail Express'=>'Priority Mail Express','Standard Post'=>'Standard Post','Media Mail'=>'Media Mail','Library Mail'=>'Library Mail','Priority Mail Express; Flat Rate Envelope'=>'Priority Mail Express; Flat Rate Envelope','First-Class Mail; Large Postcards'=>'First-Class Mail; Large Postcards','Priority Mail; Flat Rate Envelope'=>'Priority Mail; Flat Rate Envelope','Priority Mail; Medium Flat Rate Box'=>'Priority Mail; Medium Flat Rate Box','Priority Mail; Large Flat Rate Box'=>'Priority Mail; Large Flat Rate Box','Priority Mail Express; Sunday/Holiday Delivery'=>'Priority Mail Express; Sunday/Holiday Delivery','Priority Mail Express; Sunday/Holiday Delivery Flat Rate Envelope'=>'Priority Mail Express; Sunday/Holiday Delivery Flat Rate Envelope','Priority Mail Express; Flat Rate Envelope Hold For Pickup'=>'Priority Mail Express; Flat Rate Envelope Hold For Pickup','Priority Mail; Small Flat Rate Box'=>'Priority Mail; Small Flat Rate Box','Priority Mail; Padded Flat Rate Envelope'=>'Priority Mail; Padded Flat Rate Envelope','Priority Mail Express; Legal Flat Rate Envelope'=>'Priority Mail Express; Legal Flat Rate Envelope','Priority Mail Express; Legal Flat Rate Envelope Hold For Pickup'=>'Priority Mail Express; Legal Flat Rate Envelope Hold For Pickup','Priority Mail Express; Sunday/Holiday Delivery Legal Flat Rate Envelope'=>'Priority Mail Express; Sunday/Holiday Delivery Legal Flat Rate Envelope','Priority Mail; Hold For Pickup'=>'Priority Mail; Hold For Pickup','Priority Mail; Large Flat Rate Box Hold For Pickup'=>'Priority Mail; Large Flat Rate Box Hold For Pickup','Priority Mail; Medium Flat Rate Box Hold For Pickup'=>'Priority Mail; Medium Flat Rate Box Hold For Pickup','Priority Mail; Small Flat Rate Box Hold For Pickup'=>'Priority Mail; Small Flat Rate Box Hold For Pickup','Priority Mail; Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Flat Rate Envelope Hold For Pickup','Priority Mail; Gift Card Flat Rate Envelope'=>'Priority Mail; Gift Card Flat Rate Envelope','Priority Mail; Gift Card Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Gift Card Flat Rate Envelope Hold For Pickup','Priority Mail; Window Flat Rate Envelope'=>'Priority Mail; Window Flat Rate Envelope','Priority Mail; Window Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Window Flat Rate Envelope Hold For Pickup','Priority Mail; Small Flat Rate Envelope'=>'Priority Mail; Small Flat Rate Envelope','Priority Mail; Small Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Small Flat Rate Envelope Hold For Pickup','Priority Mail; Legal Flat Rate Envelope'=>'Priority Mail; Legal Flat Rate Envelope','Priority Mail; Legal Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Legal Flat Rate Envelope Hold For Pickup','Priority Mail; Padded Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Padded Flat Rate Envelope Hold For Pickup','Priority Mail; Regional Rate Box A'=>'Priority Mail; Regional Rate Box A','Priority Mail; Regional Rate Box A Hold For Pickup'=>'Priority Mail; Regional Rate Box A Hold For Pickup','Priority Mail; Regional Rate Box B'=>'Priority Mail; Regional Rate Box B','Priority Mail; Regional Rate Box B Hold For Pickup'=>'Priority Mail; Regional Rate Box B Hold For Pickup','First-Class; Package Service Hold For Pickup'=>'First-Class; Package Service Hold For Pickup','Priority Mail Express; Flat Rate Boxes'=>'Priority Mail Express; Flat Rate Boxes','Priority Mail Express; Flat Rate Boxes Hold For Pickup'=>'Priority Mail Express; Flat Rate Boxes Hold For Pickup','Priority Mail Express; Sunday/Holiday Delivery Flat Rate Boxes'=>'Priority Mail Express; Sunday/Holiday Delivery Flat Rate Boxes','Priority Mail; Regional Rate Box C'=>'Priority Mail; Regional Rate Box C','Priority Mail; Regional Rate Box C Hold For Pickup'=>'Priority Mail; Regional Rate Box C Hold For Pickup','First-Class; Package Service'=>'First-Class; Package Service','Priority Mail Express; Padded Flat Rate Envelope'=>'Priority Mail Express; Padded Flat Rate Envelope','Priority Mail Express; Padded Flat Rate Envelope Hold For Pickup'=>'Priority Mail Express; Padded Flat Rate Envelope Hold For Pickup','Priority Mail Express; Sunday/Holiday Delivery Padded Flat Rate Envelope'=>'Priority Mail Express; Sunday/Holiday Delivery Padded Flat Rate Envelope','Priority Mail 1-Day Return Service Padded Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Padded Flat Rate Envelope','Priority Mail 1-Day Return Service Gift Card Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Gift Card Flat Rate Envelope','Priority Mail 1-Day Return Service Window Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Window Flat Rate Envelope','Priority Mail 1-Day Return Service Small Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Small Flat Rate Envelope','Priority Mail 1-Day Return Service Legal Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Legal Flat Rate Envelope','Priority Mail 1-Day Return Service Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Flat Rate Envelope','Priority Mail 1-Day Return Service Regional Rate Box A'=>'Priority Mail 1-Day Return Service Regional Rate Box A','Priority Mail 1-Day Return Service Regional Rate Box B'=>'Priority Mail 1-Day Return Service Regional Rate Box B','Priority Mail 1-Day Return Service'=>'Priority Mail 1-Day Return Service','Priority Mail 1-Day Return Service Large Flat Rate Box'=>'Priority Mail 1-Day Return Service Large Flat Rate Box','Priority Mail 1-Day Return Service Medium Flat Rate Box'=>'Priority Mail 1-Day Return Service Medium Flat Rate Box','Priority Mail 1-Day Return Service Small Flat Rate Box'=>'Priority Mail 1-Day Return Service Small Flat Rate Box','Priority Mail Military Return Service Large Flat Rate Box APO/FPO/DPO'=>'Priority Mail Military Return Service Large Flat Rate Box APO/FPO/DPO','Priority Mail 1-Day Return Service Cubic'=>'Priority Mail 1-Day Return Service Cubic','First-Class Package Return Service'=>'First-Class Package Return Service','Ground Return Service'=>'Ground Return Service','Priority Mail Express International'=>'Priority Mail Express International','Priority Mail International'=>'Priority Mail International','Global Express Guaranteed; (GXG)**'=>'Global Express Guaranteed; (GXG)**','Global Express Guaranteed; Document'=>'Global Express Guaranteed; Document','Global Express Guarantee; Non-Document Rectangular'=>'Global Express Guarantee; Non-Document Rectangular','Global Express Guaranteed; Non-Document Non-Rectangular'=>'Global Express Guaranteed; Non-Document Non-Rectangular','Priority Mail International; Flat Rate Envelope**'=>'Priority Mail International; Flat Rate Envelope**','Priority Mail International; Medium Flat Rate Box'=>'Priority Mail International; Medium Flat Rate Box','Priority Mail Express International; Flat Rate Envelope'=>'Priority Mail Express International; Flat Rate Envelope','Priority Mail International; Large Flat Rate Box'=>'Priority Mail International; Large Flat Rate Box','USPS GXG; Envelopes**'=>'USPS GXG; Envelopes**','First-Class Mail; International Letter**'=>'First-Class Mail; International Letter**','First-Class Mail; International Large Envelope**'=>'First-Class Mail; International Large Envelope**','First-Class Package International Service**'=>'First-Class Package International Service**','Priority Mail International; Small Flat Rate Box**'=>'Priority Mail International; Small Flat Rate Box**','Priority Mail Express International; Legal Flat Rate Envelope'=>'Priority Mail Express International; Legal Flat Rate Envelope','Priority Mail International; Gift Card Flat Rate Envelope**'=>'Priority Mail International; Gift Card Flat Rate Envelope**','Priority Mail International; Window Flat Rate Envelope**'=>'Priority Mail International; Window Flat Rate Envelope**','Priority Mail International; Small Flat Rate Envelope**'=>'Priority Mail International; Small Flat Rate Envelope**','USPS'=>'USPS','These are the USPS Live Rates plugin general settings. In order to learn more about its configuration please refer to its %1$sdedicated documentation →%2$s'=>'Dies sind die allgemeinen Einstellungen des USPS WooCommerce Live-Versandtarife Plugins. Um mehr über die Konfiguration des Plugins zu erfahren, lies bitte die %1$sDokumentation →%2$s','API Settings'=>'API Einstellungen','Advanced Options'=>'Erweiterte Einstellungen','Debug Mode'=>'Debug-Modus','Enable debug mode'=>'Debug-Modus aktivieren','Enable debug mode to display the additional tech information, incl. the data sent to USPS API, visible only for Admins and Shop Managers in the cart and checkout.'=>'Aktiviere den Debug-Modus, um die technischen Nachrichten im Warenkorb/an der Kasse anzuzeigen. Administratoren und Shop-Manager sehen alle Nachrichten und Daten, die an UPS gesendet werden. Der Kunde sieht nur Nachrichten von der USPS API.','Method Settings'=>'Einstellungen der Methode','Manage the way how the USPS services are displayed in the cart and checkout.'=>'Verwalte wie die USPS-Services im Warenkorb und an der Kasse angezeigt werden.','Method Title'=>'Name der Methode','Define the USPS shipping method title which should be used in the cart/checkout when the Fallback option was triggered.'=>'Definiere den USPS Versandmethodetitel, den der Benutzer im Warenkorb und an der Kasse sieht, wenn Fallback verwendet wird.','USPS Live rates'=>'USPS Live-Tarife','Enable to offer flat rate cost for shipping so that the user can still checkout, if API for some reason returns no matching rates.'=>'Aktiviere die Möglichkeit die Pauschalpreisen für den Versand anzubieten, damit der Benutzer immer noch zur Kasse gehen kann, wenn die API aus irgendeinem Grund keine passenden Preise liefert.','Free Shipping'=>'Kostenloser Versand','Services'=>'USPS Dienstleistungen','Enable the services\' custom settings'=>'Benutzerdefinierte Einstellungen für Dienste aktivieren','Decide which services should be displayed and which not, change their names and order. Please mind that enabling a service does not guarantee it will be visible in the cart/checkout. It has to be available for the provided package weight, origin and destination in order to be displayed.'=>'Entscheide welche Dienstleistungen angezeigt werden sollen, ändere ihre Namen und Reihenfolge. Aktiviere diese Option, wenn du die verfügbaren Dienste auswählen möchtest. Die Aktivierung einer Dienstleistung garantiert nicht, dass sie angeboten wird, weil das Plugin nur die verfügbaren Tarife auf der Grundlage des Paketgewichts, der Absenderadresse und des Ziels anbietet.','Domestic Services Table'=>'Tabelle der USPS-Inlandsdienstleistungen','International Services Table'=>'Tabelle der internationalen USPS Dienstleistungen','Type'=>'Typ','Rates Adjustments'=>'Tarifanpassungen','Adjust these settings to get more accurate rates. Read %swhat affects the USPS rates in USPS WooCommerce plugin →%s'=>'Passe diese Einstellungen an, um genauere Tarife zu erhalten. Lese %swas die USPS Tarife im USPS WooCommerce Plugin beeinflusst →%s','https://octol.io/usps-free-rates'=>'https://octol.io/usps-free-rates','Commercial Rates'=>'Kommerzielle Tarife','Use the USPS Commercial Pricing if available'=>'Verwende die USPS Commercial Pricing, falls verfügbar','Tick this checkbox if you want to use and display the Commercial Pricing shipping rates to your customers instead of the standard ones.'=>'Aktiviere dieses Kontrollkästchen, wenn du die Versandtarife für gewerbliche Kunden anstelle der Standardtarife verwendest und deinen Kunden anzeigen möchten.','Insurance'=>'Versicherung','Request insurance to be included in the USPS rates'=>'Beantrage die Aufnahme der Versicherung in die USPS Tarife','Enable if you want to include insurance in the USPS rates if possible.'=>'Aktiviere diese Option, wenn du eine Versicherung in die USPS Tarife aufnehmen möchtest, sofern diese verfügbar ist.','Empty response!'=>'Leere Antwort!','USPS Live Rates'=>'USPS Live-Tarife','USPS integration'=>'USPS-Integration','None'=>'Keine','The &#8220;%s&#8221; plugin cannot run on PHP versions older than %s. Please contact your host and ask them to upgrade.'=>'Das &#8220;%s&#8221; Plugin kann nicht mit PHP-Versionen älter als %s ausgeführt werden. Bitte kontaktiere deinen Hoster und bitte ihn darum, ein Upgrade durchzuführen.','The &#8220;%s&#8221; plugin cannot run on WordPress versions older than %s. Please update WordPress.'=>'Das &#8220;%s&#8221; Plugin kann nicht auf WordPress-Versionen ausgeführt werden, die älter als %s sind. Bitte aktualisiere WordPress.','The &#8220;%s&#8221; plugin cannot run on WooCommerce versions older than %s. Please update WooCommerce.'=>'Das &#8220;%s&#8221; Plugin kann nicht auf WooCommerce-Versionen ausgeführt werden, die älter als %s sind. Bitte aktualisiere WooCommerce.','The &#8220;%s&#8221; plugin cannot run without OpenSSL module version at least %s. Please update OpenSSL module.'=>'Das &#8220;%s&#8221; Plugin kann nicht ohne mindestens %s OpenSSL-Modulversion ausgeführt werden. Bitte aktualisiere das OpenSSL-Modul.','The &#8220;%1$s&#8221; plugin requires at least %2$s version of %3$s to work correctly. Please update it to its latest release.'=>'Das Plugin &#8220;%1$s&#8221; benötigt mindestens %2$s Version von %3$s, um korrekt zu funktionieren. Bitte aktualisiere es auf die neueste Version.','The &#8220;%s&#8221; plugin cannot run without %s active. Please install and activate %s plugin.'=>'Das &#8220;%s&#8221; Plugin kann nicht ohne aktives %s ausgeführt werden. Bitte installiere und aktiviere das %s Plugin.','The &#8220;%s&#8221; plugin requires free %s plugin. <a href="%s">Install %s</a>'=>'Das &#8220;%s&#8221; Plugin erfordert ein kostenloses %s Plugin. <a href="%s">Installiere %s</a>','The &#8220;%s&#8221; plugin requires activating %s plugin. <a href="%s">Activate %s</a>'=>'Das &#8220;%s&#8221; Plugin erfordert die Aktivierung des %s Plugins. <a href="%s">Aktiviere %s</a>','The &#8220;%s&#8221; plugin cannot run without %s PHP module installed. Please contact your host and ask them to install %s.'=>'Das &#8220;%s&#8221; Plugin kann nicht laufen, wenn kein %s PHP-Modul installiert ist. Bitte kontaktiere deinen Hoster und bitte ihn, %s zu installieren.','The &#8220;%s&#8221; plugin cannot run without %s PHP setting set to %s. Please contact your host and ask them to set %s.'=>'Das &#8220;%s&#8221; Plugin kann nicht laufen ohne die %s PHP-Einstellung auf %s zu setzen. Bitte kontaktiere deinen Hoster und bitte ihn, %s zu setzen.','The &#8220;%s&#8221; plugin is temporarily disabled since the required %s plugin is being upgraded.'=>'Das Plugin &#8220;%s&#8221; ist vorübergehend deaktiviert, da das erforderliche %s Plugin gerade aktualisiert wird.','Set image'=>'Bild einstellen','Remove image'=>'Bild entfernen','Select or Upload Media'=>'Medien auswählen oder hochladen','Use this media'=>'Dieses Medium verwenden','Search for a product&hellip;'=>'Suche nach einem Produkt&hellip;','Disabled'=>'Deaktiviert','Emergency'=>'Notfall','Alert'=>'Alarm','Critical'=>'Kritisch','Error'=>'Fehler','Warning'=>'Warnung','Notice'=>'Hinweis','Info'=>'Info','Debug'=>'Debug','Can not enable WP Desk Debug log! Cannot create directory %s or this directory is not writeable!'=>'WP Desk Debug-Log kann nicht aktiviert werden! Verzeichnis %s kann nicht erstellt werden, oder dieses Verzeichnis kann nicht geschrieben werden!','Can not enable WP Desk Debug log! Cannot create file %s!'=>'WP Desk Debug-Log kann nicht aktiviert werden! %s Datei kann nicht erstellt werden!','Length %1$s'=>'Länge %1$s','Width %1$s'=>'Breite %1$s','Height %1$s'=>'Höhe %1$s','Max Weight %1$s'=>'Maximales Gewicht %1$s','Padding %1$s'=>'Polsterung %1$s','Box Weight %1$s'=>'Verpackungsgewicht %1$s','Delete'=>'Löschen','Add'=>'Hinzufügen','Select the package type the ordered products will be matched to. You can choose one or as many different packagings as you need. If selected, filling in the products\' weight and dimensions fields is required. %1$sLearn more about the sizes and package types →%2$s'=>'Wähle die Verpackungsart, der die bestellten Produkte zugeordnet werden sollen. Du kannst eine oder so viele verschiedene Verpackungen wählen, wie du benötigst. Falls ausgewählt, müsst du die Felder für Gewicht und Abmessungen der Produkte ausfüllen. %1$sErfahre mehr über die Größen und Verpackungsarten →%2$s','Shipping boxes'=>'Versandverpackungen','Pack into one box by weight'=>'Alle Produkte zusammen in einer Versandverpackung','Pack into custom boxes (3D bin packing)'=>'In eigenen Versandverpackungen (3D-Bin-Verpacken)','Pack into custom boxes (volume packing)'=>'In eigenen Versandverpackungen (nach Volumen)','Pack items separately'=>'Alle Produkte separat verpacken','Parcel Packing Method'=>'Verpackungsmethode','Define the way how the ordered products should be packed. Changing your choice here may affect the rates.'=>'Definiere die Art und Weise, wie die bestellten Produkte verpackt werden sollen. Sollst du unterschiedliche Optionen hier markieren, kann es sich auf die Tarife auswirken.','Cannot pack item: %1$s. %2$s'=>'Dieses Produkt kann nicht gepackt werden: %1$s. %2$s','Cannot pack items: %1$s.'=>'Diese Produkte können nicht gepackt werden: %1$s.','weight'=>'Gewicht','width'=>'Breite','length'=>'Länge','height'=>'Höhe','Item %1$s not set!'=>'Produktdaten %1$s nicht gesetzt!','Custom (%1$sx%2$sx%3$s/%4$s)'=>'Benutzerdefinierter (%1$sx%2$sx%3$s/%4$s)','Dynamically calculated USPS live rates based on the established USPS API connection. %1$sLearn more →%2$s'=>'Dynamisch berechnete USPS Live-Tarife auf der Grundlage der hergestellten USPS API-Verbindung. %1$sMehr erfahren →%2$s','%1$sPlease add now the %2$s shipping method within the shipping zone you want it to be available in →%3$s'=>'%1$sFüge jetzt die %2$s Versandmethode innerhalb der Versandzone hinzu, in der sie verfügbar sein soll →%3$s','API Connection Status'=>'Status der API-Verbindung','Checking...'=>'Wird geprüft…','If you encounter any problems with establishing the API connection, the detailed information on its cause will be displayed here.'=>'Sollten Probleme beim Aufbau der API-Verbindung auftreten, werden hier detaillierte Informationen über die Ursache angezeigt.','OK'=>'OK','Not found HTML view for custom field %1$s.'=>'HTML-Ansicht für benutzerdefiniertes Feld %1$s nicht gefunden.','Fixed value'=>'Festwert','Percentage'=>'Prozentsatz','Use this option to apply the handling fees to the rates. You can use either fixed or percentage values, including the negative ones for discounts.'=>'Verwende diese Option, um die Bearbeitungsgebühren auf die Tarife anzuwenden. Du kannst entweder feste oder prozentuale Werte verwenden, einschließlich der negativen Werte für Rabatte.','Fees Amount'=>'Gebühren betragen','Enter a positive value to charge your customers additionally or a negative one to apply the discount. If you use the currency switcher, the final price incl. Handling Fee will be automatically converted to the currently active currency in your shop. The rates will also include the taxes based on your current Tax settings.'=>'Gib einen positiven Wert ein, um deinen Kunden einen Aufschlag zu berechnen, oder einen negativen Wert, um den Rabatt anzuwenden. Wenn du den Währungsumrechner verwenden, wird der Endpreis inkl. Bearbeitungsgebühr automatisch in die aktuell in deinem Shop aktive Währung umgerechnet. Die Preise enthalten auch die Steuern, die du in deinen aktuellen Steuereinstellungen festgelegt hast.','Code'=>'Code','Name'=>'Name','Enabled'=>'Aktiviert','Drag and drop the services to control their display order. Confirm by clicking Save changes button below.'=>'Ziehe die Dienste per Drag & Drop, um die Reihenfolge der Anzeige zu bestimmen. Bestätige, indem du unten auf die Schaltfläche Änderungen speichern klickst.','Custom Origin'=>'Benutzerdefinierter Ursprung','Origin Address'=>'Absenderadresse','Origin City'=>'Sendeort','Origin Postcode'=>'Versand-Postleitzahl','Origin Country/State'=>'Versandsland','Origin Settings for the whole store'=>'Adres nadania','Origin Settings'=>'Adres nadania','Enable custom origin for the whole store'=>'Benutzerdefinierten Ursprung aktivieren','Enable custom origin'=>'Benutzerdefinierten Ursprung aktivieren','Use this option to use a different shipper\'s address than the one defined in the WooCommerce settings. If not enabled, the one you\'ve entered in WooCommerce → Settings → General → Store Address will be used by default. Each shipping method can have their own custom origin.'=>'Verwende diese Option, um eine andere Versandadresse zu verwenden als die, die in den WooCommerce-Einstellungen definiert ist. Wenn diese Option nicht aktiviert ist, wird standardmäßig die Adresse verwendet, die du unter WooCommerce → Einstellungen → Allgemein → Ladenadresse eingegeben hast.','Use this option to use a different shipper\'s address than the one defined in the WooCommerce settings. If not enabled, the one you\'ve entered in WooCommerce → Settings → General → Store Address will be used by default.'=>'Verwende diese Option, um eine andere Versandadresse zu verwenden als die, die in den WooCommerce-Einstellungen definiert ist. Wenn diese Option nicht aktiviert ist, wird standardmäßig die Adresse verwendet, die du unter WooCommerce → Einstellungen → Allgemein → Ladenadresse eingegeben hast.','Enable custom origin for this shipping method'=>'Benutzerdefinierten Ursprung aktivieren','Cart value exceeds %s. Free shipping has been applied.'=>'Warenkorb-Wert überschreitet %s. Der kostenlose Versand wurde angewendet.','Cart value doesn\'t exceed %s. Free shipping hasn\'t been applied.'=>'Der Warenkorb-Wert übersteigt nicht %s. Der kostenlose Versand wurde nicht angewendet.','Subtotal'=>'Zwischensumme','Free Shipping Amount'=>'Betrag für kostenlosen Versand','Enable the free shipping over amount'=>'Aktiviere den kostenlosen Versand über den Betrag','Tick this checkbox to enter the order amount above which the shipping becomes free.'=>'Aktiviere dieses Kontrollkästchen, um den Bestellwert einzugeben, ab dem der Versand kostenlos wird.','Free Shipping Threshold'=>'Schwellenwert für kostenlosen Versand','Enter only a numeric value without the currency symbol.'=>'Gib nur einen numerischen Wert ohne das Währungssymbol ein.','Show %1$s'=>'Anzeigen %1$s','Hide %1$s'=>'Ausblenden %1$s','Copy %1$s to clipboard'=>'Kopieren %1$s in die Zwischenablage','Fallback reason'=>'Ursache des Fallbacks','Packages'=>'Pakete','Custom'=>'Benutzerdefiniert','No rates added from collection point rates!'=>'Keine Zuschläge von Sammelstellentarifen!','Collection point not found.'=>'Sammelstelle nicht gefunden.','There was no valid services available.'=>'Es waren keine gültigen Dienstleistungen verfügbar.','Fallback rate added with reason: %1$s'=>'Ursache für das Fallback: %1$s','No rates added from standard service rates!'=>'Keine Aufschläge auf die Standardtarife!','Fallback'=>'Fallback','Enable fallback'=>'Fallback einschalten','Fallback Cost'=>'Fallback-Kosten','We noticed that you are using the Germanized plugin along with %1$s. In order to display the accurate rates, please select the WooCommerce default option in the Germanized Taxes > Additional Costs > Tax calculation mode settings. %2$sNeed assistance? Get in touch →%3$s'=>'Wir haben festgestellt, dass du das Germanized Plugin zusammen mit %1$s verwendest. Um die korrekten Tarife anzuzeigen, wähle die WooCommerce Standard in den Germanized Einstellungen Steuern > Nebenkosten > Steuerberechnung. %2$sBrauchst du Hilfe? Nimm Kontakt auf →%3$s','(Delivery Days: 0 days)'=>'(Liefertage: 0 Tage)','(Delivery Days: %1$d day)'=>'(Liefertage: %1$d Tag)' . "\0" . '','(Delivery Date: %1$s)'=>'(Lieferdatum: %1$s)','Awesome, you\'ve been using %s for more than 2 weeks. Could you please do me a BIG favor and give it a 5-star rating on WordPress? ~ Peter'=>'Toll, du hast %s schon seit mehr als 2 Wochen benutzt. Könntest du mir bitte einen GROSSEN Gefallen tun und ihm eine 5-Sterne-Bewertung auf WordPress geben? ~ Peter','%1$sOk, you deserved it%2$s'=>'%1$s Ok, du hast es verdient %2$s','%1$sNope, maybe later%2$s'=>'%1$s Nein, vielleicht später%2$s','%1$sI already did%2$s'=>'%1$sSchon getan%2$s','Created with %1$s by %2$s - If you like %3$s you can %4$srate us %5$s in plugins repository &rarr;%6$s'=>'Erstellt mit %1$s von %2$s - Hat es dir %3$s gefallen, könntest du uns %4$s%5$s in plugins Repository bewerten &rarr;%6$s','The plugin suddenly stopped working'=>'Das PlugIn funktioniert plötzlich nicht mehr','The plugin broke my site'=>'Das Plugin hat meine Seite gestört','I have found a better plugin'=>'Ich habe ein besseres Plugin gefunden','What\'s the plugin\'s name?'=>'Wie heißt das PlugIn?','I only needed the plugin for a short period'=>'Ich brauchte das Plugin nur für einen kurzen Zeitraum','I no longer need the plugin'=>'Ich brauche das Plugin nicht mehr','It\'s a temporary deactivation. I\'m just debugging an issue.'=>'Es ist eine vorübergehende Deaktivierung. Ich debugge gerade ein Problem.','Other'=>'Andere','Please let us know how we can improve our plugin'=>'Bitte teile uns mit, wie wir unser Plugin verbessern können','Plugin deactivation'=>'Plugin-Deaktivierung','You are deactivating %s plugin.'=>'Das %s Plugin wird gerade deaktiviert.','Skip &amp; Deactivate'=>'Überspringen &amp; deaktivieren','Submit &amp; Deactivate'=>'Absenden &amp; Deaktivieren','Cancel'=>'Abbrechen','You successfully opted out of collecting usage data by %1$s. If you change your mind, you can always opt in later in the plugin\'s quick links.'=>'Du hast dich bis %1$s erfolgreich gegen die Erfassung von Nutzungsdaten entschieden. Wenn du deine Meinung änderst, kannst du dich später über die Quicklinks des Plugins wieder anmelden.','Enable usage tracking'=>'Nutzung verfolgen','Disable usage tracking'=>'Nutzung nicht verfolgen','Hey %s,'=>'Hallo %s,','Please help us improve our plugins! If you opt-in, we will collect some non-sensitive data and usage information anonymously. If you skip this, that\'s okay! All plugins will work just fine.'=>'Bitte hilf uns, unsere Plugins zu verbessern! Wenn du zustimmst, sammele wir einige nicht-sensible Daten und Nutzungsinformationen. Wenn du stattdessen überspringst, es ist okay! Alle Plugins werden dennoch ganz normal funktionieren.','Allow & Continue &rarr;'=>'Zulassen & Weiter &rarr;','Skip'=>'Überspringen','What permissions are being granted?'=>'Welche Berechtigungen werden erteilt?','Your Site Overview'=>'Deine Website-Übersicht','WP version, PHP info'=>'WP Version, PHP Info','Plugin Usage'=>'Plugin Nutzung','Current settings and usage information of %1$s plugins'=>'Aktuelle Einstellungen und Nutzungsinformationen von %1$s Plugins','Your Store Overview'=>'Überblick über deinen Shop','Anonymized and non-sensitive store usage information'=>'Anonymisierte und nicht sensible Speichernutzungsinformationen','Find out more &raquo;'=>'Erfahre mehr &raquo;','Current settings and usage information of WP Desk plugins'=>'Aktuelle Einstellungen und Nutzungsinformationen von WP Desk Plugins',' If you have a moment, please let us know why you are deactivating plugin (anonymous feedback):'=>' Wenn du einen Moment Zeit hast, lass uns bitte wissen, warum du das Plugin deaktivierst (anonymes Feedback):','I found a better plugin'=>'Ich habe ein besseres Plugin gefunden','Kindly tell us the reason so we can improve'=>'Nenne bitte den Grund, sodass wir uns stetig verbessern können','We need your help to improve <strong>WP Desk plugins</strong>, so they are more useful for you and the rest of <strong>30,000+ users</strong>. By collecting data on how you use our plugins, you will help us a lot. We will not collect any sensitive data, so you can feel safe.'=>'Wir benötigen Ihre Hilfe, um die <strong>WP-Desk-Plugins</strong> zu verbessern, so dass sie noch nützlicher für dich und den Rest der <strong>30,000+ Benutzer</strong> sind. Indem wir Daten darüber sammeln, wie du unsere Plugins verwendest, hilfst du uns sehr. Wir sorgen dafür, dass keine sensible Daten gesammelt werden.','You successfully opted out of collecting usage data by WP Desk. If you change your mind, you can always opt in later in the plugin\'s quick links.'=>'Du hast erfolgreich angegeben, dass du keine Nutzungsdaten mit WP Desk teilen möchtest. Wenn du deine Meinung ändern solltest, könntest du jederzeit über die Quicklinks des Plugins zustimmen.','Upgrade'=>'Upgrade','Upgrade Now &rarr;'=>'Upgrade jetzt &rarr;','USPS WooCommerce shipping methods with real-time calculated shipping rates based on the established USPS API connection.'=>'USPS WooCommerce-Versandmethoden mit in Echtzeit berechneten Versandtarifen basierend auf der bestehenden USPS API-Verbindung.','You need to provide the USPS account credentials to get live rates. If you do not have the USPS account yet, you can register a new account directly at %1$susps.com →%2$s'=>'Du musst die Anmeldedaten für dein USPS-Konto angeben, um die aktuellen Tarife zu erhalten. Wenn du noch kein USPS-Konto hast, kannst du ein neues Konto direkt auf %1$susps.com%2$s registrieren.','USPS User ID *'=>'USPS Benutzer-ID *','Enter the credentials you acquired during the USPS account registration process.'=>'Gib bitte die Anmeldedaten ein, die du bei der Registrierung des USPS-Kontos erhalten hast.','Test Credentials'=>'Testanmeldeinformationen','Enable to use test credentials'=>'Aktiviere die Verwendung von Test-Anmeldeinformationen','Pack into custom boxes'=>'In eigenen Versandverpackungen','Shipping service should implements CanTestSettings interface! Or overwrite this method'=>'Der Versanddienst sollte die CanTestSettings-Schnittstelle implementieren oder überschreiben Sie diese Methode','If you have a moment, please let us know why you are deactivating the plugin (anonymous feedback):'=>'Wenn du einen Moment Zeit hast, lass uns bitte wissen, warum du das Plugin deaktiviert hast (anonymes Feedback):']];
