@@ -1,4 +1,4 @@
 <?php
-return ['domain'=>'flexible-shipping-usps','plural-forms'=>'nplurals=3; plural=(n==1 ? 0 : n%10>=2 && n%10<=4 && (n%100<12 || n%100>14) ? 1 : 2);','language'=>'pl_PL','project-id-version'=>'Shipping Live Rates for USPS for WooCommerce 3.0.7','pot-creation-date'=>'2025-06-23T23:09:39+00:00','po-revision-date'=>'YEAR-MO-DA HO:MI+ZONE','x-generator'=>'WP-CLI 2.11.0','messages'=>['https://wordpress.org/plugins/flexible-shipping-usps/'=>'https://pl.wordpress.org/plugins/flexible-shipping-usps/','Octolize'=>'Octolize','https://octol.io/usps-author'=>'https://octol.io/usps-author-pl','Get USPS WooCommerce Live Rates PRO!'=>'Kup USPS Stawki live PRO WooCommerce!','Handling Fees'=>'Prowizje dla stawek','Automatic Box Packing'=>'Podział na paczki','Premium Support'=>'Priorytetowe wsparcie techniczne','Multicurrency Support'=>'Obsługa wielu walut','Service Code'=>'Kod usługi','Live rates for USPS and WooCommerce'=>'USPS Stawki live WooCommerce','Settings'=>'Ustawienia','Docs'=>'Docs','Support'=>'Wsparcie','Can you let us know, what functionality you\'re looking for?'=>'Czy możesz opisać, jakiej funkcjonalności potrzebujesz?','I had difficulties configuring the plugin'=>'Miałem trudności w konfiguracji wtyczki','Sorry to hear that! We\'re certain that with a little help, configuring the plugin will be a breeze. Before you deactivate, try to find a solution in our %1$sdocumentation%2$s%3$s.'=>'Mamy pewność, że przy odrobinie pomocy konfiguracja wtyczki będzie bezproblemowa. Zanim dokończysz dezaktywację spróbuj znaleźć rozwiązanie w naszej %1$sdokumentacji%2$s%3$s.',' or post a question on the %1$sforum%2$s'=>' lub zadaj pytanie na %1$sforum%2$s','The plugin stopped working'=>'Wtyczka przestała działać','We take any issues with our plugins very seriously. Try to find a reason in our %1$sdocumentation%2$s%3$s.'=>'Poważnie traktujemy problemy związane z użytkowaniem naszych wtyczek Spróbuj znaleźć rozwiązanie w %1$sdokumentacji%2$s%3$s.',' or post the problem on the %1$sforum%2$s'=>' lub zadaj pytanie na %1$sforum%2$s','I have found another plugin'=>'Znalazłem inną wtyczkę','That hurts a little bit, but we\'re tough! Can you let us know which plugin you are switching to?'=>'Czy możesz nam powiedzieć, z jakiej wtyczki będziesz teraz korzystał?','Which plugin are you switching to?'=>'Z jakiej wtyczki będziesz teraz korzystał?','The plugin doesn\'t have the functionality I need'=>'Wtyczka nie posiada funkcjonalności, której potrzebuję','Good news! There\'s a great chance that the functionality you need is already implemented in the PRO version of the plugin. %1$sContact us%2$s to receive a discount for %3$s. Also, can you describe what functionality you\'re looking for?'=>'Dobra wiadomość! Jest duża szansa, że funkcjonalność, której potrzebujesz, jest już zaimplementowana w wersji PRO wtyczki. %1$sSkontaktuj się z nami%2$s, aby otrzymać zniżkę na %3$s. Ponadto, czy możesz opisać, jakiej funkcjonalności poszukujesz?','We\'re sorry to hear that. Can you describe what functionality you\'re looking for?'=>'Przykro nam to słyszeć! Jakiej funkcjonalności potrzebujesz?','What functionality are you looking for?'=>'Jakiej funkcjonalności potrzebujesz?','I don\'t need the plugin anymore'=>'Nie potrzebuję już wtyczki','Sorry to hear that! Can you let us know why the plugin is not needed anymore?'=>'Przykro nam to słyszeć! Czy możesz nam powiedzieć, dlaczego już nie potrzebujesz wtyczki?','Why is the plugin not needed anymore?'=>'Dlaczego już nie potrzebujesz wtyczki?','I\'m deactivating temporarily for debugging purposes'=>'Dezaktywuję tymczasowo w celu debugowania','Other reason'=>'Inny powód','Can you provide some details on the reason behind deactivation?'=>'Czy możesz podać kilka szczegółów dotyczących powodu dezaktywacji?','Please provide details'=>'Napisz dlaczego dezaktywujesz wtyczkę','Help us improve Octolize plugins\' experience'=>'Pomóż nam ulepszyć i rozwijać wtyczki Octolize','Hi %1$s, with your helping hand we can build effective solutions, launch the new features and shape better plugins experience. By agreeing to anonymously share non-sensitive %2$susage data%3$s of our plugins, you will help us develop them in the right direction. No personal data is tracked or stored and you can opt-out any time. Will you give the thumbs up to our efforts?'=>'Cześć %1$s, z Twoją pomocą możemy tworzyć skuteczne rozwiązania, wdrażać nowe funkcjonalności i stale ulepszać nasze wtyczki. Zgadzając się na udostępnianie anonimowo niewrażliwych %2$sdanych użytkowania%3$s wtyczek, pomożesz nam rozwijać je we właściwym kierunku. Żadne dane osobowe nie są udostępniane ani przechowywane i możesz zrezygnować w dowolnym momencie. Czy możemy na Ciebie liczyć?','Allow'=>'Pozwól','We\'ve added integration with Flexible Shipping Table Rate'=>'Dodaliśmy integrację z Flexible Shipping Table Rate','The new feature allows you to combine Live Rates and Table Rate, providing the ability to use automatically calculated Live Rates while also adjusting shipping costs using Table Rate with the free version of %1$sFlexible Shipping%2$s or %3$sFlexible Shipping PRO%4$s.%5$sThis way, you can now have greater control over the final delivery cost, precisely defining additional charges or discounts for each shipping method.'=>'Nowa funkcja pozwala połączyć Live Rates i Table Rate, dając możliwość skorzystania z automatycznie kalkulowanych stawek Live Rates, a jednocześnie dostosowania kosztów wysyłki za pomocą Table Rate z wykorzystaniem darmowej wersji %1$sFlexible Shipping%2$s lub %3$sFlexible Shipping PRO%4$s.%5$sDzięki temu, możesz mieć teraz większą kontrolę nad finalnym kosztem dostawy, mogąc precyzyjnie określać dodatkowe opłaty lub zniżki na każdej metodzie wysyłki.','Thank you for updating %1$s!'=>'Dziękujemy za aktualizację %1$s!','It is really important to keep the plugins up to date. We have implemented some improvements and new functionalities. Find out what has changed:'=>'To ważne, by utrzymywać wszystkie pluginy w najbardziej aktualnej wersji. Wprowadziliśmy pewne usprawnienia i nowe funkcjonalności. Zobacz co się zmieniło:','I\'m not interested'=>'Nie jestem zainteresowany/-a','Thanks for letting me know'=>'Dzięki za informację!','Page titleShipping Extensions'=>'Rozszerzenia wysyłki','Menu TitleShipping
+return ['domain'=>'flexible-shipping-usps','plural-forms'=>'nplurals=3; plural=(n==1 ? 0 : n%10>=2 && n%10<=4 && (n%100<12 || n%100>14) ? 1 : 2);','language'=>'pl_PL','project-id-version'=>'Shipping Live Rates for USPS for WooCommerce 3.1.0','pot-creation-date'=>'2025-07-31T07:25:50+00:00','po-revision-date'=>'YEAR-MO-DA HO:MI+ZONE','x-generator'=>'WP-CLI 2.12.0','messages'=>['https://wordpress.org/plugins/flexible-shipping-usps/'=>'https://pl.wordpress.org/plugins/flexible-shipping-usps/','Octolize'=>'Octolize','https://octol.io/usps-author'=>'https://octol.io/usps-author-pl','Get USPS WooCommerce Live Rates PRO!'=>'Kup USPS Stawki live PRO WooCommerce!','Handling Fees'=>'Prowizje dla stawek','Automatic Box Packing'=>'Podział na paczki','Premium Support'=>'Priorytetowe wsparcie techniczne','Multicurrency Support'=>'Obsługa wielu walut','Service Code'=>'Kod usługi','Live rates for USPS and WooCommerce'=>'USPS Stawki live WooCommerce','Settings'=>'Ustawienia','Docs'=>'Docs','Support'=>'Wsparcie','Can you let us know, what functionality you\'re looking for?'=>'Czy możesz opisać, jakiej funkcjonalności potrzebujesz?','I had difficulties configuring the plugin'=>'Miałem trudności w konfiguracji wtyczki','Sorry to hear that! We\'re certain that with a little help, configuring the plugin will be a breeze. Before you deactivate, try to find a solution in our %1$sdocumentation%2$s%3$s.'=>'Mamy pewność, że przy odrobinie pomocy konfiguracja wtyczki będzie bezproblemowa. Zanim dokończysz dezaktywację spróbuj znaleźć rozwiązanie w naszej %1$sdokumentacji%2$s%3$s.',' or post a question on the %1$sforum%2$s'=>' lub zadaj pytanie na %1$sforum%2$s','The plugin stopped working'=>'Wtyczka przestała działać','We take any issues with our plugins very seriously. Try to find a reason in our %1$sdocumentation%2$s%3$s.'=>'Poważnie traktujemy problemy związane z użytkowaniem naszych wtyczek Spróbuj znaleźć rozwiązanie w %1$sdokumentacji%2$s%3$s.',' or post the problem on the %1$sforum%2$s'=>' lub zadaj pytanie na %1$sforum%2$s','I have found another plugin'=>'Znalazłem inną wtyczkę','That hurts a little bit, but we\'re tough! Can you let us know which plugin you are switching to?'=>'Czy możesz nam powiedzieć, z jakiej wtyczki będziesz teraz korzystał?','Which plugin are you switching to?'=>'Z jakiej wtyczki będziesz teraz korzystał?','The plugin doesn\'t have the functionality I need'=>'Wtyczka nie posiada funkcjonalności, której potrzebuję','Good news! There\'s a great chance that the functionality you need is already implemented in the PRO version of the plugin. %1$sContact us%2$s to receive a discount for %3$s. Also, can you describe what functionality you\'re looking for?'=>'Dobra wiadomość! Jest duża szansa, że funkcjonalność, której potrzebujesz, jest już zaimplementowana w wersji PRO wtyczki. %1$sSkontaktuj się z nami%2$s, aby otrzymać zniżkę na %3$s. Ponadto, czy możesz opisać, jakiej funkcjonalności poszukujesz?','We\'re sorry to hear that. Can you describe what functionality you\'re looking for?'=>'Przykro nam to słyszeć! Jakiej funkcjonalności potrzebujesz?','What functionality are you looking for?'=>'Jakiej funkcjonalności potrzebujesz?','I don\'t need the plugin anymore'=>'Nie potrzebuję już wtyczki','Sorry to hear that! Can you let us know why the plugin is not needed anymore?'=>'Przykro nam to słyszeć! Czy możesz nam powiedzieć, dlaczego już nie potrzebujesz wtyczki?','Why is the plugin not needed anymore?'=>'Dlaczego już nie potrzebujesz wtyczki?','I\'m deactivating temporarily for debugging purposes'=>'Dezaktywuję tymczasowo w celu debugowania','Other reason'=>'Inny powód','Can you provide some details on the reason behind deactivation?'=>'Czy możesz podać kilka szczegółów dotyczących powodu dezaktywacji?','Please provide details'=>'Napisz dlaczego dezaktywujesz wtyczkę','Help us improve Octolize plugins\' experience'=>'Pomóż nam ulepszyć i rozwijać wtyczki Octolize','Hi %1$s, with your helping hand we can build effective solutions, launch the new features and shape better plugins experience. By agreeing to anonymously share non-sensitive %2$susage data%3$s of our plugins, you will help us develop them in the right direction. No personal data is tracked or stored and you can opt-out any time. Will you give the thumbs up to our efforts?'=>'Cześć %1$s, z Twoją pomocą możemy tworzyć skuteczne rozwiązania, wdrażać nowe funkcjonalności i stale ulepszać nasze wtyczki. Zgadzając się na udostępnianie anonimowo niewrażliwych %2$sdanych użytkowania%3$s wtyczek, pomożesz nam rozwijać je we właściwym kierunku. Żadne dane osobowe nie są udostępniane ani przechowywane i możesz zrezygnować w dowolnym momencie. Czy możemy na Ciebie liczyć?','Allow'=>'Pozwól','We\'ve added integration with Flexible Shipping Table Rate'=>'Dodaliśmy integrację z Flexible Shipping Table Rate','The new feature allows you to combine Live Rates and Table Rate, providing the ability to use automatically calculated Live Rates while also adjusting shipping costs using Table Rate with the free version of %1$sFlexible Shipping%2$s or %3$sFlexible Shipping PRO%4$s.%5$sThis way, you can now have greater control over the final delivery cost, precisely defining additional charges or discounts for each shipping method.'=>'Nowa funkcja pozwala połączyć Live Rates i Table Rate, dając możliwość skorzystania z automatycznie kalkulowanych stawek Live Rates, a jednocześnie dostosowania kosztów wysyłki za pomocą Table Rate z wykorzystaniem darmowej wersji %1$sFlexible Shipping%2$s lub %3$sFlexible Shipping PRO%4$s.%5$sDzięki temu, możesz mieć teraz większą kontrolę nad finalnym kosztem dostawy, mogąc precyzyjnie określać dodatkowe opłaty lub zniżki na każdej metodzie wysyłki.','Thank you for updating %1$s!'=>'Dziękujemy za aktualizację %1$s!','It is really important to keep the plugins up to date. We have implemented some improvements and new functionalities. Find out what has changed:'=>'To ważne, by utrzymywać wszystkie pluginy w najbardziej aktualnej wersji. Wprowadziliśmy pewne usprawnienia i nowe funkcjonalności. Zobacz co się zmieniło:','I\'m not interested'=>'Nie jestem zainteresowany/-a','Thanks for letting me know'=>'Dzięki za informację!','Page titleShipping Extensions'=>'Rozszerzenia wysyłki','Menu TitleShipping
 Extensions'=>'Rozszerzenia
-wysyłki','Buy bundle →'=>'Kup pakiet wtyczek →','Buy plugin →'=>'Kup wtyczkę →','Flexible Shipping PRO'=>'Flexible Shipping PRO','The best and the most powerful Table Rate shipping plugin for WooCommerce. Define the shipping rules based on numerous conditions and configure even the most complex shipping scenarios with ease.'=>'Najlepsza i najpopularniejsza wtyczka wysyłkowa typu Table Rate dla WooCommerce. Zdefiniuj zasady wysyłki w oparciu o liczne warunki i skonfiguruj nawet najbardziej złożone scenariusze wysyłki.','All Plugins Bundle'=>'Pakiet Wszystkich Wtyczek','Grab a pack of all Octolize plugins as a cut-price tailor-made limited offer for developers, agencies and freelancers. Move the WooCommerce shipping to a whole new level. No strings attached, each plugin\'s 25‑sites subscription included.'=>'Zgarnij pakiet wszystkich wtyczek Octolize w ramach specjalnej, ograniczonej czasowo oferty dla programistów, agencji i freelancerów. Przenieś swoją wysyłkę WooCommerce na zupełnie nowy poziom. Bez żadnych zobowiązań, w zestawie subskrypcja na 25 witryn dla każdej wtyczki.','Complete UPS Integration Bundle'=>'Kompletny Pakiet Integracji UPS','Connect your WooCommerce store with your UPS account, offer your customers real-time shipping rates and generate printable shipping labels for each placed order.'=>'Połącz swój sklep WooCommerce z kontem UPS, aby oferować klientom stawki za przesyłkę w czasie rzeczywistym i generować wydrukowane etykiety wysyłkowe dla każdego złożonego zamówienia.','Advanced DPD UK Integration Bundle'=>'Pakiet Integracji DPD UK','Integrate your WooCommerce store with your DPD UK account, customize the UK locations and automate the whole order processing including generating and printing the shipping labels.'=>'Zintegruj swój sklep WooCommerce z kontem DPD UK, dostosuj opcje dostawy dla Wielkiej Brytanii i zautomatyzuj cały proces realizacji zamówienia, w tym generowanie i drukowanie etykiet wysyłkowych.','Flexible Shipping Bundle'=>'Pakiet Flexible Shipping','Fully customize the shipping cost in your shop, define its calculation rules based on numerous conditions, hide and display the shipping methods and divide orders into separate packages.'=>'W pełni dostosuj koszty wysyłki w swoim sklepie.  Możesz określić reguły ich naliczania w oparciu o wiele warunków, ukrywać i wyświetlać metody wysyłki oraz dzielić zamówienia na osobne paczki.','Live Rates Bundle'=>'Pakiet Stawki Live','Serve your customers the UPS, FedEx, DHL Express or USPS shipping methods with automatically calculated rates and use multiple conditions to display them according to your terms and needs.'=>'Możesz zaoferować swoim klientom metody wysyłki UPS, FedEx, DHL Express lub USPS z automatycznie naliczanymi stawkami i warunkowym wyświetlaniem w zależności od Twoich potrzeb.','UPS Live Rates and Access Points PRO'=>'UPS WooCommerce - Stawki Live i punkty odbioru PRO','WooCommerce UPS integration packed with many advanced features. Display the dynamically calculated live rates for UPS shipping methods and adjust them to your needs.'=>'Integracja WooCommerce UPS z wieloma zaawansowanymi funkcjami. Wyświetl dynamicznie obliczane stawki wysyłki dla metod wysyłki UPS i dostosuj je do swoich potrzeb.','FedEx WooCommerce Live Rates PRO'=>'FedEx WooCommerce Stawki Live PRO','Enable the FedEx live rates for international delivery and integrate it with your shop in less than 5 minutes. Save your time and money – let the shipping cost be calculated automatically.'=>'Zintegruj swój sklep z najpopularniejszym kurierem na świecie. Włącz stawki live oraz automatyczne usługi FedEx w WooCommerce i zapomnij o ręcznej konfiguracji cen wysyłki.','USPS Live Rates PRO'=>'USPS Stawki Live PRO','Serve your customers the automatically and real-time calculated USPS shipping rates. Add the handling fees, insurance and adjust them to your needs with just a few clicks.'=>'Integracja WooCommerce USPS z wieloma zaawansowanymi funkcjami. Wyświetl dynamicznie obliczane stawki wysyłki dla metod wysyłki USPS i dostosuj je do swoich potrzeb.','Flexible Shipping Import / Export'=>'Flexible Shipping Import / Eksport','Use the CSV files to import or export your shipping methods. Edit, update, move or backup the ready configurations and shipping scenarios.'=>'Użyj plików CSV, aby importować lub eksportować swoje metody wysyłki. Edytuj, aktualizuj, przenieś lub utwórz kopię zapasową gotowych konfiguracji i scenariuszy wysyłki.','Flexible Shipping Locations'=>'Flexible Shipping Lokalizacje','Calculate the shipping cost based on location. Define your own custom locations, use the WooCommerce defaults or the ones created by 3rd party plugins.'=>'Twórz reguły wysyłki oparte na lokalizacjach. Różnicuj koszty wysyłki dla grup, pojedynczych krajów lub województw.','Distance Based Shipping Rates'=>'Koszt wysyłki na podstawie odległości','Offer shipping rates based on Distance or Total Travel Time calculated by Google Distance Matrix API and don\'t overpay for shipping.'=>'Uzależnij koszt wysyłki od odległości lub czasu dostawy ze sklepu do klienta.','Pickup Points PRO WooCommerce Plugin'=>'Punkty Odbioru PRO','Provide your customers with multiple carriers\' pickup points at the checkout and let them choose the preferred one to collect their order from.'=>'Oferuj klientom możliwość wyboru punktów odbioru przesyłek od różnych firm kurierskich bezpośrednio podczas finalizacji zamówienia. Dzięki temu klient będzie mógł odebrać zamówienie w najbardziej dogodnym dla siebie miejscu.','Shipping Cost on Product Page PRO'=>'Koszt wysyłki na stronie produktu PRO','Let your customers calculate and see the shipping cost on product pages based on the entered shipping destination and cart contents. Decide how and when exactly you want the shipping cost calculator to display.'=>'Pozwól klientom sklepu obliczyć koszt wysyłki bezpośrednio na stronie produktu. Zdecyduj, jak i kiedy dokładnie chcesz, aby kalkulator kosztów wysyłki był wyświetlany.','Flexible Shipping Box Packing'=>'Flexible Shipping Pakowanie do wysyłki','Use the advanced box packing WooCommerce algorithm to fit the ordered products into your shipping boxes the most optimal way. Configure the shipping cost based on the type and number of the used shipping boxes.'=>'Pozwól algorytmowi pakowania dopasować zamawiane produkty do opakowań wysyłkowych i dobrać ich optymalną konfigurację. Zarządzaj kosztem wysyłki w zależności od typu i ilości użytych opakowań.','Shipping Packages'=>'Shipping Packages','Split the WooCommerce cart content into multiple packages based on various conditions like shipping class.'=>'Podziel koszyk na kilka paczek do wysyłki w oparciu o liczne warunki, takie jak choćby klasa wysyłkowa produktów w koszyku.','Multi Vendor Shipping'=>'Multi Vendor Shipping','Define precisely the shipping cost calculation rules for each Vendor / Product Author in your marketplace or multivendor store.'=>'Zdefiniuj dokładnie zasady obliczania kosztów wysyłki dla każdego sprzedawcy / autora produktu w swoim marketplace lub sklepie typu multivendor.','Delivery Date Picker'=>'Wybór czasu dostawy','Let your customers choose a convenient delivery date for the ordered products and make the shipping cost dependent on the date they choose.'=>'Pozwól klientom wybrać datę dostawy i ustalaj koszt wysyłki dla wybranego przez nich terminu.','DPD UK & DPD Local'=>'DPD UK & DPD Local','Ship your DPD orders faster with advanced DPD UK & DPD Local WooCommerce integration. Gather shipping details, download printable shipping labels and track parcels - everything is automated.'=>'Wysyłaj swoje zamówienia DPD szybciej dzięki zaawansowanej integracji DPD UK & DPD Local WooCommerce. Zbierz szczegóły wysyłki, pobierz drukowalne etykiety wysyłkowe i śledź paczki - wszystko jest zautomatyzowane.','Flexible Printing'=>'Drukowanie PrintNode','Automate your shipping process. Print the shipping labels on thermal printers via PrintNode service. Let the labels be printed automatically the same time the order is placed.'=>'Połącz WooCommerce z drukarkami za pomocą PrintNode. Automatycznie drukuj etykiety nadawcze, faktury, zamówienia i etykiety adresowe.','UPS Labels and Tracking'=>'UPS WooCommerce - Etykiety nadawcze i śledzenie przesyłek','Create the shipments, generate the printable UPS shipping labels for the placed orders and track the parcels directly from your WooCommerce store.'=>'Twórz przesyłki i generuj etykiety wysyłkowe UPS w WooCommerce. Drukuj etykiety i śledź przesyłki UPS bezpośrednio ze swojego sklepu.','DHL Express Live Rates PRO'=>'DHL Express Stawki Live PRO','WooCommerce DHL Express integration packed with many advanced features. Display the dynamically calculated live rates for DHL Express shipping methods and adjust them to your needs.'=>'Integracja WooCommerce DHL Express z wieloma zaawansowanymi funkcjami. Wyświetl dynamicznie obliczane stawki wysyłki dla metod wysyłki DHL Express i dostosuj je do swoich potrzeb.','Canada Post WooCommerce Plugin - Live Rates PRO'=>'Canada Post - Stawki Live','Offer your customers the Canada Post services with real-time calculated shipping rates. Add the handling fees, insurance and adjust them to your needs with just a few clicks.'=>'Zaoferuj swoim klientom usługi Canada Post z kosztami wysyłki naliczanymi w czasie rzeczywistym. Dodaj opłaty manipulacyjne i ubezpieczenie, a następnie dostosuj je do swoich potrzeb kilkoma kliknięciami.','Royal Mail WooCommerce - Live Rates PRO'=>'Royal Mail - Stawki Live','Let your customers choose the Royal Mail shipping methods with the real-time calculated shipping rates. Add the handling fees, insurance and adjust them to your needs in no time.'=>'Oferuj swoim klientom metody wysyłki Royal Mail z kosztami naliczanymi w czasie rzeczywistym. Dodaj opłaty manipulacyjne i ubezpieczenie, a następnie błyskawicznie dostosuj je do swoich potrzeb.','Conditional Shipping Methods'=>'Conditional Shipping Methods','Conditionally display and hide the shipping methods in your WooCommerce store. Define the rules when the specific shipping methods, e.g., live rates should be available to pick and when not to.'=>'Wyświetlaj i ukrywaj warunkowo metody wysyłki w swoim sklepie. Określ w prosty sposób reguły, kiedy konkretne metody wysyłki powinny być dostępne, a kiedy nie.','All'=>'Wszystkie','Bundles'=>'Pakiety wtyczek','Live Rates'=>'Stawki Live','Customizable Rates'=>'Zarządzanie wysyłką','Shipping Labels'=>'Etykiety wysyłkowe','Link on plugin list pageExtensions'=>'Rozszerzenia','Shipping Extensions by'=>'Rozszerzenia wysyłki od','Dive into a system of Octolize ecommerce shipping plugins for WooCommerce. Don’t lose your customers, time and money. Let our plugins secure your sales!'=>'Odkryj pełną ofertę wtyczek wysyłkowych WooCommerce od Octolize. Zarządzaj wysyłką, integruj swój sklep z kurierami i wyświetlaj najdokładniejsze stawki za wysyłkę.','Filter plugins:'=>'Filtruj wtyczki:','Blackout Lead Days'=>'Wykluczanie dni realizacji','Blackout Lead Days are used to define days of the week when shop is not processing orders.'=>'Wykluczanie dni realizacji służy do ustalenia dni tygodnia, w których sklep nie realizuje zamówień.','Monday'=>'Poniedziałek','Tuesday'=>'Wtorek','Wednesday'=>'Środa','Thursday'=>'Czwartek','Friday'=>'Piątek','Saturday'=>'Sobota','Sunday'=>'Niedziela','Can\'t convert weight to target unit.'=>'Nie można przeliczyć wagi na jednostkę docelową.','Please enter USPS User ID.'=>'Wprowadź ID użytkownika USPS.','Multi-currency is supported only in the USPS Live Rates PRO version. %1$sLearn more →%2$s'=>'Wielowalutowość jest wspierana jedynie w wersji PRO wtyczki USPS Stawki live. %1$sDowiedz się więcej →%2$s','Priority Mail'=>'Priority Mail','Priority Mail Express; Hold For Pickup'=>'Priority Mail Express; Hold For Pickup','Priority Mail Express'=>'Priority Mail Express','Standard Post'=>'Standard Post','Media Mail'=>'Media Mail','Library Mail'=>'Library Mail','Priority Mail Express; Flat Rate Envelope'=>'Priority Mail Express; Flat Rate Envelope','First-Class Mail; Large Postcards'=>'First-Class Mail; Large Postcards','Priority Mail; Flat Rate Envelope'=>'Priority Mail; Flat Rate Envelope','Priority Mail; Medium Flat Rate Box'=>'Priority Mail; Medium Flat Rate Box','Priority Mail; Large Flat Rate Box'=>'Priority Mail; Large Flat Rate Box','Priority Mail Express; Sunday/Holiday Delivery'=>'Priority Mail Express; Sunday/Holiday Delivery','Priority Mail Express; Sunday/Holiday Delivery Flat Rate Envelope'=>'Priority Mail Express; Sunday/Holiday Delivery Flat Rate Envelope','Priority Mail Express; Flat Rate Envelope Hold For Pickup'=>'Priority Mail Express; Flat Rate Envelope Hold For Pickup','Priority Mail; Small Flat Rate Box'=>'Priority Mail; Small Flat Rate Box','Priority Mail; Padded Flat Rate Envelope'=>'Priority Mail; Padded Flat Rate Envelope','Priority Mail Express; Legal Flat Rate Envelope'=>'Priority Mail Express; Legal Flat Rate Envelope','Priority Mail Express; Legal Flat Rate Envelope Hold For Pickup'=>'Priority Mail Express; Legal Flat Rate Envelope Hold For Pickup','Priority Mail Express; Sunday/Holiday Delivery Legal Flat Rate Envelope'=>'Priority Mail Express; Sunday/Holiday Delivery Legal Flat Rate Envelope','Priority Mail; Hold For Pickup'=>'Priority Mail; Hold For Pickup','Priority Mail; Large Flat Rate Box Hold For Pickup'=>'Priority Mail; Large Flat Rate Box Hold For Pickup','Priority Mail; Medium Flat Rate Box Hold For Pickup'=>'Priority Mail; Medium Flat Rate Box Hold For Pickup','Priority Mail; Small Flat Rate Box Hold For Pickup'=>'Priority Mail; Small Flat Rate Box Hold For Pickup','Priority Mail; Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Flat Rate Envelope Hold For Pickup','Priority Mail; Gift Card Flat Rate Envelope'=>'Priority Mail; Gift Card Flat Rate Envelope','Priority Mail; Gift Card Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Gift Card Flat Rate Envelope Hold For Pickup','Priority Mail; Window Flat Rate Envelope'=>'Priority Mail; Window Flat Rate Envelope','Priority Mail; Window Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Window Flat Rate Envelope Hold For Pickup','Priority Mail; Small Flat Rate Envelope'=>'Priority Mail; Small Flat Rate Envelope','Priority Mail; Small Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Small Flat Rate Envelope Hold For Pickup','Priority Mail; Legal Flat Rate Envelope'=>'Priority Mail; Legal Flat Rate Envelope','Priority Mail; Legal Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Legal Flat Rate Envelope Hold For Pickup','Priority Mail; Padded Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Padded Flat Rate Envelope Hold For Pickup','Priority Mail; Regional Rate Box A'=>'Priority Mail; Regional Rate Box A','Priority Mail; Regional Rate Box A Hold For Pickup'=>'Priority Mail; Regional Rate Box A Hold For Pickup','Priority Mail; Regional Rate Box B'=>'Priority Mail; Regional Rate Box B','Priority Mail; Regional Rate Box B Hold For Pickup'=>'Priority Mail; Regional Rate Box B Hold For Pickup','First-Class; Package Service Hold For Pickup'=>'First-Class; Package Service Hold For Pickup','Priority Mail Express; Flat Rate Boxes'=>'Priority Mail Express; Flat Rate Boxes','Priority Mail Express; Flat Rate Boxes Hold For Pickup'=>'Priority Mail Express; Flat Rate Boxes Hold For Pickup','Priority Mail Express; Sunday/Holiday Delivery Flat Rate Boxes'=>'Priority Mail Express; Sunday/Holiday Delivery Flat Rate Boxes','Priority Mail; Regional Rate Box C'=>'Priority Mail; Regional Rate Box C','Priority Mail; Regional Rate Box C Hold For Pickup'=>'Priority Mail; Regional Rate Box C Hold For Pickup','First-Class; Package Service'=>'First-Class; Package Service','Priority Mail Express; Padded Flat Rate Envelope'=>'Priority Mail Express; Padded Flat Rate Envelope','Priority Mail Express; Padded Flat Rate Envelope Hold For Pickup'=>'Priority Mail Express; Padded Flat Rate Envelope Hold For Pickup','Priority Mail Express; Sunday/Holiday Delivery Padded Flat Rate Envelope'=>'Priority Mail Express; Sunday/Holiday Delivery Padded Flat Rate Envelope','Priority Mail 1-Day Return Service Padded Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Padded Flat Rate Envelope','Priority Mail 1-Day Return Service Gift Card Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Gift Card Flat Rate Envelope','Priority Mail 1-Day Return Service Window Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Window Flat Rate Envelope','Priority Mail 1-Day Return Service Small Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Small Flat Rate Envelope','Priority Mail 1-Day Return Service Legal Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Legal Flat Rate Envelope','Priority Mail 1-Day Return Service Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Flat Rate Envelope','Priority Mail 1-Day Return Service Regional Rate Box A'=>'Priority Mail 1-Day Return Service Regional Rate Box A','Priority Mail 1-Day Return Service Regional Rate Box B'=>'Priority Mail 1-Day Return Service Regional Rate Box B','Priority Mail 1-Day Return Service'=>'Priority Mail 1-Day Return Service','Priority Mail 1-Day Return Service Large Flat Rate Box'=>'Priority Mail 1-Day Return Service Large Flat Rate Box','Priority Mail 1-Day Return Service Medium Flat Rate Box'=>'Priority Mail 1-Day Return Service Medium Flat Rate Box','Priority Mail 1-Day Return Service Small Flat Rate Box'=>'Priority Mail 1-Day Return Service Small Flat Rate Box','Priority Mail Military Return Service Large Flat Rate Box APO/FPO/DPO'=>'Priority Mail Military Return Service Large Flat Rate Box APO/FPO/DPO','Priority Mail 1-Day Return Service Cubic'=>'Priority Mail 1-Day Return Service Cubic','First-Class Package Return Service'=>'First-Class Package Return Service','Ground Return Service'=>'Ground Return Service','Priority Mail Express International'=>'Priority Mail Express International','Priority Mail International'=>'Priority Mail International','Global Express Guaranteed; (GXG)**'=>'Global Express Guaranteed; (GXG)**','Global Express Guaranteed; Document'=>'Global Express Guaranteed; Document','Global Express Guarantee; Non-Document Rectangular'=>'Global Express Guarantee; Non-Document Rectangular','Global Express Guaranteed; Non-Document Non-Rectangular'=>'Global Express Guaranteed; Non-Document Non-Rectangular','Priority Mail International; Flat Rate Envelope**'=>'Priority Mail International; Flat Rate Envelope**','Priority Mail International; Medium Flat Rate Box'=>'Priority Mail International; Medium Flat Rate Box','Priority Mail Express International; Flat Rate Envelope'=>'Priority Mail Express International; Flat Rate Envelope','Priority Mail International; Large Flat Rate Box'=>'Priority Mail International; Large Flat Rate Box','USPS GXG; Envelopes**'=>'USPS GXG; Envelopes**','First-Class Mail; International Letter**'=>'First-Class Mail; International Letter**','First-Class Mail; International Large Envelope**'=>'First-Class Mail; International Large Envelope**','First-Class Package International Service**'=>'First-Class Package International Service**','Priority Mail International; Small Flat Rate Box**'=>'Priority Mail International; Small Flat Rate Box**','Priority Mail Express International; Legal Flat Rate Envelope'=>'Priority Mail Express International; Legal Flat Rate Envelope','Priority Mail International; Gift Card Flat Rate Envelope**'=>'Priority Mail International; Gift Card Flat Rate Envelope**','Priority Mail International; Window Flat Rate Envelope**'=>'Priority Mail International; Window Flat Rate Envelope**','Priority Mail International; Small Flat Rate Envelope**'=>'Priority Mail International; Small Flat Rate Envelope**','USPS'=>'USPS','These are the USPS Live Rates plugin general settings. In order to learn more about its configuration please refer to its %1$sdedicated documentation →%2$s'=>'Ta strona zawiera ogólne ustawienia wtyczki USPS Live Rates. Aby dowiedzieć się więcej o jej konfiguracji, zapoznaj się z jej %1$sdedykowaną dokumentacją →%2$s','API Settings'=>'Ustawienia API','Advanced Options'=>'Opcje zaawansowane','Debug Mode'=>'Tryb debugowania','Enable debug mode'=>'Włącz tryb debugowania','Enable debug mode to display the additional tech information, incl. the data sent to USPS API, visible only for Admins and Shop Managers in the cart and checkout.'=>'Włącz Tryb debugowania, aby wyświetlić dodatkowe informacje techniczne, m.in. dane wysyłane do API USPS, widoczne tylko dla Administratorów i Menadżerów Sklepu w koszyku i na stronie finalizacji zamówienia.','Method Settings'=>'Ustawienia metody','Manage the way how the USPS services are displayed in the cart and checkout.'=>'Zarządzaj w jaki sposób metody wysyłki USPS zostaną wyświetlone w koszyku i na stronie finalizacji zamówienia.','Method Title'=>'Tytuł metody wysyłki','Define the USPS shipping method title which should be used in the cart/checkout when the Fallback option was triggered.'=>'Wprowadź tytuł dla metody wysyłki USPS, który zostanie wyświetlony klientom w koszyku i na stronie finalizacji zamówienia, jeżeli zostanie zastosowana stawka awaryjna.','USPS Live rates'=>'USPS Stawki live','Enable to offer flat rate cost for shipping so that the user can still checkout, if API for some reason returns no matching rates.'=>'Zaznacz tą opcję, aby zastosować stałą stawkę awaryjną w przypadku, gdy z jakiegoś powodu nie będzie możliwe wyświetlenie stawek rzeczywistych.','Free Shipping'=>'Darmowa wysyłka','Services'=>'Usługi USPS','Enable the services\' custom settings'=>'Włącz własne ustawienia usług','Decide which services should be displayed and which not, change their names and order. Please mind that enabling a service does not guarantee it will be visible in the cart/checkout. It has to be available for the provided package weight, origin and destination in order to be displayed.'=>'Włącz, aby określić, które usługi mają być wyświetlane, a które nie, zmienić ich nazwy i/lub kolejność. Należy mieć na uwadze, że samo włączenie usługi nie gwarantuje jeszcze, że będzie ona widoczna w koszyku lub na stronie finalizacji zamówienia. Dana usługa musi być dostępna dla podanej wagi paczki oraz adresów nadania i dostawy, aby została wyświetlona.','Domestic Services Table'=>'Tabela usług krajowych USPS','International Services Table'=>'Tabela usług międzynarodowych USPS','Type'=>'Typ','Rates Adjustments'=>'Dostosowanie stawek','Adjust these settings to get more accurate rates. Read %swhat affects the USPS rates in USPS WooCommerce plugin →%s'=>'Użyj poniższych ustawień, aby dostosować wysokość stawek do swoich potrzeb. Dowiedz się %sco wpływa na stawki w naszej wtyczce (ENG) →%s','https://octol.io/usps-free-rates'=>'https://octol.io/usps-free-rates','Commercial Rates'=>'Stawki komercyjne','Use the USPS Commercial Pricing if available'=>'Użyj USPS Commercial Pricing, jeśli jest dostępny','Tick this checkbox if you want to use and display the Commercial Pricing shipping rates to your customers instead of the standard ones.'=>'Zaznacz ten checkbox, jeżeli zamiast standardowych stawek USPS chcesz używać stawek Commercial Pricing.','Insurance'=>'Ubezpieczenie','Request insurance to be included in the USPS rates'=>'Uwzględniaj ubezpieczenie w stawkach USPS','Enable if you want to include insurance in the USPS rates if possible.'=>'Włącz opcję jeśli chcesz uwzględniać ubezpieczenie w stawkach USPS, gdy jest dostępne.','Empty response!'=>'Pusta odpowiedź API!','USPS Live Rates'=>'USPS Stawki live','USPS integration'=>'Integracja USPS','None'=>'Brak','Set image'=>'Wstaw obraz','Remove image'=>'Usuń obraz','Select or Upload Media'=>'Wybierz lub prześlij media','Use this media'=>'Wybierz','Search for a product&hellip;'=>'Szukaj produktu&hellip;','Disabled'=>'Wyłączono','Emergency'=>'Wypadek','Alert'=>'Ostrzeżenie','Critical'=>'Błąd krytyczny','Error'=>'Błąd','Warning'=>'Ostrzeżenie','Notice'=>'Powiadomienie','Info'=>'Informacja','Debug'=>'Debugowanie','Can not enable WP Desk Debug log! Cannot create directory %s or this directory is not writeable!'=>'Nie można włączyć WP Desk Debug log! Nie można utworzyć katalogu %s lub brak uprawnień do zapisu w tym katalogu!','Can not enable WP Desk Debug log! Cannot create file %s!'=>'Nie można włączyć logowania błędów za pomocą WP Desk Debug! Nie można utworzyć pliku %s!','Length %1$s'=>'Długość %1$s','Width %1$s'=>'Szerokość %1$s','Height %1$s'=>'Wysokość %1$s','Max Weight %1$s'=>'Waga maks. %1$s','Padding %1$s'=>'Wypełnienie %1$s','Box Weight %1$s'=>'Waga pudełka %1$s','Delete'=>'Usuń','Add'=>'Dodaj','Select the package type the ordered products will be matched to. You can choose one or as many different packagings as you need. If selected, filling in the products\' weight and dimensions fields is required. %1$sLearn more about the sizes and package types →%2$s'=>'Wybierz typ lub typy opakowań, do których zostaną dopasowane i zapakowane zamówione produkty. Możesz w tym miejscu wybrać ich dowolną ilość. Wybór tej opcji wymaga uzupełnienia wagi i wymiarów każdego z produktów. %1$sDowiedz się więcej o wymiarach i typach opakowań →%2$s','Shipping boxes'=>'Pudełka','Pack into one box by weight'=>'Pakuj do jednego pudełka według wagi','Pack into custom boxes'=>'Pakuj do własnych pudełek','Pack items separately'=>'Pakuj produkty oddzielnie','Pack into custom boxes (3D bin packing)'=>'Pakuj do własnych pudełek (Pakowanie 3D bin)','Pack into custom boxes (volume packing)'=>'Pakuj do własnych pudełek (Pakowanie wg objętości)','Parcel Packing Method'=>'Metoda pakowania paczek','Define the way how the ordered products should be packed. Changing your choice here may affect the rates.'=>'Określ sposób pakowania zamawianych produktów. Wybór poszczególnych opcji w tym miejscu może mieć wpływ na wysokość stawek.','Cannot pack item: %1$s. %2$s'=>'Nie można zapakować produktu: %1$s. %2$s','Cannot pack items: %1$s.'=>'Nie można zapakować produktu: %1$s.','weight'=>'waga','width'=>'szerokość','length'=>'długość','height'=>'wysokość','Item %1$s not set!'=>'Dane produktu %1$s nie zostały zdefiniowane!','Custom (%1$sx%2$sx%3$s/%4$s)'=>'Własne (%1$sx%2$sx%3$s/%4$s)','Dynamically calculated USPS live rates based on the established USPS API connection. %1$sLearn more →%2$s'=>'Dynamicznie obliczane w oparciu o połączenie z API USPS stawki rzeczywiste. %1$sDowiedz się więcej →%2$s','%1$sPlease add now the %2$s shipping method within the shipping zone you want it to be available in →%3$s'=>'%1$sDodaj teraz metodę wysyłki %2$s w strefie wysyłkowej, w której chcesz, aby była dostępna →%3$s','API Connection Status'=>'Status połączenia z API','Checking...'=>'Sprawdzanie...','If you encounter any problems with establishing the API connection, the detailed information on its cause will be displayed here.'=>'Jeśli wystąpią problemy z nawiązaniem połączenia z API, w tym miejscu zostaną wyświetlone szczegółowe informacje o ich przyczynie.','OK'=>'OK','Not found HTML view for custom field %1$s.'=>'Nie znaleziono widoku HTML dla pola niestandardowego %1$s.','Fixed value'=>'Wartość stała','Percentage'=>'Wartość procentowa','Use this option to apply the handling fees to the rates. You can use either fixed or percentage values, including the negative ones for discounts.'=>'Użyj tej opcji by doliczyć stały lub procentowy koszt prowizji do zwracanych przez wtyczkę stawek, włącznie z minusowymi wartościami dla obniżek.','Fees Amount'=>'Wartość prowizji','Enter a positive value to charge your customers additionally or a negative one to apply the discount. If you use the currency switcher, the final price incl. Handling Fee will be automatically converted to the currently active currency in your shop. The rates will also include the taxes based on your current Tax settings.'=>'Dodatnia liczba = dopłata, liczba ujemna = zniżka. Jeżeli używasz switcherów walut, cena końcowa zawierająca prowizję zostanie automatycznie przeliczona na obecnie aktywną walutę w sklepie. W cenie zostaną również uwzględnione aktualne ustawienia podatków.','Code'=>'Kod','Name'=>'Nazwa','Enabled'=>'Włączone','Drag and drop the services to control their display order. Confirm by clicking Save changes button below.'=>'Przeciągnij i upuść usługę, aby zmienić kolejność wyświetlania. Zatwierdź zmiany klikając Zapisz zmiany poniżej.','Custom Origin'=>'Własny adres nadania','Origin Address'=>'Ulica','Origin City'=>'Miasto','Origin Postcode'=>'Kod pocztowy','Origin Country/State'=>'Kraj','Origin Settings for the whole store'=>'Adres nadania dla całego sklepu','Origin Settings'=>'Adres nadania','Enable custom origin for the whole store'=>'Ustaw własny adres nadania dla całego sklepu','Enable custom origin'=>'Ustaw własny adres nadania','Use this option to use a different shipper\'s address than the one defined in the WooCommerce settings. If not enabled, the one you\'ve entered in WooCommerce → Settings → General → Store Address will be used by default. Each shipping method can have their own custom origin.'=>'Zaznacz tę opcję, by użyć innego adresu nadawcy niż domyślny adres sklepu, zdefiniowany w ustawieniach WooCommerce (WooCommerce → Ustawienia → Ogólne → Adres sklepu). Metody wysyłki mogą mieć własny adres nadania.','Use this option to use a different shipper\'s address than the one defined in the WooCommerce settings. If not enabled, the one you\'ve entered in WooCommerce → Settings → General → Store Address will be used by default.'=>'Zaznacz tę opcję, by użyć innego adresu nadawcy niż domyślny adres sklepu, zdefiniowany w ustawieniach WooCommerce (WooCommerce → Ustawienia → Ogólne → Adres sklepu).','Origin Settings for this Shipping Method'=>'Adres nadania dla tej metody wysyłki','Enable custom origin for this shipping method'=>'Ustaw własny adres nadania dla tej metody wysyłki','Use this option to use a different shipper\'s address for this shipping method.'=>'Zaznacz tę opcję, by użyć innego adresu nadawcy dla tej metody wysyłki.','Cart value exceeds %s. Free shipping has been applied.'=>'Wartość koszyka przekracza %s. Zastosowano bezpłatną wysyłkę.','Cart value doesn\'t exceed %s. Free shipping hasn\'t been applied.'=>'Wartość koszyka nie przekracza %s. Bezpłatna wysyłka nie została zastosowana.','Subtotal'=>'Kwota','Free Shipping Amount'=>'Kwota darmowej wysyłki','Enable the free shipping over amount'=>'Włącz darmową wysyłkę powyżej określonej kwoty','Tick this checkbox to enter the order amount above which the shipping becomes free.'=>'Zaznacz, aby wprowadzić wartość zamówienia, po przekroczeniu której wysyłka stanie się darmowa.','Free Shipping Threshold'=>'Próg darmowej wysyłki','Enter only a numeric value without the currency symbol.'=>'Wprowadź samą wartość, bez symbolu waluty.','Fallback reason'=>'Powód stawki awaryjnej','Packages'=>'Paczki','Custom'=>'Własna','No rates added from collection point rates!'=>'Nie zostały dodane żadne stawki dla punktów odbioru!','Collection point not found.'=>'Punkt odbioru nie został znaleziony.','There was no valid services available.'=>'Żadne stawki wysyłki nie były dostępne.','Fallback rate added with reason: %1$s'=>'Została użyta stawka awaryjna. Powód: %1$s','No rates added from standard service rates!'=>'Nie zostały dodane żadne stawki standardowe!','Additional costs by Flexible Shipping Table Rate'=>'Dodatkowy koszt przez Flexible Shipping Table Rate','Additional costs'=>'Dodatkowe koszty','Enable Flexible Shipping Rules Table'=>'Włącz tabelę reguł Flexible Shipping','Enabling this option will allow you to adjust the calculated shipping cost, with additional charges applied according to the shipping cost calculation rules.'=>'Włączenie tej opcji umożliwi Ci dostosowanie wyliczonego kosztu przesyłki, naliczając dodatkowe opłaty wg zadeklarowanych reguł kalkulacji kosztów.','Shipping Cost Calculation Rules'=>'Reguły kalkulacji kosztów wysyłki','You can update %1$sFlexible Shipping%2$s plugin to define additional costs for this shipping method. %3$s'=>'Możesz zaktualizować wtyczkę %1$sFlexible Shipping%2$s, aby zdefiniować dodatkowe koszty dla tej metody wysyłki. %3$s','Update Now'=>'Zaktualizuj Teraz','You can activate %1$sFlexible Shipping%2$s plugin to define additional costs for this shipping method. %3$s'=>'Możesz aktywować wtyczkę %1$sFlexible Shipping%2$s, aby zdefiniować dodatkowe koszty dla tej metody wysyłki. %3$s','Activate Now'=>'Aktywuj Teraz','You can install %1$sFlexible Shipping%2$s plugin to define additional costs for this shipping method. %3$s'=>'Możesz zainstalować wtyczkę %1$sFlexible Shipping%2$s, aby zdefiniować dodatkowe koszty dla tej metody wysyłki. %3$s','Install Now'=>'Zainstaluj teraz','Fallback'=>'Stawka awaryjna','Enable fallback'=>'Włącz stawkę awaryjną','Fallback Cost'=>'Koszt stawki awaryjnej','We noticed that you are using the Germanized plugin along with %1$s. In order to display the accurate rates, please select the WooCommerce default option in the Germanized Taxes > Additional Costs > Tax calculation mode settings. %2$sNeed assistance? Get in touch →%3$s'=>'Zauważyliśmy, że używasz wtyczki Germanized wraz z %1$s. Aby wyświetlić dokładne stawki, wybierz domyślną opcję WooCommerce w ustawieniach trybu obliczania podatku zgermanizowanych podatków > Koszty dodatkowe >. %2$sPotrzebujesz pomocy? Skontaktuj się z →%3$s','Show %1$s'=>'Pokaż %1$s','Hide %1$s'=>'Ukryj %1$s','Copy %1$s to clipboard'=>'Skopiuj %1$s do schowka','(Delivery Days: 0 days)'=>'(Dni do dostawy: 0 dni)','(Delivery Days: %1$d day)'=>'(Dni do dostawy: %1$d dzień)' . "\0" . '' . "\0" . '','(Delivery Date: %1$s)'=>'(Data dostawy: %1$s)','Awesome, you\'ve been using %s for more than 2 weeks. Could you please do me a BIG favor and give it a 5-star rating on WordPress? ~ Peter'=>'Fantastycznie, korzystasz z %s już ponad 2 tygodnie. Czy możesz wyświadczyć mi wielką przysługę i wystawić 5-gwiazdkową ocenę? ~Piotr','%1$sOk, you deserved it%2$s'=>'%1$sTak, zasługujecie na nią%2$s','%1$sNope, maybe later%2$s'=>'%1$sNie, może później%2$s','%1$sI already did%2$s'=>'%1$sJuż to zrobiłem/zrobiłam%2$s','Created with %1$s by %2$s - If you like %3$s you can %4$srate us %5$s in plugins repository &rarr;%6$s'=>'Stworzone z %1$s przez %2$s - jeżeli podoba Ci się %3$s możesz %4$sdać nam %5$s w repozytorium wtyczek &rarr;%6$s','The plugin suddenly stopped working'=>'Wtyczka nagle przestała działać','The plugin broke my site'=>'Wtyczka zepsuła moją stronę','I have found a better plugin'=>'Znalazłem lepszą wtyczkę','What\'s the plugin\'s name?'=>'Jaka to wtyczka?','I only needed the plugin for a short period'=>'Potrzebowałem wtyczki tylko przez krótki okres','I no longer need the plugin'=>'Nie potrzebuję już wtyczki','It\'s a temporary deactivation. I\'m just debugging an issue.'=>'Jest to tymczasowa dezaktywacja, debuguję problem.','Other'=>'Inny','Please let us know how we can improve our plugin'=>'Napisz co moglibyśmy zmienić, aby ulepszyć naszą wtyczkę','Plugin deactivation'=>'Wyłączanie wtyczki','You are deactivating %s plugin.'=>'Dezaktywujesz wtyczkę %s.','Skip &amp; Deactivate'=>'Pomiń i dezaktywuj','Submit &amp; Deactivate'=>'Wyślij i dezaktywuj','Cancel'=>'Anuluj','You successfully opted out of collecting usage data by %1$s. If you change your mind, you can always opt in later in the plugin\'s quick links.'=>'Twoja rezygnacja ze zbierania danych użytkowania wtyczek przez %1$s została przyjęta. Jeśli zmienisz zdanie, zawsze możesz włączyć zgodę ponownie za pomocą szybkich linków na liście wtyczek.','Enable usage tracking'=>'Włącz śledzenie użycia','Disable usage tracking'=>'Wyłącz śledzenie użycia','Hey %s,'=>'Hej %s,','Please help us improve our plugins! If you opt-in, we will collect some non-sensitive data and usage information anonymously. If you skip this, that\'s okay! All plugins will work just fine.'=>'Proszę pomóż nam udoskonalić nasze wtyczki! Jeśli się zgodzisz, będziemy zbierać niewrażliwe dane użytkowania anonimowo. Jeśli pominiesz ten krok, w porządku! Wszystkie wtyczki będą działać niezależnie od Twojej decyzji.','Allow & Continue &rarr;'=>'Pozwól i kontynuuj &rarr;','Skip'=>'Pomiń','What permissions are being granted?'=>'Jakie uprawnienia są przyznawane?','Your Site Overview'=>'Dane witryny','WP version, PHP info'=>'Wersja WP, wersja PHP','Plugin Usage'=>'Dane użytkowania wtyczek','Current settings and usage information of %1$s plugins'=>'Ustawienia i informacje o użytkowaniu wtyczek %1$s','Your Store Overview'=>'Dane sklepu','Anonymized and non-sensitive store usage information'=>'Tylko niewrażliwe i anonimowe dane o użytkowaniu sklepu','Find out more &raquo;'=>'Dowiedz się więcej &raquo;','Current settings and usage information of WP Desk plugins'=>'Ustawienia i informacje o użytkowaniu wtyczek WP Desk',' If you have a moment, please let us know why you are deactivating plugin (anonymous feedback):'=>' Jeśli masz chwilę, prosimy daj nam znać dlaczego dezaktywujesz wtyczkę (anonimowa opinia):','I found a better plugin'=>'Znalazłem lepszą wtyczkę','Kindly tell us the reason so we can improve'=>'Napisz nam powód abyśmy mogli go poprawić','We need your help to improve <strong>WP Desk plugins</strong>, so they are more useful for you and the rest of <strong>30,000+ users</strong>. By collecting data on how you use our plugins, you will help us a lot. We will not collect any sensitive data, so you can feel safe.'=>'Potrzebujemy Twojej pomocy, aby dalej rozwijać <strong>wtyczki WP Desk</strong> i były one jeszcze bardziej pomocne dla Ciebie i pozostałych ponad <strong>30 000 użytkowników</strong>. Zbierając dane na temat tego jak korzystasz z naszych wtyczek bardzo nam w tym pomożesz. Nie będziemy zbierać żadnych danych wrażliwych, więc możesz czuć się bezpiecznie.','You successfully opted out of collecting usage data by WP Desk. If you change your mind, you can always opt in later in the plugin\'s quick links.'=>'Twoja rezygnacja ze zbierania danych użytkowania wtyczek przez WP Desk została przyjęta. Jeśli zmienisz zdanie, zawsze możesz włączyć zgodę ponownie za pomocą szybkich linków na liście wtyczek.','Upgrade'=>'Kup PRO','Upgrade Now &rarr;'=>'Zaktualizuj do wersji PRO &rarr;','USPS WooCommerce shipping methods with real-time calculated shipping rates based on the established USPS API connection.'=>'Metody wysyłki USPS dla WooCommerce ze stawkami obliczanymi w czasie rzeczywistym w oparciu o nawiązane połączenie z API USPS.','You need to provide the USPS account credentials to get live rates. If you do not have the USPS account yet, you can register a new account directly at %1$susps.com →%2$s'=>'Musisz wprowadzić dane dostępowe konta USPS, aby uzyskać stawki rzeczywiste. Jeżeli nie posiadasz jeszcze konta USPS, możesz utworzyć nowe konto bezpośrednio na stronie %1$susps.com →%2$s','USPS User ID *'=>'ID użytkownika USPS *','Enter the credentials you acquired during the USPS account registration process.'=>'Wprowadź dane dostępowe uzyskane podczas procesu rejestracji konta USPS.','Test Credentials'=>'Testowe dane dostępowe','Enable to use test credentials'=>'Używaj testowych danych dostępowych','The &#8220;%s&#8221; plugin cannot run on PHP versions older than %s. Please contact your host and ask them to upgrade.'=>'Wtyczka &#8220;%s&#8221; nie może działać na wersjach PHP starszych niż %s. Skontaktuj się z administratorem hostingu i poproś o aktualizację wersji PHP.','The &#8220;%s&#8221; plugin cannot run on WordPress versions older than %s. Please update WordPress.'=>'Wtyczka &#8220;%s&#8221; nie może działać na wersjach WordPressa starszych niż %s. Wykonaj aktualizację WordPressa.','The &#8220;%s&#8221; plugin cannot run on WooCommerce versions older than %s. Please update WooCommerce.'=>'Wtyczka &#8220;%s&#8221; nie może działać na wersjach WooCommerce starszych niż %s. Wykonaj aktualizację WooCommerce.','The &#8220;%s&#8221; plugin cannot run without OpenSSL module version at least %s. Please update OpenSSL module.'=>'Wtyczka &#8220;%s&#8221; nie może działać bez aktywacji OpenSSL co najmniej w wersji %s. Wykonaj aktualizację OpenSSL.','The &#8220;%1$s&#8221; plugin requires at least %2$s version of %3$s to work correctly. Please update it to its latest release.'=>'Do poprawnego działania wtyczki &#8220;%1$s&#8221; wymagana jest wtyczka %3$s co najmniej w wersji %2$s. Wykonaj jej aktualizację.','The &#8220;%s&#8221; plugin cannot run without %s active. Please install and activate %s plugin.'=>'Wtyczka &#8220;%s&#8221; nie może działać bez aktywacji %s. Zainstaluj i aktywuj wtyczkę %s.','The &#8220;%s&#8221; plugin requires free %s plugin. <a href="%s">Install%s</a>'=>'Wtyczka &#8220;%s&#8221; wymaga bezpłatnej wtyczki %s. <a href="%s">Zainstaluj %s</a>','The &#8220;%s&#8221; plugin requires activating %s plugin. <a href="%s">Activate %s</a>'=>'Wtyczka &#8220;%s&#8221; wymaga aktywnej wtyczki %s. <a href="%s">Włącz %s</a>','The &#8220;%s&#8221; plugin cannot run without %s PHP module installed. Please contact your host and ask them to install %s.'=>'Wtyczka &#8220;%s&#8221; nie może działać bez modułu PHP: %s. Skontaktuj się z administratorem hostingu i poproś o instalację %s.','The &#8220;%s&#8221; plugin cannot run without %s PHP setting set to %s. Please contact your host and ask them to set %s.'=>'Wtyczka &#8220;%s&#8221; nie może działać bez ustawienia PHP %s ustawionego na %s. Skontaktuj się z administratorem hostingu i poproś o ustawienie %s.','The &#8220;%s&#8221; plugin is temporarily disabled since the required %s plugin is being upgraded.'=>'Wtyczka &#8220;%s&#8221; została chwilowo wyłączona, ponieważ wymagana wtyczka %s jest właśnie aktualizowana.','Shipping service should implements CanTestSettings interface! Or overwrite this method'=>'Shipping service should implements CanTestSettings interface! Or overwrite this method','If you have a moment, please let us know why you are deactivating the plugin (anonymous feedback):'=>'Jeśli masz chwilę, prosimy daj nam znać dlaczego wyłączasz wtyczkę (anonimowa opinia):']];
+wysyłki','Buy bundle →'=>'Kup pakiet wtyczek →','Buy plugin →'=>'Kup wtyczkę →','Flexible Shipping PRO'=>'Flexible Shipping PRO','The best and the most powerful Table Rate shipping plugin for WooCommerce. Define the shipping rules based on numerous conditions and configure even the most complex shipping scenarios with ease.'=>'Najlepsza i najpopularniejsza wtyczka wysyłkowa typu Table Rate dla WooCommerce. Zdefiniuj zasady wysyłki w oparciu o liczne warunki i skonfiguruj nawet najbardziej złożone scenariusze wysyłki.','All Plugins Bundle'=>'Pakiet Wszystkich Wtyczek','Grab a pack of all Octolize plugins as a cut-price tailor-made limited offer for developers, agencies and freelancers. Move the WooCommerce shipping to a whole new level. No strings attached, each plugin\'s 25‑sites subscription included.'=>'Zgarnij pakiet wszystkich wtyczek Octolize w ramach specjalnej, ograniczonej czasowo oferty dla programistów, agencji i freelancerów. Przenieś swoją wysyłkę WooCommerce na zupełnie nowy poziom. Bez żadnych zobowiązań, w zestawie subskrypcja na 25 witryn dla każdej wtyczki.','Complete UPS Integration Bundle'=>'Kompletny Pakiet Integracji UPS','Connect your WooCommerce store with your UPS account, offer your customers real-time shipping rates and generate printable shipping labels for each placed order.'=>'Połącz swój sklep WooCommerce z kontem UPS, aby oferować klientom stawki za przesyłkę w czasie rzeczywistym i generować wydrukowane etykiety wysyłkowe dla każdego złożonego zamówienia.','Advanced DPD UK Integration Bundle'=>'Pakiet Integracji DPD UK','Integrate your WooCommerce store with your DPD UK account, customize the UK locations and automate the whole order processing including generating and printing the shipping labels.'=>'Zintegruj swój sklep WooCommerce z kontem DPD UK, dostosuj opcje dostawy dla Wielkiej Brytanii i zautomatyzuj cały proces realizacji zamówienia, w tym generowanie i drukowanie etykiet wysyłkowych.','Flexible Shipping Bundle'=>'Pakiet Flexible Shipping','Fully customize the shipping cost in your shop, define its calculation rules based on numerous conditions, hide and display the shipping methods and divide orders into separate packages.'=>'W pełni dostosuj koszty wysyłki w swoim sklepie.  Możesz określić reguły ich naliczania w oparciu o wiele warunków, ukrywać i wyświetlać metody wysyłki oraz dzielić zamówienia na osobne paczki.','Live Rates Bundle'=>'Pakiet Stawki Live','Serve your customers the UPS, FedEx, DHL Express or USPS shipping methods with automatically calculated rates and use multiple conditions to display them according to your terms and needs.'=>'Możesz zaoferować swoim klientom metody wysyłki UPS, FedEx, DHL Express lub USPS z automatycznie naliczanymi stawkami i warunkowym wyświetlaniem w zależności od Twoich potrzeb.','UPS Live Rates and Access Points PRO'=>'UPS WooCommerce - Stawki Live i punkty odbioru PRO','WooCommerce UPS integration packed with many advanced features. Display the dynamically calculated live rates for UPS shipping methods and adjust them to your needs.'=>'Integracja WooCommerce UPS z wieloma zaawansowanymi funkcjami. Wyświetl dynamicznie obliczane stawki wysyłki dla metod wysyłki UPS i dostosuj je do swoich potrzeb.','FedEx WooCommerce Live Rates PRO'=>'FedEx WooCommerce Stawki Live PRO','Enable the FedEx live rates for international delivery and integrate it with your shop in less than 5 minutes. Save your time and money – let the shipping cost be calculated automatically.'=>'Zintegruj swój sklep z najpopularniejszym kurierem na świecie. Włącz stawki live oraz automatyczne usługi FedEx w WooCommerce i zapomnij o ręcznej konfiguracji cen wysyłki.','USPS Live Rates PRO'=>'USPS Stawki Live PRO','Serve your customers the automatically and real-time calculated USPS shipping rates. Add the handling fees, insurance and adjust them to your needs with just a few clicks.'=>'Integracja WooCommerce USPS z wieloma zaawansowanymi funkcjami. Wyświetl dynamicznie obliczane stawki wysyłki dla metod wysyłki USPS i dostosuj je do swoich potrzeb.','Flexible Shipping Import / Export'=>'Flexible Shipping Import / Eksport','Use the CSV files to import or export your shipping methods. Edit, update, move or backup the ready configurations and shipping scenarios.'=>'Użyj plików CSV, aby importować lub eksportować swoje metody wysyłki. Edytuj, aktualizuj, przenieś lub utwórz kopię zapasową gotowych konfiguracji i scenariuszy wysyłki.','Flexible Shipping Locations'=>'Flexible Shipping Lokalizacje','Calculate the shipping cost based on location. Define your own custom locations, use the WooCommerce defaults or the ones created by 3rd party plugins.'=>'Twórz reguły wysyłki oparte na lokalizacjach. Różnicuj koszty wysyłki dla grup, pojedynczych krajów lub województw.','Distance Based Shipping Rates'=>'Koszt wysyłki na podstawie odległości','Offer shipping rates based on Distance or Total Travel Time calculated by Google Distance Matrix API and don\'t overpay for shipping.'=>'Uzależnij koszt wysyłki od odległości lub czasu dostawy ze sklepu do klienta.','Pickup Points PRO WooCommerce Plugin'=>'Punkty Odbioru PRO','Provide your customers with multiple carriers\' pickup points at the checkout and let them choose the preferred one to collect their order from.'=>'Oferuj klientom możliwość wyboru punktów odbioru przesyłek od różnych firm kurierskich bezpośrednio podczas finalizacji zamówienia. Dzięki temu klient będzie mógł odebrać zamówienie w najbardziej dogodnym dla siebie miejscu.','Shipping Cost on Product Page PRO'=>'Koszt wysyłki na stronie produktu PRO','Let your customers calculate and see the shipping cost on product pages based on the entered shipping destination and cart contents. Decide how and when exactly you want the shipping cost calculator to display.'=>'Pozwól klientom sklepu obliczyć koszt wysyłki bezpośrednio na stronie produktu. Zdecyduj, jak i kiedy dokładnie chcesz, aby kalkulator kosztów wysyłki był wyświetlany.','Flexible Shipping Box Packing'=>'Flexible Shipping Pakowanie do wysyłki','Use the advanced box packing WooCommerce algorithm to fit the ordered products into your shipping boxes the most optimal way. Configure the shipping cost based on the type and number of the used shipping boxes.'=>'Pozwól algorytmowi pakowania dopasować zamawiane produkty do opakowań wysyłkowych i dobrać ich optymalną konfigurację. Zarządzaj kosztem wysyłki w zależności od typu i ilości użytych opakowań.','Shipping Packages'=>'Shipping Packages','Split the WooCommerce cart content into multiple packages based on various conditions like shipping class.'=>'Podziel koszyk na kilka paczek do wysyłki w oparciu o liczne warunki, takie jak choćby klasa wysyłkowa produktów w koszyku.','Multi Vendor Shipping'=>'Multi Vendor Shipping','Define precisely the shipping cost calculation rules for each Vendor / Product Author in your marketplace or multivendor store.'=>'Zdefiniuj dokładnie zasady obliczania kosztów wysyłki dla każdego sprzedawcy / autora produktu w swoim marketplace lub sklepie typu multivendor.','Delivery Date Picker'=>'Wybór czasu dostawy','Let your customers choose a convenient delivery date for the ordered products and make the shipping cost dependent on the date they choose.'=>'Pozwól klientom wybrać datę dostawy i ustalaj koszt wysyłki dla wybranego przez nich terminu.','DPD UK & DPD Local'=>'DPD UK & DPD Local','Ship your DPD orders faster with advanced DPD UK & DPD Local WooCommerce integration. Gather shipping details, download printable shipping labels and track parcels - everything is automated.'=>'Wysyłaj swoje zamówienia DPD szybciej dzięki zaawansowanej integracji DPD UK & DPD Local WooCommerce. Zbierz szczegóły wysyłki, pobierz drukowalne etykiety wysyłkowe i śledź paczki - wszystko jest zautomatyzowane.','Flexible Printing'=>'Drukowanie PrintNode','Automate your shipping process. Print the shipping labels on thermal printers via PrintNode service. Let the labels be printed automatically the same time the order is placed.'=>'Połącz WooCommerce z drukarkami za pomocą PrintNode. Automatycznie drukuj etykiety nadawcze, faktury, zamówienia i etykiety adresowe.','UPS Labels and Tracking'=>'UPS WooCommerce - Etykiety nadawcze i śledzenie przesyłek','Create the shipments, generate the printable UPS shipping labels for the placed orders and track the parcels directly from your WooCommerce store.'=>'Twórz przesyłki i generuj etykiety wysyłkowe UPS w WooCommerce. Drukuj etykiety i śledź przesyłki UPS bezpośrednio ze swojego sklepu.','DHL Express Live Rates PRO'=>'DHL Express Stawki Live PRO','WooCommerce DHL Express integration packed with many advanced features. Display the dynamically calculated live rates for DHL Express shipping methods and adjust them to your needs.'=>'Integracja WooCommerce DHL Express z wieloma zaawansowanymi funkcjami. Wyświetl dynamicznie obliczane stawki wysyłki dla metod wysyłki DHL Express i dostosuj je do swoich potrzeb.','Canada Post WooCommerce Plugin - Live Rates PRO'=>'Canada Post - Stawki Live','Offer your customers the Canada Post services with real-time calculated shipping rates. Add the handling fees, insurance and adjust them to your needs with just a few clicks.'=>'Zaoferuj swoim klientom usługi Canada Post z kosztami wysyłki naliczanymi w czasie rzeczywistym. Dodaj opłaty manipulacyjne i ubezpieczenie, a następnie dostosuj je do swoich potrzeb kilkoma kliknięciami.','Royal Mail WooCommerce - Live Rates PRO'=>'Royal Mail - Stawki Live','Let your customers choose the Royal Mail shipping methods with the real-time calculated shipping rates. Add the handling fees, insurance and adjust them to your needs in no time.'=>'Oferuj swoim klientom metody wysyłki Royal Mail z kosztami naliczanymi w czasie rzeczywistym. Dodaj opłaty manipulacyjne i ubezpieczenie, a następnie błyskawicznie dostosuj je do swoich potrzeb.','Conditional Shipping Methods'=>'Conditional Shipping Methods','Conditionally display and hide the shipping methods in your WooCommerce store. Define the rules when the specific shipping methods, e.g., live rates should be available to pick and when not to.'=>'Wyświetlaj i ukrywaj warunkowo metody wysyłki w swoim sklepie. Określ w prosty sposób reguły, kiedy konkretne metody wysyłki powinny być dostępne, a kiedy nie.','All'=>'Wszystkie','Bundles'=>'Pakiety wtyczek','Live Rates'=>'Stawki Live','Customizable Rates'=>'Zarządzanie wysyłką','Shipping Labels'=>'Etykiety wysyłkowe','Link on plugin list pageExtensions'=>'Rozszerzenia','Shipping Extensions by'=>'Rozszerzenia wysyłki od','Dive into a system of Octolize ecommerce shipping plugins for WooCommerce. Don’t lose your customers, time and money. Let our plugins secure your sales!'=>'Odkryj pełną ofertę wtyczek wysyłkowych WooCommerce od Octolize. Zarządzaj wysyłką, integruj swój sklep z kurierami i wyświetlaj najdokładniejsze stawki za wysyłkę.','Filter plugins:'=>'Filtruj wtyczki:','Blackout Lead Days'=>'Wykluczanie dni realizacji','Blackout Lead Days are used to define days of the week when shop is not processing orders.'=>'Wykluczanie dni realizacji służy do ustalenia dni tygodnia, w których sklep nie realizuje zamówień.','Monday'=>'Poniedziałek','Tuesday'=>'Wtorek','Wednesday'=>'Środa','Thursday'=>'Czwartek','Friday'=>'Piątek','Saturday'=>'Sobota','Sunday'=>'Niedziela','Can\'t convert weight to target unit.'=>'Nie można przeliczyć wagi na jednostkę docelową.','Please enter USPS User ID.'=>'Wprowadź ID użytkownika USPS.','Multi-currency is supported only in the USPS Live Rates PRO version. %1$sLearn more →%2$s'=>'Wielowalutowość jest wspierana jedynie w wersji PRO wtyczki USPS Stawki live. %1$sDowiedz się więcej →%2$s','Priority Mail'=>'Priority Mail','Priority Mail Express; Hold For Pickup'=>'Priority Mail Express; Hold For Pickup','Priority Mail Express'=>'Priority Mail Express','Standard Post'=>'Standard Post','Media Mail'=>'Media Mail','Library Mail'=>'Library Mail','Priority Mail Express; Flat Rate Envelope'=>'Priority Mail Express; Flat Rate Envelope','First-Class Mail; Large Postcards'=>'First-Class Mail; Large Postcards','Priority Mail; Flat Rate Envelope'=>'Priority Mail; Flat Rate Envelope','Priority Mail; Medium Flat Rate Box'=>'Priority Mail; Medium Flat Rate Box','Priority Mail; Large Flat Rate Box'=>'Priority Mail; Large Flat Rate Box','Priority Mail Express; Sunday/Holiday Delivery'=>'Priority Mail Express; Sunday/Holiday Delivery','Priority Mail Express; Sunday/Holiday Delivery Flat Rate Envelope'=>'Priority Mail Express; Sunday/Holiday Delivery Flat Rate Envelope','Priority Mail Express; Flat Rate Envelope Hold For Pickup'=>'Priority Mail Express; Flat Rate Envelope Hold For Pickup','Priority Mail; Small Flat Rate Box'=>'Priority Mail; Small Flat Rate Box','Priority Mail; Padded Flat Rate Envelope'=>'Priority Mail; Padded Flat Rate Envelope','Priority Mail Express; Legal Flat Rate Envelope'=>'Priority Mail Express; Legal Flat Rate Envelope','Priority Mail Express; Legal Flat Rate Envelope Hold For Pickup'=>'Priority Mail Express; Legal Flat Rate Envelope Hold For Pickup','Priority Mail Express; Sunday/Holiday Delivery Legal Flat Rate Envelope'=>'Priority Mail Express; Sunday/Holiday Delivery Legal Flat Rate Envelope','Priority Mail; Hold For Pickup'=>'Priority Mail; Hold For Pickup','Priority Mail; Large Flat Rate Box Hold For Pickup'=>'Priority Mail; Large Flat Rate Box Hold For Pickup','Priority Mail; Medium Flat Rate Box Hold For Pickup'=>'Priority Mail; Medium Flat Rate Box Hold For Pickup','Priority Mail; Small Flat Rate Box Hold For Pickup'=>'Priority Mail; Small Flat Rate Box Hold For Pickup','Priority Mail; Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Flat Rate Envelope Hold For Pickup','Priority Mail; Gift Card Flat Rate Envelope'=>'Priority Mail; Gift Card Flat Rate Envelope','Priority Mail; Gift Card Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Gift Card Flat Rate Envelope Hold For Pickup','Priority Mail; Window Flat Rate Envelope'=>'Priority Mail; Window Flat Rate Envelope','Priority Mail; Window Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Window Flat Rate Envelope Hold For Pickup','Priority Mail; Small Flat Rate Envelope'=>'Priority Mail; Small Flat Rate Envelope','Priority Mail; Small Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Small Flat Rate Envelope Hold For Pickup','Priority Mail; Legal Flat Rate Envelope'=>'Priority Mail; Legal Flat Rate Envelope','Priority Mail; Legal Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Legal Flat Rate Envelope Hold For Pickup','Priority Mail; Padded Flat Rate Envelope Hold For Pickup'=>'Priority Mail; Padded Flat Rate Envelope Hold For Pickup','Priority Mail; Regional Rate Box A'=>'Priority Mail; Regional Rate Box A','Priority Mail; Regional Rate Box A Hold For Pickup'=>'Priority Mail; Regional Rate Box A Hold For Pickup','Priority Mail; Regional Rate Box B'=>'Priority Mail; Regional Rate Box B','Priority Mail; Regional Rate Box B Hold For Pickup'=>'Priority Mail; Regional Rate Box B Hold For Pickup','First-Class; Package Service Hold For Pickup'=>'First-Class; Package Service Hold For Pickup','Priority Mail Express; Flat Rate Boxes'=>'Priority Mail Express; Flat Rate Boxes','Priority Mail Express; Flat Rate Boxes Hold For Pickup'=>'Priority Mail Express; Flat Rate Boxes Hold For Pickup','Priority Mail Express; Sunday/Holiday Delivery Flat Rate Boxes'=>'Priority Mail Express; Sunday/Holiday Delivery Flat Rate Boxes','Priority Mail; Regional Rate Box C'=>'Priority Mail; Regional Rate Box C','Priority Mail; Regional Rate Box C Hold For Pickup'=>'Priority Mail; Regional Rate Box C Hold For Pickup','First-Class; Package Service'=>'First-Class; Package Service','Priority Mail Express; Padded Flat Rate Envelope'=>'Priority Mail Express; Padded Flat Rate Envelope','Priority Mail Express; Padded Flat Rate Envelope Hold For Pickup'=>'Priority Mail Express; Padded Flat Rate Envelope Hold For Pickup','Priority Mail Express; Sunday/Holiday Delivery Padded Flat Rate Envelope'=>'Priority Mail Express; Sunday/Holiday Delivery Padded Flat Rate Envelope','Priority Mail 1-Day Return Service Padded Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Padded Flat Rate Envelope','Priority Mail 1-Day Return Service Gift Card Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Gift Card Flat Rate Envelope','Priority Mail 1-Day Return Service Window Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Window Flat Rate Envelope','Priority Mail 1-Day Return Service Small Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Small Flat Rate Envelope','Priority Mail 1-Day Return Service Legal Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Legal Flat Rate Envelope','Priority Mail 1-Day Return Service Flat Rate Envelope'=>'Priority Mail 1-Day Return Service Flat Rate Envelope','Priority Mail 1-Day Return Service Regional Rate Box A'=>'Priority Mail 1-Day Return Service Regional Rate Box A','Priority Mail 1-Day Return Service Regional Rate Box B'=>'Priority Mail 1-Day Return Service Regional Rate Box B','Priority Mail 1-Day Return Service'=>'Priority Mail 1-Day Return Service','Priority Mail 1-Day Return Service Large Flat Rate Box'=>'Priority Mail 1-Day Return Service Large Flat Rate Box','Priority Mail 1-Day Return Service Medium Flat Rate Box'=>'Priority Mail 1-Day Return Service Medium Flat Rate Box','Priority Mail 1-Day Return Service Small Flat Rate Box'=>'Priority Mail 1-Day Return Service Small Flat Rate Box','Priority Mail Military Return Service Large Flat Rate Box APO/FPO/DPO'=>'Priority Mail Military Return Service Large Flat Rate Box APO/FPO/DPO','Priority Mail 1-Day Return Service Cubic'=>'Priority Mail 1-Day Return Service Cubic','First-Class Package Return Service'=>'First-Class Package Return Service','Ground Return Service'=>'Ground Return Service','Priority Mail Express International'=>'Priority Mail Express International','Priority Mail International'=>'Priority Mail International','Global Express Guaranteed; (GXG)**'=>'Global Express Guaranteed; (GXG)**','Global Express Guaranteed; Document'=>'Global Express Guaranteed; Document','Global Express Guarantee; Non-Document Rectangular'=>'Global Express Guarantee; Non-Document Rectangular','Global Express Guaranteed; Non-Document Non-Rectangular'=>'Global Express Guaranteed; Non-Document Non-Rectangular','Priority Mail International; Flat Rate Envelope**'=>'Priority Mail International; Flat Rate Envelope**','Priority Mail International; Medium Flat Rate Box'=>'Priority Mail International; Medium Flat Rate Box','Priority Mail Express International; Flat Rate Envelope'=>'Priority Mail Express International; Flat Rate Envelope','Priority Mail International; Large Flat Rate Box'=>'Priority Mail International; Large Flat Rate Box','USPS GXG; Envelopes**'=>'USPS GXG; Envelopes**','First-Class Mail; International Letter**'=>'First-Class Mail; International Letter**','First-Class Mail; International Large Envelope**'=>'First-Class Mail; International Large Envelope**','First-Class Package International Service**'=>'First-Class Package International Service**','Priority Mail International; Small Flat Rate Box**'=>'Priority Mail International; Small Flat Rate Box**','Priority Mail Express International; Legal Flat Rate Envelope'=>'Priority Mail Express International; Legal Flat Rate Envelope','Priority Mail International; Gift Card Flat Rate Envelope**'=>'Priority Mail International; Gift Card Flat Rate Envelope**','Priority Mail International; Window Flat Rate Envelope**'=>'Priority Mail International; Window Flat Rate Envelope**','Priority Mail International; Small Flat Rate Envelope**'=>'Priority Mail International; Small Flat Rate Envelope**','USPS'=>'USPS','These are the USPS Live Rates plugin general settings. In order to learn more about its configuration please refer to its %1$sdedicated documentation →%2$s'=>'Ta strona zawiera ogólne ustawienia wtyczki USPS Live Rates. Aby dowiedzieć się więcej o jej konfiguracji, zapoznaj się z jej %1$sdedykowaną dokumentacją →%2$s','API Settings'=>'Ustawienia API','Advanced Options'=>'Opcje zaawansowane','Debug Mode'=>'Tryb debugowania','Enable debug mode'=>'Włącz tryb debugowania','Enable debug mode to display the additional tech information, incl. the data sent to USPS API, visible only for Admins and Shop Managers in the cart and checkout.'=>'Włącz Tryb debugowania, aby wyświetlić dodatkowe informacje techniczne, m.in. dane wysyłane do API USPS, widoczne tylko dla Administratorów i Menadżerów Sklepu w koszyku i na stronie finalizacji zamówienia.','Method Settings'=>'Ustawienia metody','Manage the way how the USPS services are displayed in the cart and checkout.'=>'Zarządzaj w jaki sposób metody wysyłki USPS zostaną wyświetlone w koszyku i na stronie finalizacji zamówienia.','Method Title'=>'Tytuł metody wysyłki','Define the USPS shipping method title which should be used in the cart/checkout when the Fallback option was triggered.'=>'Wprowadź tytuł dla metody wysyłki USPS, który zostanie wyświetlony klientom w koszyku i na stronie finalizacji zamówienia, jeżeli zostanie zastosowana stawka awaryjna.','USPS Live rates'=>'USPS Stawki live','Enable to offer flat rate cost for shipping so that the user can still checkout, if API for some reason returns no matching rates.'=>'Zaznacz tą opcję, aby zastosować stałą stawkę awaryjną w przypadku, gdy z jakiegoś powodu nie będzie możliwe wyświetlenie stawek rzeczywistych.','Free Shipping'=>'Darmowa wysyłka','Services'=>'Usługi USPS','Enable the services\' custom settings'=>'Włącz własne ustawienia usług','Decide which services should be displayed and which not, change their names and order. Please mind that enabling a service does not guarantee it will be visible in the cart/checkout. It has to be available for the provided package weight, origin and destination in order to be displayed.'=>'Włącz, aby określić, które usługi mają być wyświetlane, a które nie, zmienić ich nazwy i/lub kolejność. Należy mieć na uwadze, że samo włączenie usługi nie gwarantuje jeszcze, że będzie ona widoczna w koszyku lub na stronie finalizacji zamówienia. Dana usługa musi być dostępna dla podanej wagi paczki oraz adresów nadania i dostawy, aby została wyświetlona.','Domestic Services Table'=>'Tabela usług krajowych USPS','International Services Table'=>'Tabela usług międzynarodowych USPS','Type'=>'Typ','Rates Adjustments'=>'Dostosowanie stawek','Adjust these settings to get more accurate rates. Read %swhat affects the USPS rates in USPS WooCommerce plugin →%s'=>'Użyj poniższych ustawień, aby dostosować wysokość stawek do swoich potrzeb. Dowiedz się %sco wpływa na stawki w naszej wtyczce (ENG) →%s','https://octol.io/usps-free-rates'=>'https://octol.io/usps-free-rates','Commercial Rates'=>'Stawki komercyjne','Use the USPS Commercial Pricing if available'=>'Użyj USPS Commercial Pricing, jeśli jest dostępny','Tick this checkbox if you want to use and display the Commercial Pricing shipping rates to your customers instead of the standard ones.'=>'Zaznacz ten checkbox, jeżeli zamiast standardowych stawek USPS chcesz używać stawek Commercial Pricing.','Insurance'=>'Ubezpieczenie','Request insurance to be included in the USPS rates'=>'Uwzględniaj ubezpieczenie w stawkach USPS','Enable if you want to include insurance in the USPS rates if possible.'=>'Włącz opcję jeśli chcesz uwzględniać ubezpieczenie w stawkach USPS, gdy jest dostępne.','Empty response!'=>'Pusta odpowiedź API!','USPS Live Rates'=>'USPS Stawki live','USPS integration'=>'Integracja USPS','None'=>'Brak','The &#8220;%s&#8221; plugin cannot run on PHP versions older than %s. Please contact your host and ask them to upgrade.'=>'Wtyczka &#8220;%s&#8221; nie może działać na wersjach PHP starszych niż %s. Skontaktuj się z administratorem hostingu i poproś o aktualizację wersji PHP.','The &#8220;%s&#8221; plugin cannot run on WordPress versions older than %s. Please update WordPress.'=>'Wtyczka &#8220;%s&#8221; nie może działać na wersjach WordPressa starszych niż %s. Wykonaj aktualizację WordPressa.','The &#8220;%s&#8221; plugin cannot run on WooCommerce versions older than %s. Please update WooCommerce.'=>'Wtyczka &#8220;%s&#8221; nie może działać na wersjach WooCommerce starszych niż %s. Wykonaj aktualizację WooCommerce.','The &#8220;%s&#8221; plugin cannot run without OpenSSL module version at least %s. Please update OpenSSL module.'=>'Wtyczka &#8220;%s&#8221; nie może działać bez aktywacji OpenSSL co najmniej w wersji %s. Wykonaj aktualizację OpenSSL.','The &#8220;%1$s&#8221; plugin requires at least %2$s version of %3$s to work correctly. Please update it to its latest release.'=>'Do poprawnego działania wtyczki &#8220;%1$s&#8221; wymagana jest wtyczka %3$s co najmniej w wersji %2$s. Wykonaj jej aktualizację.','The &#8220;%s&#8221; plugin cannot run without %s active. Please install and activate %s plugin.'=>'Wtyczka &#8220;%s&#8221; nie może działać bez aktywacji %s. Zainstaluj i aktywuj wtyczkę %s.','The &#8220;%s&#8221; plugin requires activating %s plugin. <a href="%s">Activate %s</a>'=>'Wtyczka &#8220;%s&#8221; wymaga aktywnej wtyczki %s. <a href="%s">Włącz %s</a>','The &#8220;%s&#8221; plugin cannot run without %s PHP module installed. Please contact your host and ask them to install %s.'=>'Wtyczka &#8220;%s&#8221; nie może działać bez modułu PHP: %s. Skontaktuj się z administratorem hostingu i poproś o instalację %s.','The &#8220;%s&#8221; plugin cannot run without %s PHP setting set to %s. Please contact your host and ask them to set %s.'=>'Wtyczka &#8220;%s&#8221; nie może działać bez ustawienia PHP %s ustawionego na %s. Skontaktuj się z administratorem hostingu i poproś o ustawienie %s.','The &#8220;%s&#8221; plugin is temporarily disabled since the required %s plugin is being upgraded.'=>'Wtyczka &#8220;%s&#8221; została chwilowo wyłączona, ponieważ wymagana wtyczka %s jest właśnie aktualizowana.','Set image'=>'Wstaw obraz','Remove image'=>'Usuń obraz','Select or Upload Media'=>'Wybierz lub prześlij media','Use this media'=>'Wybierz','Search for a product&hellip;'=>'Szukaj produktu&hellip;','Disabled'=>'Wyłączono','Emergency'=>'Wypadek','Alert'=>'Ostrzeżenie','Critical'=>'Błąd krytyczny','Error'=>'Błąd','Warning'=>'Ostrzeżenie','Notice'=>'Powiadomienie','Info'=>'Informacja','Debug'=>'Debugowanie','Can not enable WP Desk Debug log! Cannot create directory %s or this directory is not writeable!'=>'Nie można włączyć WP Desk Debug log! Nie można utworzyć katalogu %s lub brak uprawnień do zapisu w tym katalogu!','Can not enable WP Desk Debug log! Cannot create file %s!'=>'Nie można włączyć logowania błędów za pomocą WP Desk Debug! Nie można utworzyć pliku %s!','Length %1$s'=>'Długość %1$s','Width %1$s'=>'Szerokość %1$s','Height %1$s'=>'Wysokość %1$s','Max Weight %1$s'=>'Waga maks. %1$s','Padding %1$s'=>'Wypełnienie %1$s','Box Weight %1$s'=>'Waga pudełka %1$s','Delete'=>'Usuń','Add'=>'Dodaj','Select the package type the ordered products will be matched to. You can choose one or as many different packagings as you need. If selected, filling in the products\' weight and dimensions fields is required. %1$sLearn more about the sizes and package types →%2$s'=>'Wybierz typ lub typy opakowań, do których zostaną dopasowane i zapakowane zamówione produkty. Możesz w tym miejscu wybrać ich dowolną ilość. Wybór tej opcji wymaga uzupełnienia wagi i wymiarów każdego z produktów. %1$sDowiedz się więcej o wymiarach i typach opakowań →%2$s','Shipping boxes'=>'Pudełka','Pack into one box by weight'=>'Pakuj do jednego pudełka według wagi','Pack into custom boxes (3D bin packing)'=>'Pakuj do własnych pudełek (Pakowanie 3D bin)','Pack into custom boxes (volume packing)'=>'Pakuj do własnych pudełek (Pakowanie wg objętości)','Pack items separately'=>'Pakuj produkty oddzielnie','Parcel Packing Method'=>'Metoda pakowania paczek','Define the way how the ordered products should be packed. Changing your choice here may affect the rates.'=>'Określ sposób pakowania zamawianych produktów. Wybór poszczególnych opcji w tym miejscu może mieć wpływ na wysokość stawek.','Packs all items into a single box.'=>'Pakuje wszystkie przedmioty do jednego pudełka.','Optimizes item placement in custom boxes using a 3D algorithm that considers item dimensions and available space.'=>'Optymalizuje rozmieszczenie przedmiotów w niestandardowych pudełkach za pomocą algorytmu 3D, uwzględniając wymiary i dostępną przestrzeń.','Groups items by total volume and fits them into boxes to maximize capacity. Also ensures each item can physically fit into an empty box.'=>'Grupuje przedmioty według całkowitej objętości i umieszcza je w pudełkach, aby maksymalnie wykorzystać ich pojemność. Dodatkowo sprawdza, czy każdy przedmiot mieści się fizycznie w pustym pudełku.','Packs each item in a separate box.'=>'Pakuje każdy przedmiot do osobnego pudełka.','Cannot pack item: %1$s. %2$s'=>'Nie można zapakować produktu: %1$s. %2$s','Cannot pack items: %1$s.'=>'Nie można zapakować produktu: %1$s.','weight'=>'waga','width'=>'szerokość','length'=>'długość','height'=>'wysokość','Item %1$s not set!'=>'Dane produktu %1$s nie zostały zdefiniowane!','Custom (%1$sx%2$sx%3$s/%4$s)'=>'Własne (%1$sx%2$sx%3$s/%4$s)','Dynamically calculated USPS live rates based on the established USPS API connection. %1$sLearn more →%2$s'=>'Dynamicznie obliczane w oparciu o połączenie z API USPS stawki rzeczywiste. %1$sDowiedz się więcej →%2$s','%1$sPlease add now the %2$s shipping method within the shipping zone you want it to be available in →%3$s'=>'%1$sDodaj teraz metodę wysyłki %2$s w strefie wysyłkowej, w której chcesz, aby była dostępna →%3$s','API Connection Status'=>'Status połączenia z API','Checking...'=>'Sprawdzanie...','If you encounter any problems with establishing the API connection, the detailed information on its cause will be displayed here.'=>'Jeśli wystąpią problemy z nawiązaniem połączenia z API, w tym miejscu zostaną wyświetlone szczegółowe informacje o ich przyczynie.','OK'=>'OK','Not found HTML view for custom field %1$s.'=>'Nie znaleziono widoku HTML dla pola niestandardowego %1$s.','Fixed value'=>'Wartość stała','Percentage'=>'Wartość procentowa','Use this option to apply the handling fees to the rates. You can use either fixed or percentage values, including the negative ones for discounts.'=>'Użyj tej opcji by doliczyć stały lub procentowy koszt prowizji do zwracanych przez wtyczkę stawek, włącznie z minusowymi wartościami dla obniżek.','Fees Amount'=>'Wartość prowizji','Enter a positive value to charge your customers additionally or a negative one to apply the discount. If you use the currency switcher, the final price incl. Handling Fee will be automatically converted to the currently active currency in your shop. The rates will also include the taxes based on your current Tax settings.'=>'Dodatnia liczba = dopłata, liczba ujemna = zniżka. Jeżeli używasz switcherów walut, cena końcowa zawierająca prowizję zostanie automatycznie przeliczona na obecnie aktywną walutę w sklepie. W cenie zostaną również uwzględnione aktualne ustawienia podatków.','Code'=>'Kod','Name'=>'Nazwa','Enabled'=>'Włączone','Drag and drop the services to control their display order. Confirm by clicking Save changes button below.'=>'Przeciągnij i upuść usługę, aby zmienić kolejność wyświetlania. Zatwierdź zmiany klikając Zapisz zmiany poniżej.','Custom Origin'=>'Własny adres nadania','Origin Address'=>'Ulica','Origin City'=>'Miasto','Origin Postcode'=>'Kod pocztowy','Origin Country/State'=>'Kraj','Origin Settings for the whole store'=>'Adres nadania dla całego sklepu','Origin Settings'=>'Adres nadania','Enable custom origin for the whole store'=>'Ustaw własny adres nadania dla całego sklepu','Enable custom origin'=>'Ustaw własny adres nadania','Use this option to use a different shipper\'s address than the one defined in the WooCommerce settings. If not enabled, the one you\'ve entered in WooCommerce → Settings → General → Store Address will be used by default. Each shipping method can have their own custom origin.'=>'Zaznacz tę opcję, by użyć innego adresu nadawcy niż domyślny adres sklepu, zdefiniowany w ustawieniach WooCommerce (WooCommerce → Ustawienia → Ogólne → Adres sklepu). Metody wysyłki mogą mieć własny adres nadania.','Use this option to use a different shipper\'s address than the one defined in the WooCommerce settings. If not enabled, the one you\'ve entered in WooCommerce → Settings → General → Store Address will be used by default.'=>'Zaznacz tę opcję, by użyć innego adresu nadawcy niż domyślny adres sklepu, zdefiniowany w ustawieniach WooCommerce (WooCommerce → Ustawienia → Ogólne → Adres sklepu).','Origin Settings for this Shipping Method'=>'Adres nadania dla tej metody wysyłki','Enable custom origin for this shipping method'=>'Ustaw własny adres nadania dla tej metody wysyłki','Use this option to use a different shipper\'s address for this shipping method.'=>'Zaznacz tę opcję, by użyć innego adresu nadawcy dla tej metody wysyłki.','Cart value exceeds %s. Free shipping has been applied.'=>'Wartość koszyka przekracza %s. Zastosowano bezpłatną wysyłkę.','Cart value doesn\'t exceed %s. Free shipping hasn\'t been applied.'=>'Wartość koszyka nie przekracza %s. Bezpłatna wysyłka nie została zastosowana.','Subtotal'=>'Kwota','Free Shipping Amount'=>'Kwota darmowej wysyłki','Enable the free shipping over amount'=>'Włącz darmową wysyłkę powyżej określonej kwoty','Tick this checkbox to enter the order amount above which the shipping becomes free.'=>'Zaznacz, aby wprowadzić wartość zamówienia, po przekroczeniu której wysyłka stanie się darmowa.','Free Shipping Threshold'=>'Próg darmowej wysyłki','Enter only a numeric value without the currency symbol.'=>'Wprowadź samą wartość, bez symbolu waluty.','Show %1$s'=>'Pokaż %1$s','Hide %1$s'=>'Ukryj %1$s','Copy %1$s to clipboard'=>'Skopiuj %1$s do schowka','Fallback reason'=>'Powód stawki awaryjnej','Packages'=>'Paczki','Custom'=>'Własna','No rates added from collection point rates!'=>'Nie zostały dodane żadne stawki dla punktów odbioru!','Collection point not found.'=>'Punkt odbioru nie został znaleziony.','There was no valid services available.'=>'Żadne stawki wysyłki nie były dostępne.','Fallback rate added with reason: %1$s'=>'Została użyta stawka awaryjna. Powód: %1$s','No rates added from standard service rates!'=>'Nie zostały dodane żadne stawki standardowe!','Additional costs by Flexible Shipping Table Rate'=>'Dodatkowy koszt przez Flexible Shipping Table Rate','Additional costs'=>'Dodatkowe koszty','Enable Flexible Shipping Rules Table'=>'Włącz tabelę reguł Flexible Shipping','Enabling this option will allow you to adjust the calculated shipping cost, with additional charges applied according to the shipping cost calculation rules.'=>'Włączenie tej opcji umożliwi Ci dostosowanie wyliczonego kosztu przesyłki, naliczając dodatkowe opłaty wg zadeklarowanych reguł kalkulacji kosztów.','Shipping Cost Calculation Rules'=>'Reguły kalkulacji kosztów wysyłki','You can update %1$sFlexible Shipping%2$s plugin to define additional costs for this shipping method. %3$s'=>'Możesz zaktualizować wtyczkę %1$sFlexible Shipping%2$s, aby zdefiniować dodatkowe koszty dla tej metody wysyłki. %3$s','Update Now'=>'Zaktualizuj Teraz','You can activate %1$sFlexible Shipping%2$s plugin to define additional costs for this shipping method. %3$s'=>'Możesz aktywować wtyczkę %1$sFlexible Shipping%2$s, aby zdefiniować dodatkowe koszty dla tej metody wysyłki. %3$s','Activate Now'=>'Aktywuj Teraz','You can install %1$sFlexible Shipping%2$s plugin to define additional costs for this shipping method. %3$s'=>'Możesz zainstalować wtyczkę %1$sFlexible Shipping%2$s, aby zdefiniować dodatkowe koszty dla tej metody wysyłki. %3$s','Install Now'=>'Zainstaluj teraz','Fallback'=>'Stawka awaryjna','Enable fallback'=>'Włącz stawkę awaryjną','Fallback Cost'=>'Koszt stawki awaryjnej','We noticed that you are using the Germanized plugin along with %1$s. In order to display the accurate rates, please select the WooCommerce default option in the Germanized Taxes > Additional Costs > Tax calculation mode settings. %2$sNeed assistance? Get in touch →%3$s'=>'Zauważyliśmy, że używasz wtyczki Germanized wraz z %1$s. Aby wyświetlić dokładne stawki, wybierz domyślną opcję WooCommerce w ustawieniach trybu obliczania podatku zgermanizowanych podatków > Koszty dodatkowe >. %2$sPotrzebujesz pomocy? Skontaktuj się z →%3$s','(Delivery Days: 0 days)'=>'(Dni do dostawy: 0 dni)','(Delivery Days: %1$d day)'=>'(Dni do dostawy: %1$d dzień)' . "\0" . '' . "\0" . '','(Delivery Date: %1$s)'=>'(Data dostawy: %1$s)','Awesome, you\'ve been using %s for more than 2 weeks. Could you please do me a BIG favor and give it a 5-star rating on WordPress? ~ Peter'=>'Fantastycznie, korzystasz z %s już ponad 2 tygodnie. Czy możesz wyświadczyć mi wielką przysługę i wystawić 5-gwiazdkową ocenę? ~Piotr','%1$sOk, you deserved it%2$s'=>'%1$sTak, zasługujecie na nią%2$s','%1$sNope, maybe later%2$s'=>'%1$sNie, może później%2$s','%1$sI already did%2$s'=>'%1$sJuż to zrobiłem/zrobiłam%2$s','Created with %1$s by %2$s - If you like %3$s you can %4$srate us %5$s in plugins repository &rarr;%6$s'=>'Stworzone z %1$s przez %2$s - jeżeli podoba Ci się %3$s możesz %4$sdać nam %5$s w repozytorium wtyczek &rarr;%6$s','The plugin suddenly stopped working'=>'Wtyczka nagle przestała działać','The plugin broke my site'=>'Wtyczka zepsuła moją stronę','I have found a better plugin'=>'Znalazłem lepszą wtyczkę','What\'s the plugin\'s name?'=>'Jaka to wtyczka?','I only needed the plugin for a short period'=>'Potrzebowałem wtyczki tylko przez krótki okres','I no longer need the plugin'=>'Nie potrzebuję już wtyczki','It\'s a temporary deactivation. I\'m just debugging an issue.'=>'Jest to tymczasowa dezaktywacja, debuguję problem.','Other'=>'Inny','Please let us know how we can improve our plugin'=>'Napisz co moglibyśmy zmienić, aby ulepszyć naszą wtyczkę','Plugin deactivation'=>'Wyłączanie wtyczki','You are deactivating %s plugin.'=>'Dezaktywujesz wtyczkę %s.','Skip &amp; Deactivate'=>'Pomiń i dezaktywuj','Submit &amp; Deactivate'=>'Wyślij i dezaktywuj','Cancel'=>'Anuluj','You successfully opted out of collecting usage data by %1$s. If you change your mind, you can always opt in later in the plugin\'s quick links.'=>'Twoja rezygnacja ze zbierania danych użytkowania wtyczek przez %1$s została przyjęta. Jeśli zmienisz zdanie, zawsze możesz włączyć zgodę ponownie za pomocą szybkich linków na liście wtyczek.','Enable usage tracking'=>'Włącz śledzenie użycia','Disable usage tracking'=>'Wyłącz śledzenie użycia','Hey %s,'=>'Hej %s,','Please help us improve our plugins! If you opt-in, we will collect some non-sensitive data and usage information anonymously. If you skip this, that\'s okay! All plugins will work just fine.'=>'Proszę pomóż nam udoskonalić nasze wtyczki! Jeśli się zgodzisz, będziemy zbierać niewrażliwe dane użytkowania anonimowo. Jeśli pominiesz ten krok, w porządku! Wszystkie wtyczki będą działać niezależnie od Twojej decyzji.','Allow & Continue &rarr;'=>'Pozwól i kontynuuj &rarr;','Skip'=>'Pomiń','What permissions are being granted?'=>'Jakie uprawnienia są przyznawane?','Your Site Overview'=>'Dane witryny','WP version, PHP info'=>'Wersja WP, wersja PHP','Plugin Usage'=>'Dane użytkowania wtyczek','Current settings and usage information of %1$s plugins'=>'Ustawienia i informacje o użytkowaniu wtyczek %1$s','Your Store Overview'=>'Dane sklepu','Anonymized and non-sensitive store usage information'=>'Tylko niewrażliwe i anonimowe dane o użytkowaniu sklepu','Find out more &raquo;'=>'Dowiedz się więcej &raquo;','Current settings and usage information of WP Desk plugins'=>'Ustawienia i informacje o użytkowaniu wtyczek WP Desk',' If you have a moment, please let us know why you are deactivating plugin (anonymous feedback):'=>' Jeśli masz chwilę, prosimy daj nam znać dlaczego dezaktywujesz wtyczkę (anonimowa opinia):','I found a better plugin'=>'Znalazłem lepszą wtyczkę','Kindly tell us the reason so we can improve'=>'Napisz nam powód abyśmy mogli go poprawić','We need your help to improve <strong>WP Desk plugins</strong>, so they are more useful for you and the rest of <strong>30,000+ users</strong>. By collecting data on how you use our plugins, you will help us a lot. We will not collect any sensitive data, so you can feel safe.'=>'Potrzebujemy Twojej pomocy, aby dalej rozwijać <strong>wtyczki WP Desk</strong> i były one jeszcze bardziej pomocne dla Ciebie i pozostałych ponad <strong>30 000 użytkowników</strong>. Zbierając dane na temat tego jak korzystasz z naszych wtyczek bardzo nam w tym pomożesz. Nie będziemy zbierać żadnych danych wrażliwych, więc możesz czuć się bezpiecznie.','You successfully opted out of collecting usage data by WP Desk. If you change your mind, you can always opt in later in the plugin\'s quick links.'=>'Twoja rezygnacja ze zbierania danych użytkowania wtyczek przez WP Desk została przyjęta. Jeśli zmienisz zdanie, zawsze możesz włączyć zgodę ponownie za pomocą szybkich linków na liście wtyczek.','Upgrade'=>'Kup PRO','Upgrade Now &rarr;'=>'Zaktualizuj do wersji PRO &rarr;','USPS WooCommerce shipping methods with real-time calculated shipping rates based on the established USPS API connection.'=>'Metody wysyłki USPS dla WooCommerce ze stawkami obliczanymi w czasie rzeczywistym w oparciu o nawiązane połączenie z API USPS.','You need to provide the USPS account credentials to get live rates. If you do not have the USPS account yet, you can register a new account directly at %1$susps.com →%2$s'=>'Musisz wprowadzić dane dostępowe konta USPS, aby uzyskać stawki rzeczywiste. Jeżeli nie posiadasz jeszcze konta USPS, możesz utworzyć nowe konto bezpośrednio na stronie %1$susps.com →%2$s','USPS User ID *'=>'ID użytkownika USPS *','Enter the credentials you acquired during the USPS account registration process.'=>'Wprowadź dane dostępowe uzyskane podczas procesu rejestracji konta USPS.','Test Credentials'=>'Testowe dane dostępowe','Enable to use test credentials'=>'Używaj testowych danych dostępowych','The &#8220;%s&#8221; plugin requires free %s plugin. <a href="%s">Install%s</a>'=>'Wtyczka &#8220;%s&#8221; wymaga bezpłatnej wtyczki %s. <a href="%s">Zainstaluj %s</a>','Shipping service should implements CanTestSettings interface! Or overwrite this method'=>'Shipping service should implements CanTestSettings interface! Or overwrite this method','If you have a moment, please let us know why you are deactivating the plugin (anonymous feedback):'=>'Jeśli masz chwilę, prosimy daj nam znać dlaczego wyłączasz wtyczkę (anonimowa opinia):']];
