@@ -76,17 +76,6 @@ class Notice
         $this->addAction();
     }
     /**
-     * @return bool
-     */
-    public function isBlockEditor()
-    {
-        if (!function_exists('get_current_screen')) {
-            require_once \ABSPATH . '/wp-admin/includes/screen.php';
-        }
-        $screen = \get_current_screen();
-        return is_object($screen) ? $screen->is_block_editor() : \false;
-    }
-    /**
      * @return string
      */
     public function getNoticeContent()
@@ -154,7 +143,6 @@ class Notice
         if (!$this->actionAdded) {
             add_action('admin_notices', [$this, 'showNotice'], $this->priority);
             add_action('admin_footer', [$this, 'showNotice'], self::ADMIN_FOOTER_BASE_PRIORITY + intval($this->priority));
-            add_action('admin_head', [$this, 'addGutenbergScript']);
             $this->actionAdded = \true;
         }
     }
@@ -167,15 +155,6 @@ class Notice
             remove_action('admin_notices', [$this, 'showNotice'], $this->priority);
             remove_action('admin_footer', [$this, 'showNotice'], self::ADMIN_FOOTER_BASE_PRIORITY + intval($this->priority));
             $this->actionAdded = \false;
-        }
-    }
-    /**
-     * Enqueue admin scripts.
-     */
-    public function addGutenbergScript()
-    {
-        if ($this->isBlockEditor()) {
-            include_once __DIR__ . '/views/admin-head-js-gutenberg.php';
         }
     }
     /**
@@ -247,6 +226,6 @@ class Notice
         if ($this->addParagraphToContent()) {
             $noticeFormat = '<div %1$s><p>%2$s</p></div>';
         }
-        echo sprintf($noticeFormat, $this->getAttributesAsString(), $this->noticeContent);
+        echo \wp_kses_post(sprintf($noticeFormat, $this->getAttributesAsString(), $this->noticeContent));
     }
 }
