@@ -110,76 +110,4 @@ if (!\defined('ABSPATH')) {
 \esc_html_e('Skip &amp; Deactivate', 'flexible-shipping-usps');
 ?></a>
 	</div>
-</div>
-<script type="text/javascript">
-	jQuery('input[type=radio]').click(function(){
-		var reason = jQuery('input[name=selected-reason]:checked').val();
-		console.log(reason);
-		jQuery('.reason-input').hide();
-		if ( reason == 'found_better_plugin' ) {
-			jQuery('#found_better_plugin').show();
-		}
-		if ( reason == 'other' ) {
-			jQuery('#other').show();
-		}
-		jQuery('.wpdesk_tracker_deactivate .button-deactivate').html( '<?php 
-\esc_html_e("Submit &amp; Deactivate", "flexible-shipping-usps");
-?>' );
-	})
-	jQuery('.button-deactivate').click(function(e){
-		e.preventDefault();
-		console.log('deactivate');
-		var reason = jQuery('input[name=selected-reason]:checked').val();
-		var plugin = '<?php 
-echo \esc_attr($plugin);
-?>';
-		var plugin_name = '<?php 
-echo \esc_attr($plugin_name);
-?>';
-		var additional_info = '';
-		if ( reason == 'found_better_plugin' ) {
-			additional_info = jQuery('#found_better_plugin input').val();
-		}
-		if ( reason == 'other' ) {
-			additional_info = jQuery('#other input').val();
-		}
-		console.log(reason);
-		if ( typeof reason != 'undefined' ) {
-			console.log('not undefined');
-			jQuery('.button').attr('disabled',true);
-			jQuery.ajax( '<?php 
-echo \esc_url_raw(\admin_url("admin-ajax.php"));
-?>',
-				{
-					type: 'POST',
-					data: {
-						action: 'wpdesk_tracker_deactivation_handler',
-						security: '<?php 
-echo \esc_attr(\wp_create_nonce(WPDesk_Tracker::WPDESK_TRACKER_DEACTIVATION));
-?>',
-						reason: reason,
-						plugin: plugin,
-						plugin_name: plugin_name,
-						additional_info: additional_info,
-					}
-				}
-			).always(function() {
-				var url = '<?php 
-echo \str_replace("&amp;", "&", \admin_url(\wp_nonce_url("plugins.php?action=deactivate&plugin=" . $plugin . "&plugin_status=all&", "deactivate-plugin_" . $plugin)));
-?>';
-				window.location.href = url;
-			});
-		}
-		else {
-			var url = '<?php 
-echo \str_replace("&amp;", "&", \admin_url(\wp_nonce_url("plugins.php?action=deactivate&plugin=" . $plugin . "&plugin_status=all&", "deactivate-plugin_" . $plugin)));
-?>';
-			window.location.href = url;
-		}
-	})
-	jQuery('.button-close').click(function(e){
-		e.preventDefault();
-		window.history.back();
-	})
-</script>
-<?php 
+</div><?php 
